Calculate the light level where a coral reef switches from shallow to mesophotic

This function finds the light level where the shallow and mesophotic community values are equal.

boundary(Vm = 0.393, K = 13.5, a = 1.54, b = 8.92)

Arguments

Vm
Passed to shallow. The maximum similarity of a reef patch to the idealised shallow reef community, between 0 and 1. Defaults to (Tamir et al., 2019).

K
Passed to shallow. The light value where the community value = Vmax/2, between 0 and 100. Defaults to (Tamir et al., 2019). This can be interpreted as the light level at which shallow reef communities become light limited.

a
Passed to mesophotic. The scale parameter of a Weibull distribution, defaults to (Tamir et al., 2019). a describes the spread of the data, or ‘the peakyness’. Larger numbers create a flatter, wider, curve. Varying a will simultaneously change the width of the preferred light environment of mesophotic taxa, and the maximum similarity a reef patch is expected to achieve in comparison to the idealised mesophotic community.

b
Passed to mesophotic. The shape parameter of a Weibull distribution, defaults to (Tamir et al., 2019). A value of 3 approximates a normal distribution. Values larger than 3 introduce a left-skew to the curve. This means mesophotic taxa can be light-limited, while keeping mesophotic communities in low-light environments. Varying b shifts the preferred light environment of mesophotic communities.

Value

The light value when the shallow and mesophotic community values are equal.

Details

When the community values are equal for shallow and mesophotic communities, given a light level, a reef looks equally similar to a shallow or mesophotic reef. Changing the light level will tip the reef towards either a shallow or mesophotic community.

Strictly, this function finds the mathematical root of reef, searching between light levels of 0.01 and 100 percent of surface Photosynthetically Active Radiation (PAR).

depth can then be used to convert the boundary light level to boundry depths under different environmental conditions.

References

(Tamir et al., 2019)

(Laverick et al., 2020)

See also

Examples

# By default the function uses the shallow and mesophotic
# community light relationships from (Tamir et al., 2019).

   boundary()

#> [1] 0.03590561

# You can specify your own community light relationships
# by setting parameters explicitly

   boundary(Vm = 0.2, K = 11, a = 1.2, b = 8)

#> [1] 0.05945372