Calculate the depth of a light level

This function uses the Lamberte-Beere law to predict the depth in metres a given proportion of surface Photosynthetically Active Radiation (PAR) penetrates to.

depth(light = seq(0.0065, 1, length = 500), KdPAR = 0.05, shade = 1)

Arguments

light
The proportion of surface irradiance (PAR), between 0 and 1. Defaults to a vector of light values.

KdPAR
The average light attenuation coefficient of the water column above the reef. Defaults to 0.05.

shade
Additional shading at the seafloor, such as bathymetric effects, between 0 (no light available) and 1 (all light available). Defaults to 1 (all light available).

Value

A depth in metres. When a single light value is provided, a single value is returned. You can pass a vector of light levels for a vector of depths.

Details

The function takes the average light attenuation coefficient (KdPAR) of the water column above a reef and a proportion between 0 and 1 of surface PAR. The Lamberte-Beer law is then used to predict the depth in metres this light environment occurs at.

A shading parameter can be used to scale how much of the light which reaches the seafloor is available to the benthos. According to (Lesser et al., 2018) the shading effect of reef slope at a vertical wall is to limit surface irradiance to 25 percent (shade = 0.25). The default behaviour is for there to be no shading (Shade = 1). When shading is more severe, the function will return a shallower depth. This is as the function is predicting the depth you would find your specified light level in the shaded environment.

Leaving shade unspecified makes this function equivalent to the "Depth_function" in (Laverick et al. 2020).

Specifying shading makes this function equivalent to the "Shade_function" in (Laverick et al., 2020).

References

Lamberte-Beer law

Lesser

Laverick

Examples

# Specifying a single light value and light
# attenuation returns a single depth.

   depth(light = 0.1, KdPAR = 0.03)

#> [1] 76.75284

# By defualt there is no shading. You can
# switch this on by passing a value.

   depth(light = 0.1, KdPAR = 0.03, shade = 0.8)

#> [1] 69.31472

# Specifying a vector of light values and light
# attenuation returns a vector of depths.

   depth(light = seq(0.0065, 1, length = 500), KdPAR = 0.03)

#>   [1] 167.86510340 158.95835414 151.93664066 146.13977696 141.20324877
#>   [6] 136.90437923 133.09712003 129.68048244 126.58171781 123.74668029
#>  [11] 121.13397940 118.71126362 116.45276615 114.33763086 112.34873867
#>  [16] 110.47186483 108.69506111 107.00819453 105.40259730 103.87079730
#>  [21] 102.40630778 101.00346137  99.65727765  98.36335643  97.11779096
#>  [26]  95.91709682  94.75815315  93.63815375  92.55456619  91.50509728
#>  [31]  90.48766388  89.50036798  88.54147542  87.60939753  86.70267533
#>  [36]  85.81996573  84.96002965  84.12172148  83.30397995  82.50582004
#>  [41]  81.72632582  80.96464419  80.21997922  79.49158715  78.77877195
#>  [46]  78.08088132  77.39730309  76.72746199  76.07081677  75.42685753
#>  [51]  74.79510339  74.17510027  73.56641900  72.96865349  72.38141912
#>  [56]  71.80435122  71.23710377  70.67934812  70.13077182  69.59107762
#>  [61]  69.05998247  68.53721665  68.02252292  67.51565578  67.01638078
#>  [66]  66.52447381  66.03972057  65.56191598  65.09086363  64.62637533
#>  [71]  64.16827067  63.71637654  63.27052682  62.83056192  62.39632854
#>  [76]  61.96767924  61.54447224  61.12657107  60.71384434  60.30616547
#>  [81]  59.90341248  59.50546775  59.11221783  58.72355323  58.33936825
#>  [86]  57.95956081  57.58403227  57.21268729  56.84543369  56.48218229
#>  [91]  56.12284680  55.76734370  55.41559209  55.06751364  54.72303241
#>  [96]  54.38207481  54.04456950  53.71044725  53.37964093  53.05208535
#> [101]  52.72771726  52.40647521  52.08829953  51.77313223  51.46091695
#> [106]  51.15159891  50.84512482  50.54144288  50.24050266  49.94225509
#> [111]  49.64665243  49.35364817  49.06319702  48.77525489  48.48977879
#> [116]  48.20672684  47.92605822  47.64773313  47.37171275  47.09795923
#> [121]  46.82643564  46.55710595  46.28993498  46.02488840  45.76193269
#> [126]  45.50103514  45.24216376  44.98528733  44.73037533  44.47739796
#> [131]  44.22632605  43.97713114  43.72978535  43.48426144  43.24053279
#> [136]  42.99857331  42.75835752  42.51986045  42.28305770  42.04792535
#> [141]  41.81444001  41.58257876  41.35231916  41.12363925  40.89651748
#> [146]  40.67093278  40.44686448  40.22429232  40.00319646  39.78355744
#> [151]  39.56535620  39.34857402  39.13319257  38.91919387  38.70656028
#> [156]  38.49527448  38.28531950  38.07667868  37.86933567  37.66327443
#> [161]  37.45847920  37.25493452  37.05262522  36.85153639  36.65165339
#> [166]  36.45296184  36.25544763  36.05909688  35.86389597  35.66983151
#> [171]  35.47689035  35.28505955  35.09432641  34.90467843  34.71610335
#> [176]  34.52858908  34.34212376  34.15669573  33.97229349  33.78890577
#> [181]  33.60652147  33.42512966  33.24471959  33.06528071  32.88680261
#> [186]  32.70927505  32.53268797  32.35703145  32.18229574  32.00847123
#> [191]  31.83554846  31.66351814  31.49237109  31.32209830  31.15269087
#> [196]  30.98414006  30.81643724  30.64957393  30.48354176  30.31833250
#> [201]  30.15393801  29.99035032  29.82756154  29.66556390  29.50434975
#> [206]  29.34391155  29.18424187  29.02533337  28.86717884  28.70977115
#> [211]  28.55310329  28.39716832  28.24195943  28.08746989  27.93369306
#> [216]  27.78062239  27.62825142  27.47657380  27.32558323  27.17527352
#> [221]  27.02563857  26.87667233  26.72836886  26.58072228  26.43372681
#> [226]  26.28737672  26.14166638  25.99659021  25.85214271  25.70831847
#> [231]  25.56511213  25.42251839  25.28053205  25.13914795  24.99836100
#> [236]  24.85816618  24.71855853  24.57953315  24.44108520  24.30320991
#> [241]  24.16590256  24.02915849  23.89297310  23.75734183  23.62226021
#> [246]  23.48772379  23.35372819  23.22026908  23.08734218  22.95494326
#> [251]  22.82306814  22.69171270  22.56087286  22.43054458  22.30072388
#> [256]  22.17140682  22.04258951  21.91426811  21.78643880  21.65909783
#> [261]  21.53224148  21.40586607  21.27996798  21.15454361  21.02958941
#> [266]  20.90510187  20.78107751  20.65751290  20.53440465  20.41174940
#> [271]  20.28954382  20.16778463  20.04646858  19.92559246  19.80515309
#> [276]  19.68514732  19.56557204  19.44642417  19.32770068  19.20939854
#> [281]  19.09151478  18.97404645  18.85699064  18.74034444  18.62410501
#> [286]  18.50826952  18.39283518  18.27779921  18.16315887  18.04891145
#> [291]  17.93505427  17.82158467  17.70850002  17.59579772  17.48347518
#> [296]  17.37152987  17.25995925  17.14876083  17.03793213  16.92747069
#> [301]  16.81737410  16.70763995  16.59826586  16.48924948  16.38058848
#> [306]  16.27228054  16.16432338  16.05671473  15.94945236  15.84253403
#> [311]  15.73595756  15.62972075  15.52382146  15.41825754  15.31302688
#> [316]  15.20812738  15.10355696  14.99931356  14.89539515  14.79179970
#> [321]  14.68852522  14.58556971  14.48293123  14.38060781  14.27859753
#> [326]  14.17689849  14.07550878  13.97442653  13.87364989  13.77317700
#> [331]  13.67300605  13.57313523  13.47356273  13.37428679  13.27530565
#> [336]  13.17661755  13.07822077  12.98011359  12.88229432  12.78476127
#> [341]  12.68751276  12.59054715  12.49386279  12.39745805  12.30133133
#> [346]  12.20548102  12.10990554  12.01460331  11.91957279  11.82481241
#> [351]  11.73032066  11.63609602  11.54213697  11.44844203  11.35500971
#> [356]  11.26183854  11.16892708  11.07627388  10.98387749  10.89173652
#> [361]  10.79984954  10.70821516  10.61683200  10.52569867  10.43481383
#> [366]  10.34417612  10.25378419  10.16363672  10.07373239   9.98406989
#> [371]   9.89464793   9.80546521   9.71652047   9.62781242   9.53933983
#> [376]   9.45110143   9.36309600   9.27532231   9.18777913   9.10046527
#> [381]   9.01337952   8.92652070   8.83988762   8.75347912   8.66729403
#> [386]   8.58133120   8.49558949   8.41006776   8.32476489   8.23967977
#> [391]   8.15481127   8.07015830   7.98571978   7.90149461   7.81748172
#> [396]   7.73368004   7.65008852   7.56670609   7.48353173   7.40056439
#> [401]   7.31780304   7.23524667   7.15289426   7.07074481   6.98879731
#> [406]   6.90705078   6.82550424   6.74415671   6.66300721   6.58205479
#> [411]   6.50129850   6.42073738   6.34037049   6.26019690   6.18021569
#> [416]   6.10042592   6.02082669   5.94141709   5.86219621   5.78316317
#> [421]   5.70431707   5.62565704   5.54718218   5.46889165   5.39078456
#> [426]   5.31286007   5.23511732   5.15755547   5.08017367   5.00297109
#> [431]   4.92594691   4.84910030   4.77243045   4.69593654   4.61961776
#> [436]   4.54347333   4.46750243   4.39170429   4.31607812   4.24062315
#> [441]   4.16533858   4.09022367   4.01527765   3.94049976   3.86588924
#> [446]   3.79144535   3.71716735   3.64305449   3.56910605   3.49532130
#> [451]   3.42169952   3.34823998   3.27494198   3.20180480   3.12882773
#> [456]   3.05601009   2.98335118   2.91085030   2.83850677   2.76631991
#> [461]   2.69428904   2.62241348   2.55069258   2.47912566   2.40771207
#> [466]   2.33645114   2.26534224   2.19438470   2.12357790   2.05292118
#> [471]   1.98241392   1.91205548   1.84184524   1.77178257   1.70186686
#> [476]   1.63209749   1.56247384   1.49299531   1.42366131   1.35447121
#> [481]   1.28542444   1.21652040   1.14775850   1.07913815   1.01065877
#> [486]   0.94231978   0.87412062   0.80606071   0.73813947   0.67035636
#> [491]   0.60271080   0.53520224   0.46783013   0.40059391   0.33349304
#> [496]   0.26652698   0.19969518   0.13299710   0.06643222   0.00000000

# By default the function uses the reference light
# environment from (Laverick et al., 2020) with no shade.

   depth()

#>   [1] 100.71906204  95.37501248  91.16198440  87.68386618  84.72194926
#>   [6]  82.14262754  79.85827202  77.80828946  75.94903068  74.24800817
#>  [11]  72.68038764  71.22675817  69.87165969  68.60257852  67.40924320
#>  [16]  66.28311890  65.21703667  64.20491672  63.24155838  62.32247838
#>  [21]  61.44378467  60.60207682  59.79436659  59.01801386  58.27067458
#>  [26]  57.55025809  56.85489189  56.18289225  55.53273971  54.90305837
#>  [31]  54.29259833  53.70022079  53.12488525  52.56563852  52.02160520
#>  [36]  51.49197944  50.97601779  50.47303289  49.98238797  49.50349202
#>  [41]  49.03579549  48.57878652  48.13198753  47.69495229  47.26726317
#>  [46]  46.84852879  46.43838185  46.03647720  45.64249006  45.25611452
#>  [51]  44.87706203  44.50506016  44.13985140  43.78119210  43.42885147
#>  [56]  43.08261073  42.74226226  42.40760887  42.07846309  41.75464657
#>  [61]  41.43598948  41.12232999  40.81351375  40.50939347  40.20982847
#>  [66]  39.91468428  39.62383234  39.33714959  39.05451818  38.77582520
#>  [71]  38.50096240  38.22982593  37.96231609  37.69833715  37.43779712
#>  [76]  37.18060755  36.92668335  36.67594264  36.42830661  36.18369928
#>  [81]  35.94204749  35.70328065  35.46733070  35.23413194  35.00362095
#>  [86]  34.77573648  34.55041936  34.32761238  34.10726021  33.88930937
#>  [91]  33.67370808  33.46040622  33.24935526  33.04050818  32.83381944
#>  [96]  32.62924489  32.42674170  32.22626835  32.02778456  31.83125121
#> [101]  31.63663036  31.44388513  31.25297972  31.06387934  30.87655017
#> [106]  30.69095935  30.50707489  30.32486573  30.14430159  29.96535306
#> [111]  29.78799146  29.61218890  29.43791821  29.26515293  29.09386727
#> [116]  28.92403611  28.75563493  28.58863988  28.42302765  28.25877554
#> [121]  28.09586139  27.93426357  27.77396099  27.61493304  27.45715962
#> [126]  27.30062108  27.14529825  26.99117240  26.83822520  26.68643877
#> [131]  26.53579563  26.38627868  26.23787121  26.09055687  25.94431967
#> [136]  25.79914399  25.65501451  25.51191627  25.36983462  25.22875521
#> [141]  25.08866400  24.94954726  24.81139150  24.67418355  24.53791049
#> [146]  24.40255967  24.26811869  24.13457539  24.00191787  23.87013446
#> [151]  23.73921372  23.60914441  23.47991554  23.35151632  23.22393617
#> [156]  23.09716469  22.97119170  22.84600721  22.72160140  22.59796466
#> [161]  22.47508752  22.35296071  22.23157513  22.11092183  21.99099203
#> [166]  21.87177710  21.75326858  21.63545813  21.51833758  21.40189891
#> [171]  21.28613421  21.17103573  21.05659584  20.94280706  20.82966201
#> [176]  20.71715345  20.60527426  20.49401744  20.38337609  20.27334346
#> [181]  20.16391288  20.05507779  19.94683176  19.83916843  19.73208157
#> [186]  19.62556503  19.51961278  19.41421887  19.30937744  19.20508274
#> [191]  19.10132908  18.99811088  18.89542266  18.79325898  18.69161452
#> [196]  18.59048404  18.48986235  18.38974436  18.29012506  18.19099950
#> [201]  18.09236281  17.99421019  17.89653692  17.79933834  17.70260985
#> [206]  17.60634693  17.51054512  17.41520002  17.32030730  17.22586269
#> [211]  17.13186197  17.03830099  16.94517566  16.85248193  16.76021583
#> [216]  16.66837343  16.57695085  16.48594428  16.39534994  16.30516411
#> [221]  16.21538314  16.12600340  16.03702131  15.94843337  15.86023609
#> [226]  15.77242603  15.68499983  15.59795412  15.51128563  15.42499108
#> [231]  15.33906728  15.25351104  15.16831923  15.08348877  14.99901660
#> [236]  14.91489971  14.83113512  14.74771989  14.66465112  14.58192595
#> [241]  14.49954154  14.41749509  14.33578386  14.25440510  14.17335613
#> [246]  14.09263428  14.01223691  13.93216145  13.85240531  13.77296595
#> [251]  13.69384088  13.61502762  13.53652371  13.45832675  13.38043433
#> [256]  13.30284409  13.22555371  13.14856086  13.07186328  12.99545870
#> [261]  12.91934489  12.84351964  12.76798079  12.69272617  12.61775365
#> [266]  12.54306112  12.46864651  12.39450774  12.32064279  12.24704964
#> [271]  12.17372629  12.10067078  12.02788115  11.95535548  11.88309185
#> [276]  11.81108839  11.73934322  11.66785450  11.59662041  11.52563913
#> [281]  11.45490887  11.38442787  11.31419438  11.24420666  11.17446301
#> [286]  11.10496171  11.03570111  10.96667952  10.89789532  10.82934687
#> [291]  10.76103256  10.69295080  10.62510001  10.55747863  10.49008511
#> [296]  10.42291792  10.35597555  10.28925650  10.22275928  10.15648242
#> [301]  10.09042446  10.02458397   9.95895952   9.89354969   9.82835309
#> [306]   9.76336832   9.69859403   9.63402884   9.56967141   9.50552042
#> [311]   9.44157453   9.37783245   9.31429287   9.25095452   9.18781613
#> [316]   9.12487643   9.06213417   8.99958814   8.93723709   8.87507982
#> [321]   8.81311513   8.75134183   8.68975874   8.62836469   8.56715852
#> [326]   8.50613909   8.44530527   8.38465592   8.32418993   8.26390620
#> [331]   8.20380363   8.14388114   8.08413764   8.02457208   7.96518339
#> [336]   7.90597053   7.84693246   7.78806816   7.72937659   7.67085676
#> [341]   7.61250766   7.55432829   7.49631767   7.43847483   7.38079880
#> [346]   7.32328861   7.26594332   7.20876199   7.15174367   7.09488745
#> [351]   7.03819240   6.98165761   6.92528218   6.86906522   6.81300582
#> [356]   6.75710313   6.70135625   6.64576433   6.59032650   6.53504191
#> [361]   6.47990972   6.42492910   6.37009920   6.31541920   6.26088830
#> [366]   6.20650567   6.15227051   6.09818203   6.04423943   5.99044194
#> [371]   5.93678876   5.88327913   5.82991228   5.77668745   5.72360390
#> [376]   5.67066086   5.61785760   5.56519338   5.51266748   5.46027916
#> [381]   5.40802771   5.35591242   5.30393257   5.25208747   5.20037642
#> [386]   5.14879872   5.09735369   5.04604066   4.99485894   4.94380786
#> [391]   4.89288676   4.84209498   4.79143187   4.74089676   4.69048903
#> [396]   4.64020802   4.59005311   4.54002366   4.49011904   4.44033863
#> [401]   4.39068183   4.34114800   4.29173656   4.24244688   4.19327839
#> [406]   4.14423047   4.09530254   4.04649402   3.99780433   3.94923288
#> [411]   3.90077910   3.85244243   3.80422229   3.75611814   3.70812941
#> [416]   3.66025555   3.61249601   3.56485025   3.51731773   3.46989790
#> [421]   3.42259024   3.37539422   3.32830931   3.28133499   3.23447074
#> [426]   3.18771604   3.14107039   3.09453328   3.04810420   3.00178266
#> [431]   2.95556815   2.90946018   2.86345827   2.81756192   2.77177066
#> [436]   2.72608400   2.68050146   2.63502258   2.58964687   2.54437389
#> [441]   2.49920315   2.45413420   2.40916659   2.36429985   2.31953354
#> [446]   2.27486721   2.23030041   2.18583270   2.14146363   2.09719278
#> [451]   2.05301971   2.00894399   1.96496519   1.92108288   1.87729664
#> [456]   1.83360606   1.79001071   1.74651018   1.70310406   1.65979195
#> [461]   1.61657342   1.57344809   1.53041555   1.48747540   1.44462724
#> [466]   1.40187069   1.35920534   1.31663082   1.27414674   1.23175271
#> [471]   1.18944835   1.14723329   1.10510714   1.06306954   1.02112012
#> [476]   0.97925849   0.93748430   0.89579719   0.85419678   0.81268273
#> [481]   0.77125467   0.72991224   0.68865510   0.64748289   0.60639526
#> [486]   0.56539187   0.52447237   0.48363642   0.44288368   0.40221382
#> [491]   0.36162648   0.32112135   0.28069808   0.24035635   0.20009583
#> [496]   0.15991619   0.11981711   0.07979826   0.03985933   0.00000000