Calculate the mesophotic community value for a given light level

This function captures the preferred light environment of the mesophotic coral community.

mesophotic(light = seq(0.01, 100, length = 500), a = 1.54, b = 8.92)

Arguments

light
The percentage of surface irradiance (PAR) reaching a reef, positive and up to 100%.

a
The scale parameter of a Weibull distribution, defaults to (Tamir et al., 2019). a describes the spread of the data, or ‘the peakyness’. Larger numbers create a flatter, wider, curve. Varying a will simultaneously change the width of the preferred light environment of mesophotic taxa, and the maximum similarity a reef patch is expected to achieve in comparison to the idealised mesophotic community.

b
The shape parameter of a Weibull distribution, defaults to (Tamir et al., 2019). A value of 3 approximates a normal distribution. Values larger than 3 introduce a left-skew to the curve. This means mesophotic taxa can be light-limited, while keeping mesophotic communities in low-light environments. Varying b shifts the preferred light environment of mesophotic communities.

Value

A mesophotic community indicator value, as discussed in (Laverick et al., 2017). When a single light value is provided, a single value is returned. You can pass a vector of light levels for a vector of depths.

Details

A Weibull distribution is used to represent the relationship between mesophotic coral communities and light. We assume mesophotic community values (how similar a reef patch is to the idealised mesophotic community) center on a preferred light value, between light limitation and light-induced stress.

This relationship has been verified in the Red Sea (Tamir et al., 2019), but certainly needs more research!

References

(Tamir et al., 2019)

(Laverick et al., 2017)

See also

Examples

# Specifying a single light value
# returns a single community value.

   mesophotic(light = 10)

#> [1] 0

# Specifying a vector of light values
# returns a vector of community values.

   mesophotic(light = seq(0.0065, 100, length = 500))

#>   [1]  4.123242e-02  7.045382e-03  1.232628e-05  2.234118e-09  6.631360e-14
#>   [6]  4.125789e-19  6.328740e-25  2.692839e-31  3.480451e-38  1.469821e-45
#>  [11]  2.154189e-53  1.153020e-61  2.354937e-70  1.906975e-79  6.333234e-89
#>  [16]  8.890471e-99 5.420378e-109 1.471022e-119 1.817312e-130 1.043248e-141
#>  [21] 2.836070e-153 3.715595e-165 2.384470e-177 7.610256e-190 1.225225e-202
#>  [26] 1.008383e-215 4.296047e-229 9.586847e-243 1.133169e-256 7.169965e-271
#>  [31] 2.453020e-285 4.581303e-300 4.713312e-315  0.000000e+00  0.000000e+00
#>  [36]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [41]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [46]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [51]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [56]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [61]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [66]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [71]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [76]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [81]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [86]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [91]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [96]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [101]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [106]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [111]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [116]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [121]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [126]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [131]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [136]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [141]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [146]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [151]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [156]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [161]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [166]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [171]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [176]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [181]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [186]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [191]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [196]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [201]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [206]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [211]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [216]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [221]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [226]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [231]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [236]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [241]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [246]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [251]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [256]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [261]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [266]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [271]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [276]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [281]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [286]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [291]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [296]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [301]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [306]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [311]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [316]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [321]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [326]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [331]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [336]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [341]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [346]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [351]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [356]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [361]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [366]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [371]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [376]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [381]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [386]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [391]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [396]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [401]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [406]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [411]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [416]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [421]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [426]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [431]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [436]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [441]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [446]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [451]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [456]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [461]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [466]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [471]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [476]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [481]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [486]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [491]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [496]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00

# By default the function uses the shallow community
# light relationship from (Tamir et al., 2019) for a
# range of light values.

   mesophotic()

#>   [1] 4.405005e-03 2.275244e-02 3.247154e-02 3.994654e-02 4.614805e-02
#>   [6] 5.146640e-02 5.610681e-02 6.019408e-02 6.381155e-02 6.701885e-02
#>  [11] 6.986088e-02 7.237290e-02 7.458357e-02 7.651684e-02 7.819322e-02
#>  [16] 7.963065e-02 8.084507e-02 8.185084e-02 8.266110e-02 8.328795e-02
#>  [21] 8.374265e-02 8.403573e-02 8.417715e-02 8.417628e-02 8.404207e-02
#>  [26] 8.378301e-02 8.340723e-02 8.292250e-02 8.233625e-02 8.165563e-02
#>  [31] 8.088749e-02 8.003838e-02 7.911462e-02 7.812226e-02 7.706709e-02
#>  [36] 7.595470e-02 7.479040e-02 7.357932e-02 7.232633e-02 7.103610e-02
#>  [41] 6.971310e-02 6.836157e-02 6.698558e-02 6.558897e-02 6.417541e-02
#>  [46] 6.274837e-02 6.131115e-02 5.986684e-02 5.841839e-02 5.696857e-02
#>  [51] 5.551996e-02 5.407503e-02 5.263603e-02 5.120512e-02 4.978427e-02
#>  [56] 4.837533e-02 4.698000e-02 4.559986e-02 4.423635e-02 4.289081e-02
#>  [61] 4.156443e-02 4.025831e-02 3.897344e-02 3.771070e-02 3.647087e-02
#>  [66] 3.525464e-02 3.406262e-02 3.289531e-02 3.175314e-02 3.063649e-02
#>  [71] 2.954562e-02 2.848075e-02 2.744203e-02 2.642956e-02 2.544337e-02
#>  [76] 2.448342e-02 2.354967e-02 2.264198e-02 2.176020e-02 2.090413e-02
#>  [81] 2.007354e-02 1.926815e-02 1.848766e-02 1.773174e-02 1.700005e-02
#>  [86] 1.629220e-02 1.560779e-02 1.494641e-02 1.430763e-02 1.369099e-02
#>  [91] 1.309604e-02 1.252231e-02 1.196932e-02 1.143658e-02 1.092361e-02
#>  [96] 1.042991e-02 9.954975e-03 9.498313e-03 9.059423e-03 8.637807e-03
#> [101] 8.232967e-03 7.844409e-03 7.471643e-03 7.114184e-03 6.771550e-03
#> [106] 6.443266e-03 6.128864e-03 5.827883e-03 5.539869e-03 5.264375e-03
#> [111] 5.000965e-03 4.749208e-03 4.508684e-03 4.278984e-03 4.059704e-03
#> [116] 3.850453e-03 3.650849e-03 3.460518e-03 3.279099e-03 3.106237e-03
#> [121] 2.941591e-03 2.784827e-03 2.635622e-03 2.493662e-03 2.358645e-03
#> [126] 2.230275e-03 2.108269e-03 1.992351e-03 1.882256e-03 1.777727e-03
#> [131] 1.678517e-03 1.584387e-03 1.495107e-03 1.410455e-03 1.330218e-03
#> [136] 1.254191e-03 1.182176e-03 1.113984e-03 1.049433e-03 9.883477e-04
#> [141] 9.305608e-04 8.759116e-04 8.242459e-04 7.754160e-04 7.292808e-04
#> [146] 6.857050e-04 6.445592e-04 6.057199e-04 5.690688e-04 5.344932e-04
#> [151] 5.018854e-04 4.711424e-04 4.421664e-04 4.148638e-04 3.891455e-04
#> [156] 3.649267e-04 3.421267e-04 3.206684e-04 3.004789e-04 2.814884e-04
#> [161] 2.636309e-04 2.468436e-04 2.310668e-04 2.162437e-04 2.023208e-04
#> [166] 1.892469e-04 1.769736e-04 1.654552e-04 1.546481e-04 1.445113e-04
#> [171] 1.350057e-04 1.260944e-04 1.177425e-04 1.099170e-04 1.025868e-04
#> [176] 9.572219e-05 8.929542e-05 8.328012e-05 7.765142e-05 7.238587e-05
#> [181] 6.746130e-05 6.285685e-05 5.855282e-05 5.453063e-05 5.077280e-05
#> [186] 4.726286e-05 4.398528e-05 4.092545e-05 3.806964e-05 3.540491e-05
#> [191] 3.291911e-05 3.060078e-05 2.843920e-05 2.642426e-05 2.454647e-05
#> [196] 2.279694e-05 2.116731e-05 1.964972e-05 1.823682e-05 1.692170e-05
#> [201] 1.569791e-05 1.455936e-05 1.350037e-05 1.251563e-05 1.160014e-05
#> [206] 1.074924e-05 9.958555e-06 9.224001e-06 8.541753e-06 7.908236e-06
#> [211] 7.320106e-06 6.774240e-06 6.267717e-06 5.797812e-06 5.361977e-06
#> [216] 4.957837e-06 4.583171e-06 4.235911e-06 3.914124e-06 3.616011e-06
#> [221] 3.339891e-06 3.084202e-06 2.847483e-06 2.628378e-06 2.425621e-06
#> [226] 2.238034e-06 2.064521e-06 1.904062e-06 1.755707e-06 1.618574e-06
#> [231] 1.491843e-06 1.374749e-06 1.266583e-06 1.166687e-06 1.074449e-06
#> [236] 9.893001e-07 9.107123e-07 8.381958e-07 7.712962e-07 7.095916e-07
#> [241] 6.526909e-07 6.002312e-07 5.518764e-07 5.073146e-07 4.662571e-07
#> [246] 4.284364e-07 3.936045e-07 3.615322e-07 3.320070e-07 3.048322e-07
#> [251] 2.798261e-07 2.568204e-07 2.356594e-07 2.161993e-07 1.983070e-07
#> [256] 1.818597e-07 1.667438e-07 1.528543e-07 1.400945e-07 1.283747e-07
#> [261] 1.176125e-07 1.077315e-07 9.866159e-08 9.033776e-08 8.270022e-08
#> [266] 7.569380e-08 6.926764e-08 6.337487e-08 5.797230e-08 5.302015e-08
#> [271] 4.848176e-08 4.432341e-08 4.051401e-08 3.702499e-08 3.383004e-08
#> [276] 3.090493e-08 2.822742e-08 2.577702e-08 2.353491e-08 2.148380e-08
#> [281] 1.960776e-08 1.789221e-08 1.632370e-08 1.488993e-08 1.357956e-08
#> [286] 1.238221e-08 1.128835e-08 1.028921e-08 9.376779e-09 8.543685e-09
#> [291] 7.783177e-09 7.089063e-09 6.455667e-09 5.877788e-09 5.350661e-09
#> [296] 4.869920e-09 4.431564e-09 4.031934e-09 3.667677e-09 3.335723e-09
#> [301] 3.033265e-09 2.757735e-09 2.506780e-09 2.278253e-09 2.070188e-09
#> [306] 1.880787e-09 1.708409e-09 1.551552e-09 1.408846e-09 1.279038e-09
#> [311] 1.160983e-09 1.053638e-09 9.560490e-10 8.673451e-10 7.867323e-10
#> [316] 7.134860e-10 6.469453e-10 5.865071e-10 5.316218e-10 4.817883e-10
#> [321] 4.365497e-10 3.954898e-10 3.582294e-10 3.244229e-10 2.937557e-10
#> [326] 2.659413e-10 2.407188e-10 2.178508e-10 1.971211e-10 1.783332e-10
#> [331] 1.613083e-10 1.458836e-10 1.319112e-10 1.192565e-10 1.077975e-10
#> [336] 9.742283e-11 8.803164e-11 7.953217e-11 7.184110e-11 6.488277e-11
#> [341] 5.858846e-11 5.289581e-11 4.774819e-11 4.309425e-11 3.888736e-11
#> [346] 3.508524e-11 3.164954e-11 2.854549e-11 2.574155e-11 2.320916e-11
#> [351] 2.092239e-11 1.885779e-11 1.699409e-11 1.531202e-11 1.379416e-11
#> [356] 1.242469e-11 1.118933e-11 1.007513e-11 9.070385e-12 8.164487e-12
#> [361] 7.347853e-12 6.611813e-12 5.948525e-12 5.350899e-12 4.812525e-12
#> [366] 4.327611e-12 3.890920e-12 3.497725e-12 3.143751e-12 2.825139e-12
#> [371] 2.538406e-12 2.280403e-12 2.048292e-12 1.839509e-12 1.651739e-12
#> [376] 1.482897e-12 1.331099e-12 1.194648e-12 1.072012e-12 9.618107e-13
#> [381] 8.627994e-13 7.738566e-13 6.939715e-13 6.222335e-13 5.578222e-13
#> [386] 4.999989e-13 4.480981e-13 4.015209e-13 3.597279e-13 3.222340e-13
#> [391] 2.886023e-13 2.584398e-13 2.313932e-13 2.071444e-13 1.854075e-13
#> [396] 1.659255e-13 1.484672e-13 1.328250e-13 1.188122e-13 1.062611e-13
#> [401] 9.502102e-14 8.495661e-14 7.594634e-14 6.788110e-14 6.066292e-14
#> [406] 5.420386e-14 4.842501e-14 4.325555e-14 3.863196e-14 3.449725e-14
#> [411] 3.080031e-14 2.749531e-14 2.454117e-14 2.190105e-14 1.954195e-14
#> [416] 1.743429e-14 1.555156e-14 1.387002e-14 1.236840e-14 1.102767e-14
#> [421] 9.830777e-15 8.762450e-15 7.809030e-15 6.958290e-15 6.199291e-15
#> [426] 5.522244e-15 4.918393e-15 4.379910e-15 3.899791e-15 3.471778e-15
#> [431] 3.090274e-15 2.750278e-15 2.447320e-15 2.177407e-15 1.936971e-15
#> [436] 1.722827e-15 1.532127e-15 1.362333e-15 1.211174e-15 1.076626e-15
#> [441] 9.568822e-16 8.503297e-16 7.555297e-16 6.711988e-16 5.961921e-16
#> [446] 5.294889e-16 4.701789e-16 4.174505e-16 3.705806e-16 3.289245e-16
#> [451] 2.919077e-16 2.590186e-16 2.298012e-16 2.038496e-16 1.808021e-16
#> [456] 1.603369e-16 1.421673e-16 1.260382e-16 1.117227e-16 9.901861e-17
#> [461] 8.774633e-17 7.774594e-17 6.887526e-17 6.100782e-17 5.403120e-17
#> [466] 4.784545e-17 4.236173e-17 3.750108e-17 3.319333e-17 2.937617e-17
#> [471] 2.599422e-17 2.299830e-17 2.034473e-17 1.799474e-17 1.591391e-17
#> [476] 1.407167e-17 1.244091e-17 1.099755e-17 9.720261e-18 8.590087e-18
#> [481] 7.590233e-18 6.705800e-18 5.923577e-18 5.231854e-18 4.620248e-18
#> [486] 4.079559e-18 3.601632e-18 3.179243e-18 2.805993e-18 2.476212e-18
#> [491] 2.184880e-18 1.927551e-18 1.700290e-18 1.499611e-18 1.322431e-18
#> [496] 1.166020e-18 1.027965e-18 9.061270e-19 7.986179e-19 7.037656e-19

# You can specify your own community light relationship
# by setting parameters explicitly

   mesophotic(a = 1.2, b = 8)

#>   [1] 3.938498e-02 7.154145e-02 8.051516e-02 8.567735e-02 8.901373e-02
#>   [6] 9.123741e-02 9.269809e-02 9.359780e-02 9.406598e-02 9.419174e-02
#>  [11] 9.403971e-02 9.365870e-02 9.308672e-02 9.235415e-02 9.148575e-02
#>  [16] 9.050207e-02 8.942037e-02 8.825532e-02 8.701955e-02 8.572397e-02
#>  [21] 8.437812e-02 8.299035e-02 8.156802e-02 8.011766e-02 7.864506e-02
#>  [26] 7.715538e-02 7.565323e-02 7.414274e-02 7.262760e-02 7.111113e-02
#>  [31] 6.959629e-02 6.808576e-02 6.658192e-02 6.508694e-02 6.360273e-02
#>  [36] 6.213101e-02 6.067332e-02 5.923103e-02 5.780535e-02 5.639737e-02
#>  [41] 5.500804e-02 5.363818e-02 5.228852e-02 5.095970e-02 4.965224e-02
#>  [46] 4.836661e-02 4.710319e-02 4.586228e-02 4.464413e-02 4.344893e-02
#>  [51] 4.227681e-02 4.112786e-02 4.000211e-02 3.889956e-02 3.782018e-02
#>  [56] 3.676389e-02 3.573058e-02 3.472012e-02 3.373234e-02 3.276706e-02
#>  [61] 3.182407e-02 3.090315e-02 3.000405e-02 2.912651e-02 2.827026e-02
#>  [66] 2.743502e-02 2.662048e-02 2.582634e-02 2.505229e-02 2.429800e-02
#>  [71] 2.356314e-02 2.284738e-02 2.215038e-02 2.147180e-02 2.081130e-02
#>  [76] 2.016852e-02 1.954313e-02 1.893477e-02 1.834310e-02 1.776777e-02
#>  [81] 1.720844e-02 1.666475e-02 1.613637e-02 1.562296e-02 1.512417e-02
#>  [86] 1.463969e-02 1.416916e-02 1.371227e-02 1.326869e-02 1.283810e-02
#>  [91] 1.242019e-02 1.201464e-02 1.162115e-02 1.123941e-02 1.086913e-02
#>  [96] 1.051002e-02 1.016178e-02 9.824128e-03 9.496794e-03 9.179503e-03
#> [101] 8.871986e-03 8.573980e-03 8.285228e-03 8.005479e-03 7.734483e-03
#> [106] 7.472001e-03 7.217794e-03 6.971631e-03 6.733285e-03 6.502535e-03
#> [111] 6.279164e-03 6.062960e-03 5.853717e-03 5.651232e-03 5.455309e-03
#> [116] 5.265755e-03 5.082382e-03 4.905008e-03 4.733454e-03 4.567545e-03
#> [121] 4.407112e-03 4.251990e-03 4.102018e-03 3.957038e-03 3.816899e-03
#> [126] 3.681451e-03 3.550549e-03 3.424054e-03 3.301827e-03 3.183736e-03
#> [131] 3.069651e-03 2.959445e-03 2.852997e-03 2.750187e-03 2.650900e-03
#> [136] 2.555023e-03 2.462446e-03 2.373064e-03 2.286774e-03 2.203476e-03
#> [141] 2.123072e-03 2.045468e-03 1.970573e-03 1.898297e-03 1.828556e-03
#> [146] 1.761265e-03 1.696343e-03 1.633711e-03 1.573294e-03 1.515018e-03
#> [151] 1.458810e-03 1.404602e-03 1.352327e-03 1.301918e-03 1.253313e-03
#> [156] 1.206451e-03 1.161272e-03 1.117720e-03 1.075738e-03 1.035273e-03
#> [161] 9.962719e-04 9.586854e-04 9.224644e-04 8.875615e-04 8.539311e-04
#> [166] 8.215289e-04 7.903122e-04 7.602395e-04 7.312709e-04 7.033676e-04
#> [171] 6.764922e-04 6.506084e-04 6.256813e-04 6.016770e-04 5.785628e-04
#> [176] 5.563072e-04 5.348794e-04 5.142501e-04 4.943906e-04 4.752735e-04
#> [181] 4.568720e-04 4.391604e-04 4.221139e-04 4.057085e-04 3.899210e-04
#> [186] 3.747291e-04 3.601110e-04 3.460459e-04 3.325138e-04 3.194951e-04
#> [191] 3.069711e-04 2.949236e-04 2.833352e-04 2.721890e-04 2.614687e-04
#> [196] 2.511586e-04 2.412436e-04 2.317091e-04 2.225408e-04 2.137254e-04
#> [201] 2.052495e-04 1.971006e-04 1.892665e-04 1.817354e-04 1.744960e-04
#> [206] 1.675373e-04 1.608488e-04 1.544203e-04 1.482421e-04 1.423047e-04
#> [211] 1.365990e-04 1.311163e-04 1.258480e-04 1.207861e-04 1.159227e-04
#> [216] 1.112503e-04 1.067616e-04 1.024495e-04 9.830742e-05 9.432872e-05
#> [221] 9.050717e-05 8.683675e-05 8.331165e-05 7.992627e-05 7.667524e-05
#> [226] 7.355336e-05 7.055565e-05 6.767730e-05 6.491369e-05 6.226037e-05
#> [231] 5.971305e-05 5.726762e-05 5.492009e-05 5.266666e-05 5.050365e-05
#> [236] 4.842753e-05 4.643489e-05 4.452246e-05 4.268710e-05 4.092578e-05
#> [241] 3.923558e-05 3.761371e-05 3.605747e-05 3.456427e-05 3.313161e-05
#> [246] 3.175711e-05 3.043845e-05 2.917343e-05 2.795991e-05 2.679585e-05
#> [251] 2.567928e-05 2.460829e-05 2.358109e-05 2.259591e-05 2.165108e-05
#> [256] 2.074498e-05 1.987607e-05 1.904284e-05 1.824386e-05 1.747777e-05
#> [261] 1.674323e-05 1.603897e-05 1.536378e-05 1.471647e-05 1.409593e-05
#> [266] 1.350107e-05 1.293084e-05 1.238426e-05 1.186035e-05 1.135820e-05
#> [271] 1.087693e-05 1.041568e-05 9.973644e-06 9.550029e-06 9.144086e-06
#> [276] 8.755093e-06 8.382356e-06 8.025209e-06 7.683014e-06 7.355156e-06
#> [281] 7.041047e-06 6.740122e-06 6.451838e-06 6.175674e-06 5.911130e-06
#> [286] 5.657727e-06 5.415005e-06 5.182521e-06 4.959853e-06 4.746593e-06
#> [291] 4.542352e-06 4.346754e-06 4.159442e-06 3.980069e-06 3.808307e-06
#> [296] 3.643838e-06 3.486358e-06 3.335575e-06 3.191210e-06 3.052994e-06
#> [301] 2.920670e-06 2.793991e-06 2.672721e-06 2.556632e-06 2.445508e-06
#> [306] 2.339139e-06 2.237326e-06 2.139877e-06 2.046607e-06 1.957341e-06
#> [311] 1.871910e-06 1.790152e-06 1.711911e-06 1.637039e-06 1.565393e-06
#> [316] 1.496836e-06 1.431237e-06 1.368472e-06 1.308418e-06 1.250962e-06
#> [321] 1.195992e-06 1.143403e-06 1.093093e-06 1.044965e-06 9.989264e-07
#> [326] 9.548871e-07 9.127619e-07 8.724689e-07 8.339298e-07 7.970692e-07
#> [331] 7.618154e-07 7.280992e-07 6.958546e-07 6.650184e-07 6.355301e-07
#> [336] 6.073315e-07 5.803671e-07 5.545837e-07 5.299303e-07 5.063582e-07
#> [341] 4.838206e-07 4.622728e-07 4.416720e-07 4.219771e-07 4.031488e-07
#> [346] 3.851497e-07 3.679437e-07 3.514963e-07 3.357747e-07 3.207471e-07
#> [351] 3.063835e-07 2.926549e-07 2.795335e-07 2.669930e-07 2.550080e-07
#> [356] 2.435542e-07 2.326083e-07 2.221483e-07 2.121527e-07 2.026013e-07
#> [361] 1.934746e-07 1.847540e-07 1.764216e-07 1.684603e-07 1.608540e-07
#> [366] 1.535869e-07 1.466442e-07 1.400115e-07 1.336751e-07 1.276221e-07
#> [371] 1.218400e-07 1.163166e-07 1.110407e-07 1.060013e-07 1.011879e-07
#> [376] 9.659044e-08 9.219946e-08 8.800576e-08 8.400060e-08 8.017560e-08
#> [381] 7.652277e-08 7.303445e-08 6.970332e-08 6.652239e-08 6.348496e-08
#> [386] 6.058466e-08 5.781536e-08 5.517121e-08 5.264663e-08 5.023629e-08
#> [391] 4.793506e-08 4.573808e-08 4.364068e-08 4.163839e-08 3.972696e-08
#> [396] 3.790231e-08 3.616055e-08 3.449796e-08 3.291098e-08 3.139621e-08
#> [401] 2.995041e-08 2.857048e-08 2.725344e-08 2.599646e-08 2.479684e-08
#> [406] 2.365199e-08 2.255944e-08 2.151682e-08 2.052188e-08 1.957247e-08
#> [411] 1.866652e-08 1.780206e-08 1.697723e-08 1.619021e-08 1.543930e-08
#> [416] 1.472286e-08 1.403933e-08 1.338720e-08 1.276506e-08 1.217154e-08
#> [421] 1.160533e-08 1.106520e-08 1.054995e-08 1.005845e-08 9.589623e-09
#> [426] 9.142429e-09 8.715882e-09 8.309038e-09 7.920997e-09 7.550900e-09
#> [431] 7.197924e-09 6.861288e-09 6.540241e-09 6.234070e-09 5.942093e-09
#> [436] 5.663659e-09 5.398145e-09 5.144959e-09 4.903534e-09 4.673329e-09
#> [441] 4.453828e-09 4.244539e-09 4.044992e-09 3.854737e-09 3.673346e-09
#> [446] 3.500410e-09 3.335540e-09 3.178363e-09 3.028523e-09 2.885681e-09
#> [451] 2.749514e-09 2.619713e-09 2.495983e-09 2.378044e-09 2.265626e-09
#> [456] 2.158474e-09 2.056343e-09 1.959001e-09 1.866226e-09 1.777804e-09
#> [461] 1.693534e-09 1.613222e-09 1.536685e-09 1.463747e-09 1.394239e-09
#> [466] 1.328004e-09 1.264886e-09 1.204743e-09 1.147433e-09 1.092826e-09
#> [471] 1.040795e-09 9.912198e-10 9.439852e-10 8.989819e-10 8.561054e-10
#> [476] 8.152563e-10 7.763394e-10 7.392643e-10 7.039446e-10 6.702979e-10
#> [481] 6.382457e-10 6.077132e-10 5.786288e-10 5.509246e-10 5.245357e-10
#> [486] 4.994002e-10 4.754590e-10 4.526559e-10 4.309374e-10 4.102522e-10
#> [491] 3.905517e-10 3.717893e-10 3.539209e-10 3.369041e-10 3.206988e-10
#> [496] 3.052666e-10 2.905709e-10 2.765769e-10 2.632514e-10 2.505627e-10