Calculate the whole-reef community value for a given light level

This function returns the community value for the whole reef community at a light level, given shallow and mesophotic community light relationships.

reef(
  light = seq(0.01, 1, length = 500),
  Vm = 0.393,
  K = 13.5,
  a = 1.54,
  b = 8.92
)

Arguments

light
The proportion of surface irradiance (PAR), between 0 and 1.

Vm
Passed to shallow. The maximum similarity of a reef patch to the idealised shallow reef community, between 0 and 1. Defaults to (Tamir et al., 2019).

K
Passed to shallow. The light value where the community value = Vmax/2, between 0 and 100. Defaults to (Tamir et al., 2019). This can be interpreted as the light level at which shallow reef communities become light limited.

a
Passed to mesophotic. The scale parameter of a Weibull distribution, defaults to (Tamir et al., 2019). a describes the spread of the data, or ‘the peakyness’. Larger numbers create a flatter, wider, curve. Varying a will simultaneously change the width of the preferred light environment of mesophotic taxa, and the maximum similarity a reef patch is expected to achieve in comparison to the idealised mesophotic community.

b
Passed to mesophotic. The shape parameter of a Weibull distribution, defaults to (Tamir et al., 2019). A value of 3 approximates a normal distribution. Values larger than 3 introduce a left-skew to the curve. This means mesophotic taxa can be light-limited, while keeping mesophotic communities in low-light environments. Varying b shifts the preferred light environment of mesophotic communities.

Value

A reef wide community indicator value. This is the basis of the model presented in (Laverick et al., 2020). When a single light value is provided, a single value is returned. You can pass a vector of light levels for a vector of depths.

Details

The function assumes a tradeoff, where an observed patch of reef can approach a mesophotic or shallow community. A patch of reef cannot simultaneously look like the exemplar mesophotic and shallow reef patch. This is achieved by subtracting the shallow community value from the mesophotic community value.

Essentially this function is a wrapper for mesophotic() - shallow(), but it is neccessary for use in boundary when calculating the light level the communities switch at.

References

(Tamir et al., 2019)

(Laverick et al., 2020)

See also

Examples

# Specifying a single light value
# returns a single community value.

   reef(light = 10)

#> [1] -0.3877652

# Specifying a vector of light values
# returns a vector of community values.

   reef(light = seq(0.0065, 1, length = 500))

#>   [1]  0.0231794162  0.0239493283  0.0240371209  0.0236189587  0.0228025766
#>   [6]  0.0216596516  0.0202408890  0.0185839110  0.0167177412  0.0146655162
#>  [11]  0.0124462091  0.0100757669  0.0075678866  0.0049345572  0.0021864474
#>  [16] -0.0006668125 -0.0036164231 -0.0066542599 -0.0097727563 -0.0129648142
#>  [21] -0.0162237360 -0.0195431713 -0.0229170764 -0.0263396828 -0.0298054720
#>  [26] -0.0333091565 -0.0368456647 -0.0404101287 -0.0439978753 -0.0476044180
#>  [31] -0.0512254517 -0.0548568469 -0.0584946463 -0.0621350614 -0.0657744691
#>  [36] -0.0694094091 -0.0730365817 -0.0766528452 -0.0802552135 -0.0838408540
#>  [41] -0.0874070849 -0.0909513729 -0.0944713307 -0.0979647141 -0.1014294195
#>  [46] -0.1048634809 -0.1082650665 -0.1116324762 -0.1149641381 -0.1182586046
#>  [51] -0.1215145500 -0.1247307661 -0.1279061588 -0.1310397448 -0.1341306472
#>  [56] -0.1371780924 -0.1401814059 -0.1431400085 -0.1460534125 -0.1489212180
#>  [61] -0.1517431088 -0.1545188489 -0.1572482782 -0.1599313095 -0.1625679239
#>  [66] -0.1651581680 -0.1677021497 -0.1702000348 -0.1726520437 -0.1750584478
#>  [71] -0.1774195662 -0.1797357627 -0.1820074423 -0.1842350483 -0.1864190596
#>  [76] -0.1885599871 -0.1906583718 -0.1927147816 -0.1947298088 -0.1967040677
#>  [81] -0.1986381920 -0.2005328329 -0.2023886567 -0.2042063427 -0.2059865812
#>  [86] -0.2077300719 -0.2094375215 -0.2111096428 -0.2127471525 -0.2143507699
#>  [91] -0.2159212154 -0.2174592095 -0.2189654710 -0.2204407162 -0.2218856579
#>  [96] -0.2233010040 -0.2246874569 -0.2260457125 -0.2273764596 -0.2286803788
#> [101] -0.2299581421 -0.2312104124 -0.2324378426 -0.2336410754 -0.2348207430
#> [106] -0.2359774662 -0.2371118546 -0.2382245061 -0.2393160066 -0.2403869299
#> [111] -0.2414378374 -0.2424692783 -0.2434817890 -0.2444758936 -0.2454521035
#> [116] -0.2464109174 -0.2473528216 -0.2482782897 -0.2491877830 -0.2500817505
#> [121] -0.2509606287 -0.2518248421 -0.2526748034 -0.2535109132 -0.2543335606
#> [126] -0.2551431233 -0.2559399677 -0.2567244489 -0.2574969115 -0.2582576893
#> [131] -0.2590071056 -0.2597454735 -0.2604730964 -0.2611902677 -0.2618972715
#> [136] -0.2625943823 -0.2632818661 -0.2639599797 -0.2646289716 -0.2652890818
#> [141] -0.2659405424 -0.2665835777 -0.2672184043 -0.2678452314 -0.2684642612
#> [146] -0.2690756888 -0.2696797028 -0.2702764852 -0.2708662118 -0.2714490522
#> [151] -0.2720251704 -0.2725947246 -0.2731578674 -0.2737147464 -0.2742655040
#> [156] -0.2748102776 -0.2753492001 -0.2758823996 -0.2764100001 -0.2769321211
#> [161] -0.2774488783 -0.2779603832 -0.2784667437 -0.2789680641 -0.2794644453
#> [166] -0.2799559846 -0.2804427762 -0.2809249113 -0.2814024781 -0.2818755618
#> [171] -0.2823442451 -0.2828086078 -0.2832687273 -0.2837246786 -0.2841765344
#> [176] -0.2846243650 -0.2850682388 -0.2855082218 -0.2859443782 -0.2863767705
#> [181] -0.2868054590 -0.2872305025 -0.2876519580 -0.2880698810 -0.2884843253
#> [186] -0.2888953432 -0.2893029858 -0.2897073026 -0.2901083420 -0.2905061507
#> [191] -0.2909007748 -0.2912922586 -0.2916806458 -0.2920659787 -0.2924482986
#> [196] -0.2928276459 -0.2932040601 -0.2935775795 -0.2939482417 -0.2943160834
#> [201] -0.2946811406 -0.2950434483 -0.2954030408 -0.2957599516 -0.2961142138
#> [206] -0.2964658593 -0.2968149198 -0.2971614261 -0.2975054084 -0.2978468964
#> [211] -0.2981859191 -0.2985225051 -0.2988566823 -0.2991884782 -0.2995179196
#> [216] -0.2998450330 -0.3001698444 -0.3004923793 -0.3008126627 -0.3011307193
#> [221] -0.3014465731 -0.3017602481 -0.3020717676 -0.3023811545 -0.3026884316
#> [226] -0.3029936210 -0.3032967447 -0.3035978242 -0.3038968807 -0.3041939351
#> [231] -0.3044890080 -0.3047821196 -0.3050732899 -0.3053625385 -0.3056498847
#> [236] -0.3059353477 -0.3062189462 -0.3065006988 -0.3067806237 -0.3070587389
#> [241] -0.3073350622 -0.3076096110 -0.3078824027 -0.3081534542 -0.3084227823
#> [246] -0.3086904037 -0.3089563345 -0.3092205910 -0.3094831891 -0.3097441444
#> [251] -0.3100034725 -0.3102611885 -0.3105173077 -0.3107718448 -0.3110248147
#> [256] -0.3112762318 -0.3115261104 -0.3117744647 -0.3120213087 -0.3122666561
#> [261] -0.3125105206 -0.3127529157 -0.3129938546 -0.3132333504 -0.3134714162
#> [266] -0.3137080646 -0.3139433085 -0.3141771603 -0.3144096323 -0.3146407367
#> [271] -0.3148704856 -0.3150988909 -0.3153259644 -0.3155517178 -0.3157761624
#> [276] -0.3159993097 -0.3162211709 -0.3164417572 -0.3166610794 -0.3168791484
#> [281] -0.3170959750 -0.3173115698 -0.3175259431 -0.3177391055 -0.3179510670
#> [286] -0.3181618379 -0.3183714282 -0.3185798477 -0.3187871063 -0.3189932136
#> [291] -0.3191981792 -0.3194020126 -0.3196047230 -0.3198063199 -0.3200068122
#> [296] -0.3202062092 -0.3204045197 -0.3206017526 -0.3207979167 -0.3209930206
#> [301] -0.3211870729 -0.3213800821 -0.3215720565 -0.3217630046 -0.3219529344
#> [306] -0.3221418541 -0.3223297718 -0.3225166953 -0.3227026327 -0.3228875916
#> [311] -0.3230715797 -0.3232546047 -0.3234366742 -0.3236177955 -0.3237979761
#> [316] -0.3239772233 -0.3241555443 -0.3243329463 -0.3245094364 -0.3246850216
#> [321] -0.3248597087 -0.3250335048 -0.3252064166 -0.3253784508 -0.3255496141
#> [326] -0.3257199131 -0.3258893543 -0.3260579441 -0.3262256891 -0.3263925955
#> [331] -0.3265586696 -0.3267239176 -0.3268883456 -0.3270519598 -0.3272147662
#> [336] -0.3273767707 -0.3275379792 -0.3276983976 -0.3278580318 -0.3280168873
#> [341] -0.3281749700 -0.3283322854 -0.3284888391 -0.3286446367 -0.3287996835
#> [346] -0.3289539851 -0.3291075467 -0.3292603737 -0.3294124713 -0.3295638448
#> [351] -0.3297144993 -0.3298644399 -0.3300136717 -0.3301621997 -0.3303100288
#> [356] -0.3304571640 -0.3306036102 -0.3307493722 -0.3308944547 -0.3310388626
#> [361] -0.3311826004 -0.3313256729 -0.3314680847 -0.3316098402 -0.3317509442
#> [366] -0.3318914010 -0.3320312151 -0.3321703908 -0.3323089326 -0.3324468448
#> [371] -0.3325841316 -0.3327207973 -0.3328568461 -0.3329922821 -0.3331271096
#> [376] -0.3332613325 -0.3333949550 -0.3335279811 -0.3336604147 -0.3337922598
#> [381] -0.3339235203 -0.3340542001 -0.3341843030 -0.3343138329 -0.3344427936
#> [386] -0.3345711886 -0.3346990219 -0.3348262970 -0.3349530177 -0.3350791875
#> [391] -0.3352048099 -0.3353298887 -0.3354544272 -0.3355784290 -0.3357018975
#> [396] -0.3358248362 -0.3359472485 -0.3360691378 -0.3361905073 -0.3363113604
#> [401] -0.3364317005 -0.3365515307 -0.3366708543 -0.3367896745 -0.3369079945
#> [406] -0.3370258174 -0.3371431463 -0.3372599845 -0.3373763348 -0.3374922004
#> [411] -0.3376075844 -0.3377224896 -0.3378369191 -0.3379508758 -0.3380643627
#> [416] -0.3381773826 -0.3382899384 -0.3384020330 -0.3385136692 -0.3386248498
#> [421] -0.3387355776 -0.3388458554 -0.3389556858 -0.3390650716 -0.3391740156
#> [426] -0.3392825203 -0.3393905884 -0.3394982226 -0.3396054255 -0.3397121996
#> [431] -0.3398185475 -0.3399244718 -0.3400299750 -0.3401350595 -0.3402397280
#> [436] -0.3403439828 -0.3404478264 -0.3405512612 -0.3406542896 -0.3407569141
#> [441] -0.3408591370 -0.3409609606 -0.3410623873 -0.3411634193 -0.3412640591
#> [446] -0.3413643089 -0.3414641709 -0.3415636474 -0.3416627406 -0.3417614527
#> [451] -0.3418597860 -0.3419577425 -0.3420553245 -0.3421525341 -0.3422493734
#> [456] -0.3423458446 -0.3424419496 -0.3425376908 -0.3426330699 -0.3427280892
#> [461] -0.3428227507 -0.3429170564 -0.3430110082 -0.3431046082 -0.3431978583
#> [466] -0.3432907606 -0.3433833168 -0.3434755291 -0.3435673992 -0.3436589291
#> [471] -0.3437501207 -0.3438409758 -0.3439314963 -0.3440216841 -0.3441115410
#> [476] -0.3442010687 -0.3442902692 -0.3443791441 -0.3444676953 -0.3445559246
#> [481] -0.3446438336 -0.3447314242 -0.3448186980 -0.3449056568 -0.3449923023
#> [486] -0.3450786361 -0.3451646600 -0.3452503756 -0.3453357846 -0.3454208885
#> [491] -0.3455056891 -0.3455901880 -0.3456743867 -0.3457582869 -0.3458418901
#> [496] -0.3459251979 -0.3460082120 -0.3460909337 -0.3461733648 -0.3462555066

# By default the function uses the shallow and mesophotic
# community light relationships from (Tamir et al., 2019)
# for a range of light values.

   reef()

#>   [1]  0.012369365  0.018956452  0.021965507  0.023453798  0.024028888
#>   [6]  0.023970226  0.023433742  0.022516981  0.021285891  0.019787710
#>  [11]  0.018057839  0.016123825  0.014007791  0.011728001  0.009299897
#>  [16]  0.006736811  0.004050470  0.001251353 -0.001651047 -0.004648029
#>  [21] -0.007731536 -0.010894050 -0.014128503 -0.017428217 -0.020786853
#>  [26] -0.024198373 -0.027657008 -0.031157240 -0.034693779 -0.038261549
#>  [31] -0.041855682 -0.045471503 -0.049104526 -0.052750448 -0.056405145
#>  [36] -0.060064665 -0.063725230 -0.067383228 -0.071035213 -0.074677902
#>  [41] -0.078308172 -0.081923060 -0.085519755 -0.089095604 -0.092648099
#>  [46] -0.096174885 -0.099673749 -0.103142621 -0.106579570 -0.109982801
#>  [51] -0.113350653 -0.116681591 -0.119974210 -0.123227223 -0.126439464
#>  [56] -0.129609882 -0.132737536 -0.135821591 -0.138861317 -0.141856082
#>  [61] -0.144805351 -0.147708679 -0.150565708 -0.153376165 -0.156139854
#>  [66] -0.158856658 -0.161526529 -0.164149489 -0.166725623 -0.169255077
#>  [71] -0.171738054 -0.174174812 -0.176565657 -0.178910942 -0.181211065
#>  [76] -0.183466462 -0.185677609 -0.187845013 -0.189969216 -0.192050785
#>  [81] -0.194090314 -0.196088422 -0.198045746 -0.199962944 -0.201840689
#>  [86] -0.203679666 -0.205480575 -0.207244126 -0.208971033 -0.210662021
#>  [91] -0.212317818 -0.213939153 -0.215526760 -0.217081371 -0.218603717
#>  [96] -0.220094528 -0.221554529 -0.222984440 -0.224384978 -0.225756851
#> [101] -0.227100762 -0.228417404 -0.229707462 -0.230971613 -0.232210521
#> [106] -0.233424843 -0.234615223 -0.235782294 -0.236926679 -0.238048986
#> [111] -0.239149815 -0.240229749 -0.241289363 -0.242329215 -0.243349853
#> [116] -0.244351812 -0.245335613 -0.246301765 -0.247250764 -0.248183093
#> [121] -0.249099222 -0.249999608 -0.250884697 -0.251754922 -0.252610703
#> [126] -0.253452448 -0.254280554 -0.255095405 -0.255897376 -0.256686826
#> [131] -0.257464109 -0.258229562 -0.258983516 -0.259726289 -0.260458189
#> [136] -0.261179516 -0.261890558 -0.262591595 -0.263282897 -0.263964725
#> [141] -0.264637332 -0.265300961 -0.265955849 -0.266602223 -0.267240304
#> [146] -0.267870304 -0.268492428 -0.269106874 -0.269713834 -0.270313492
#> [151] -0.270906027 -0.271491611 -0.272070410 -0.272642584 -0.273208289
#> [156] -0.273767673 -0.274320882 -0.274868054 -0.275409324 -0.275944821
#> [161] -0.276474672 -0.276998997 -0.277517914 -0.278031535 -0.278539969
#> [166] -0.279043323 -0.279541697 -0.280035191 -0.280523899 -0.281007915
#> [171] -0.281487326 -0.281962220 -0.282432680 -0.282898786 -0.283360617
#> [176] -0.283818248 -0.284271753 -0.284721203 -0.285166667 -0.285608212
#> [181] -0.286045902 -0.286479801 -0.286909969 -0.287336466 -0.287759350
#> [186] -0.288178675 -0.288594498 -0.289006870 -0.289415843 -0.289821467
#> [191] -0.290223790 -0.290622861 -0.291018725 -0.291411428 -0.291801012
#> [196] -0.292187522 -0.292570998 -0.292951481 -0.293329012 -0.293703628
#> [201] -0.294075368 -0.294444270 -0.294810368 -0.295173698 -0.295534296
#> [206] -0.295892194 -0.296247426 -0.296600025 -0.296950023 -0.297297449
#> [211] -0.297642336 -0.297984712 -0.298324607 -0.298662051 -0.298997070
#> [216] -0.299329693 -0.299659948 -0.299987859 -0.300313455 -0.300636760
#> [221] -0.300957800 -0.301276599 -0.301593182 -0.301907574 -0.302219797
#> [226] -0.302529875 -0.302837830 -0.303143685 -0.303447463 -0.303749184
#> [231] -0.304048871 -0.304346544 -0.304642224 -0.304935932 -0.305227687
#> [236] -0.305517510 -0.305805419 -0.306091435 -0.306375576 -0.306657861
#> [241] -0.306938308 -0.307216936 -0.307493761 -0.307768803 -0.308042078
#> [246] -0.308313604 -0.308583397 -0.308851475 -0.309117853 -0.309382548
#> [251] -0.309645575 -0.309906952 -0.310166693 -0.310424814 -0.310681330
#> [256] -0.310936255 -0.311189606 -0.311441396 -0.311691640 -0.311940353
#> [261] -0.312187547 -0.312433238 -0.312677439 -0.312920163 -0.313161424
#> [266] -0.313401235 -0.313639609 -0.313876560 -0.314112099 -0.314346239
#> [271] -0.314578994 -0.314810375 -0.315040394 -0.315269064 -0.315496395
#> [276] -0.315722401 -0.315947092 -0.316170480 -0.316392577 -0.316613393
#> [281] -0.316832939 -0.317051227 -0.317268267 -0.317484070 -0.317698647
#> [286] -0.317912008 -0.318124162 -0.318335122 -0.318544895 -0.318753494
#> [291] -0.318960926 -0.319167203 -0.319372334 -0.319576327 -0.319779194
#> [296] -0.319980942 -0.320181582 -0.320381122 -0.320579572 -0.320776939
#> [301] -0.320973234 -0.321168465 -0.321362640 -0.321555769 -0.321747858
#> [306] -0.321938918 -0.322128956 -0.322317979 -0.322505998 -0.322693018
#> [311] -0.322879049 -0.323064098 -0.323248173 -0.323431282 -0.323613431
#> [316] -0.323794630 -0.323974884 -0.324154202 -0.324332590 -0.324510057
#> [321] -0.324686608 -0.324862252 -0.325036994 -0.325210843 -0.325383805
#> [326] -0.325555886 -0.325727093 -0.325897434 -0.326066914 -0.326235540
#> [331] -0.326403319 -0.326570256 -0.326736359 -0.326901633 -0.327066085
#> [336] -0.327229720 -0.327392545 -0.327554566 -0.327715789 -0.327876219
#> [341] -0.328035863 -0.328194726 -0.328352814 -0.328510132 -0.328666687
#> [346] -0.328822483 -0.328977527 -0.329131823 -0.329285378 -0.329438195
#> [351] -0.329590282 -0.329741642 -0.329892282 -0.330042206 -0.330191419
#> [356] -0.330339926 -0.330487733 -0.330634845 -0.330781265 -0.330927000
#> [361] -0.331072053 -0.331216430 -0.331360136 -0.331503174 -0.331645550
#> [366] -0.331787269 -0.331928334 -0.332068751 -0.332208523 -0.332347656
#> [371] -0.332486153 -0.332624019 -0.332761258 -0.332897875 -0.333033874
#> [376] -0.333169258 -0.333304033 -0.333438202 -0.333571769 -0.333704738
#> [381] -0.333837114 -0.333968900 -0.334100100 -0.334230718 -0.334360758
#> [386] -0.334490224 -0.334619119 -0.334747448 -0.334875214 -0.335002421
#> [391] -0.335129072 -0.335255171 -0.335380722 -0.335505728 -0.335630193
#> [396] -0.335754120 -0.335877513 -0.336000375 -0.336122709 -0.336244520
#> [401] -0.336365810 -0.336486583 -0.336606841 -0.336726589 -0.336845830
#> [406] -0.336964566 -0.337082801 -0.337200539 -0.337317781 -0.337434532
#> [411] -0.337550794 -0.337666571 -0.337781866 -0.337896680 -0.338011019
#> [416] -0.338124884 -0.338238278 -0.338351205 -0.338463666 -0.338575666
#> [421] -0.338687207 -0.338798292 -0.338908923 -0.339019103 -0.339128836
#> [426] -0.339238123 -0.339346968 -0.339455372 -0.339563340 -0.339670873
#> [431] -0.339777975 -0.339884646 -0.339990892 -0.340096713 -0.340202112
#> [436] -0.340307092 -0.340411656 -0.340515805 -0.340619543 -0.340722871
#> [441] -0.340825793 -0.340928310 -0.341030425 -0.341132140 -0.341233458
#> [446] -0.341334381 -0.341434911 -0.341535050 -0.341634802 -0.341734167
#> [451] -0.341833149 -0.341931749 -0.342029970 -0.342127814 -0.342225283
#> [456] -0.342322380 -0.342419105 -0.342515462 -0.342611453 -0.342707079
#> [461] -0.342802343 -0.342897247 -0.342991793 -0.343085982 -0.343179817
#> [466] -0.343273301 -0.343366434 -0.343459219 -0.343551657 -0.343643751
#> [471] -0.343735503 -0.343826915 -0.343917987 -0.344008723 -0.344099124
#> [476] -0.344189193 -0.344278930 -0.344368337 -0.344457417 -0.344546172
#> [481] -0.344634602 -0.344722710 -0.344810498 -0.344897967 -0.344985119
#> [486] -0.345071956 -0.345158479 -0.345244691 -0.345330592 -0.345416185
#> [491] -0.345501471 -0.345586452 -0.345671129 -0.345755505 -0.345839580
#> [496] -0.345923356 -0.346006836 -0.346090019 -0.346172909 -0.346255507

# You can specify your own community light relationships
# by setting parameters explicitly

   reef(Vm = 0.2, K = 11, a = 1.2, b = 8)

#>   [1]  0.0603156210  0.0709830840  0.0744227597  0.0753615465  0.0749777550
#>   [6]  0.0737699643  0.0720019795  0.0698326232  0.0673659903  0.0646745381
#>  [11]  0.0618110117  0.0588151530  0.0557177322  0.0525430991  0.0493108694
#>  [16]  0.0460370778  0.0427349912  0.0394156974  0.0360885402  0.0327614476
#>  [21]  0.0294411841  0.0261335467  0.0228435204  0.0195754026  0.0163329031
#>  [26]  0.0131192269  0.0099371421  0.0067890364  0.0036769644  0.0006026887
#>  [31] -0.0024322869 -0.0054266868 -0.0083794392 -0.0112896530 -0.0141565997
#>  [36] -0.0169796953 -0.0197584862 -0.0224926358 -0.0251819127 -0.0278261805
#>  [41] -0.0304253879 -0.0329795609 -0.0354887952 -0.0379532488 -0.0403731366
#>  [46] -0.0427487240 -0.0450803225 -0.0473682847 -0.0496130002 -0.0518148914
#>  [51] -0.0539744105 -0.0560920357 -0.0581682686 -0.0602036315 -0.0621986643
#>  [56] -0.0641539230 -0.0660699769 -0.0679474070 -0.0697868041 -0.0715887670
#>  [61] -0.0733539012 -0.0750828173 -0.0767761299 -0.0784344561 -0.0800584145
#>  [66] -0.0816486244 -0.0832057046 -0.0847302728 -0.0862229444 -0.0876843322
#>  [71] -0.0891150454 -0.0905156895 -0.0918868650 -0.0932291676 -0.0945431873
#>  [76] -0.0958295082 -0.0970887081 -0.0983213580 -0.0995280221 -0.1007092573
#>  [81] -0.1018656130 -0.1029976309 -0.1041058449 -0.1051907808 -0.1062529565
#>  [86] -0.1072928812 -0.1083110561 -0.1093079738 -0.1102841184 -0.1112399656
#>  [91] -0.1121759825 -0.1130926275 -0.1139903504 -0.1148695928 -0.1157307874
#>  [96] -0.1165743585 -0.1174007219 -0.1182102852 -0.1190034473 -0.1197805992
#> [101] -0.1205421235 -0.1212883945 -0.1220197788 -0.1227366348 -0.1234393131
#> [106] -0.1241281566 -0.1248035004 -0.1254656721 -0.1261149919 -0.1267517727
#> [111] -0.1273763200 -0.1279889323 -0.1285899010 -0.1291795109 -0.1297580396
#> [116] -0.1303257585 -0.1308829320 -0.1314298185 -0.1319666699 -0.1324937321
#> [121] -0.1330112446 -0.1335194414 -0.1340185505 -0.1345087940 -0.1349903888
#> [126] -0.1354635461 -0.1359284718 -0.1363853666 -0.1368344261 -0.1372758408
#> [131] -0.1377097964 -0.1381364738 -0.1385560492 -0.1389686942 -0.1393745760
#> [136] -0.1397738575 -0.1401666973 -0.1405532496 -0.1409336649 -0.1413080897
#> [141] -0.1416766664 -0.1420395337 -0.1423968269 -0.1427486774 -0.1430952132
#> [146] -0.1434365590 -0.1437728360 -0.1441041623 -0.1444306529 -0.1447524194
#> [151] -0.1450695708 -0.1453822130 -0.1456904491 -0.1459943793 -0.1462941014
#> [156] -0.1465897102 -0.1468812984 -0.1471689557 -0.1474527698 -0.1477328259
#> [161] -0.1480092069 -0.1482819934 -0.1485512640 -0.1488170951 -0.1490795610
#> [166] -0.1493387341 -0.1495946848 -0.1498474817 -0.1500971914 -0.1503438788
#> [171] -0.1505876073 -0.1508284382 -0.1510664314 -0.1513016452 -0.1515341363
#> [176] -0.1517639600 -0.1519911699 -0.1522158185 -0.1524379567 -0.1526576340
#> [181] -0.1528748989 -0.1530897984 -0.1533023782 -0.1535126831 -0.1537207563
#> [186] -0.1539266404 -0.1541303765 -0.1543320048 -0.1545315644 -0.1547290935
#> [191] -0.1549246292 -0.1551182077 -0.1553098643 -0.1554996334 -0.1556875485
#> [196] -0.1558736423 -0.1560579467 -0.1562404927 -0.1564213106 -0.1566004300
#> [201] -0.1567778796 -0.1569536876 -0.1571278813 -0.1573004876 -0.1574715325
#> [206] -0.1576410414 -0.1578090393 -0.1579755504 -0.1581405983 -0.1583042062
#> [211] -0.1584663967 -0.1586271917 -0.1587866129 -0.1589446812 -0.1591014171
#> [216] -0.1592568408 -0.1594109718 -0.1595638294 -0.1597154321 -0.1598657984
#> [221] -0.1600149461 -0.1601628928 -0.1603096554 -0.1604552509 -0.1605996955
#> [226] -0.1607430052 -0.1608851957 -0.1610262824 -0.1611662802 -0.1613052039
#> [231] -0.1614430677 -0.1615798859 -0.1617156721 -0.1618504399 -0.1619842024
#> [236] -0.1621169727 -0.1622487635 -0.1623795871 -0.1625094557 -0.1626383814
#> [241] -0.1627663757 -0.1628934503 -0.1630196162 -0.1631448847 -0.1632692665
#> [246] -0.1633927722 -0.1635154123 -0.1636371970 -0.1637581364 -0.1638782402
#> [251] -0.1639975181 -0.1641159797 -0.1642336343 -0.1643504910 -0.1644665589
#> [256] -0.1645818467 -0.1646963631 -0.1648101167 -0.1649231158 -0.1650353687
#> [261] -0.1651468835 -0.1652576682 -0.1653677304 -0.1654770781 -0.1655857186
#> [266] -0.1656936596 -0.1658009082 -0.1659074717 -0.1660133572 -0.1661185716
#> [271] -0.1662231218 -0.1663270146 -0.1664302566 -0.1665328544 -0.1666348143
#> [276] -0.1667361428 -0.1668368461 -0.1669369304 -0.1670364017 -0.1671352660
#> [281] -0.1672335291 -0.1673311969 -0.1674282751 -0.1675247693 -0.1676206851
#> [286] -0.1677160279 -0.1678108032 -0.1679050162 -0.1679986721 -0.1680917763
#> [291] -0.1681843337 -0.1682763493 -0.1683678282 -0.1684587753 -0.1685491953
#> [296] -0.1686390930 -0.1687284731 -0.1688173403 -0.1689056991 -0.1689935540
#> [301] -0.1690809096 -0.1691677701 -0.1692541400 -0.1693400235 -0.1694254248
#> [306] -0.1695103482 -0.1695947978 -0.1696787775 -0.1697622916 -0.1698453438
#> [311] -0.1699279383 -0.1700100787 -0.1700917691 -0.1701730130 -0.1702538144
#> [316] -0.1703341768 -0.1704141040 -0.1704935995 -0.1705726669 -0.1706513098
#> [321] -0.1707295315 -0.1708073356 -0.1708847255 -0.1709617044 -0.1710382758
#> [326] -0.1711144429 -0.1711902090 -0.1712655773 -0.1713405509 -0.1714151330
#> [331] -0.1714893267 -0.1715631351 -0.1716365613 -0.1717096081 -0.1717822786
#> [336] -0.1718545758 -0.1719265024 -0.1719980615 -0.1720692558 -0.1721400882
#> [341] -0.1722105615 -0.1722806784 -0.1723504415 -0.1724198538 -0.1724889177
#> [346] -0.1725576359 -0.1726260111 -0.1726940458 -0.1727617426 -0.1728291041
#> [351] -0.1728961326 -0.1729628308 -0.1730292011 -0.1730952458 -0.1731609674
#> [356] -0.1732263684 -0.1732914509 -0.1733562175 -0.1734206703 -0.1734848117
#> [361] -0.1735486440 -0.1736121693 -0.1736753900 -0.1737383081 -0.1738009259
#> [366] -0.1738632455 -0.1739252692 -0.1739869989 -0.1740484367 -0.1741095849
#> [371] -0.1741704453 -0.1742310201 -0.1742913112 -0.1743513207 -0.1744110505
#> [376] -0.1744705025 -0.1745296788 -0.1745885813 -0.1746472118 -0.1747055722
#> [381] -0.1747636644 -0.1748214902 -0.1748790516 -0.1749363502 -0.1749933879
#> [386] -0.1750501665 -0.1751066878 -0.1751629534 -0.1752189652 -0.1752747248
#> [391] -0.1753302339 -0.1753854943 -0.1754405076 -0.1754952755 -0.1755497995
#> [396] -0.1756040814 -0.1756581228 -0.1757119252 -0.1757654902 -0.1758188194
#> [401] -0.1758719143 -0.1759247766 -0.1759774077 -0.1760298091 -0.1760819823
#> [406] -0.1761339289 -0.1761856503 -0.1762371480 -0.1762884234 -0.1763394779
#> [411] -0.1763903130 -0.1764409301 -0.1764913306 -0.1765415159 -0.1765914874
#> [416] -0.1766412463 -0.1766907942 -0.1767401323 -0.1767892619 -0.1768381844
#> [421] -0.1768869011 -0.1769354132 -0.1769837221 -0.1770318291 -0.1770797353
#> [426] -0.1771274421 -0.1771749507 -0.1772222624 -0.1772693782 -0.1773162996
#> [431] -0.1773630276 -0.1774095635 -0.1774559085 -0.1775020636 -0.1775480302
#> [436] -0.1775938093 -0.1776394021 -0.1776848096 -0.1777300332 -0.1777750738
#> [441] -0.1778199326 -0.1778646106 -0.1779091090 -0.1779534289 -0.1779975712
#> [446] -0.1780415372 -0.1780853277 -0.1781289440 -0.1781723870 -0.1782156577
#> [451] -0.1782587572 -0.1783016865 -0.1783444466 -0.1783870385 -0.1784294632
#> [456] -0.1784717216 -0.1785138148 -0.1785557437 -0.1785975093 -0.1786391125
#> [461] -0.1786805543 -0.1787218355 -0.1787629573 -0.1788039203 -0.1788447257
#> [466] -0.1788853742 -0.1789258669 -0.1789662045 -0.1790063880 -0.1790464182
#> [471] -0.1790862961 -0.1791260225 -0.1791655982 -0.1792050241 -0.1792443012
#> [476] -0.1792834301 -0.1793224117 -0.1793612470 -0.1793999366 -0.1794384815
#> [481] -0.1794768823 -0.1795151401 -0.1795532554 -0.1795912291 -0.1796290621
#> [486] -0.1796667551 -0.1797043088 -0.1797417241 -0.1797790016 -0.1798161423
#> [491] -0.1798531467 -0.1798900157 -0.1799267500 -0.1799633504 -0.1799998175
#> [496] -0.1800361521 -0.1800723550 -0.1801084267 -0.1801443682 -0.1801801799