Calculate the shallow community value for a given light level

This function captures the preferred light environment of the "shallow" coral community.

shallow(light = seq(0.01, 100, length = 500), Vm = 0.393, K = 13.5)

Arguments

light
The percentage of surface irradiance (PAR) reaching a reef, positive and up to 100%.

Vm
The maximum similarity of a reef patch to the idealised shallow reef community, between 0 and 1. Defaults to (Tamir et al., 2019).

K
The light value where the community value = Vmax/2, between 0 and 100. Defaults to (Tamir et al., 2019). This can be interpreted as the light level at which shallow reef communities become light limited.

Value

A shallow community indicator value, as discussed in (Laverick et al., 2017). When a single light value is provided, a single value is returned. You can pass a vector of light levels for a vector of depths.

Details

A Michaelis-Menten equation is used to represent the relationship between shallow coral communities and light. We assume shallow corals are light-limited beneath a light threshold, and that community values (how similar a reef patch is to the idealised shallow community) increase asymptotically with increasing light.

This relationship has been verified in the Red Sea (Tamir et al., 2019), but certainly needs more research!

References

(Tamir et al., 2019)

(Laverick et al., 2017)

See also

Examples

# Specifying a single light value
# returns a single community value.

   shallow(light = 10)

#> [1] 0.3877652

# Specifying a vector of light values
# returns a vector of community values.

   shallow(light = seq(0.0065, 100, length = 500))

#>   [1] 0.0180530 0.2378175 0.2951623 0.3215612 0.3367411 0.3466005 0.3535195
#>   [8] 0.3586427 0.3625891 0.3657222 0.3682701 0.3703827 0.3721627 0.3736830
#>  [15] 0.3749965 0.3761428 0.3771518 0.3780469 0.3788463 0.3795645 0.3802134
#>  [22] 0.3808024 0.3813396 0.3818315 0.3822836 0.3827004 0.3830861 0.3834439
#>  [29] 0.3837768 0.3840873 0.3843776 0.3846495 0.3849048 0.3851450 0.3853713
#>  [36] 0.3855850 0.3857870 0.3859783 0.3861597 0.3863320 0.3864958 0.3866517
#>  [43] 0.3868004 0.3869422 0.3870777 0.3872073 0.3873313 0.3874502 0.3875641
#>  [50] 0.3876735 0.3877785 0.3878795 0.3879766 0.3880702 0.3881603 0.3882472
#>  [57] 0.3883310 0.3884119 0.3884901 0.3885656 0.3886386 0.3887093 0.3887777
#>  [64] 0.3888440 0.3889082 0.3889705 0.3890309 0.3890895 0.3891465 0.3892017
#>  [71] 0.3892555 0.3893077 0.3893585 0.3894079 0.3894560 0.3895028 0.3895484
#>  [78] 0.3895928 0.3896361 0.3896783 0.3897194 0.3897596 0.3897988 0.3898370
#>  [85] 0.3898744 0.3899108 0.3899464 0.3899813 0.3900153 0.3900486 0.3900811
#>  [92] 0.3901129 0.3901441 0.3901746 0.3902044 0.3902336 0.3902622 0.3902902
#>  [99] 0.3903177 0.3903446 0.3903709 0.3903968 0.3904221 0.3904470 0.3904714
#> [106] 0.3904953 0.3905188 0.3905418 0.3905644 0.3905866 0.3906084 0.3906298
#> [113] 0.3906509 0.3906715 0.3906918 0.3907118 0.3907314 0.3907507 0.3907696
#> [120] 0.3907882 0.3908066 0.3908246 0.3908423 0.3908598 0.3908769 0.3908938
#> [127] 0.3909104 0.3909268 0.3909429 0.3909588 0.3909744 0.3909898 0.3910049
#> [134] 0.3910198 0.3910345 0.3910490 0.3910633 0.3910773 0.3910912 0.3911049
#> [141] 0.3911183 0.3911316 0.3911447 0.3911576 0.3911703 0.3911829 0.3911953
#> [148] 0.3912075 0.3912196 0.3912315 0.3912432 0.3912548 0.3912662 0.3912775
#> [155] 0.3912886 0.3912996 0.3913105 0.3913212 0.3913317 0.3913422 0.3913525
#> [162] 0.3913627 0.3913728 0.3913827 0.3913925 0.3914022 0.3914118 0.3914213
#> [169] 0.3914306 0.3914399 0.3914490 0.3914581 0.3914670 0.3914758 0.3914845
#> [176] 0.3914932 0.3915017 0.3915101 0.3915185 0.3915267 0.3915348 0.3915429
#> [183] 0.3915509 0.3915588 0.3915666 0.3915743 0.3915819 0.3915895 0.3915970
#> [190] 0.3916044 0.3916117 0.3916189 0.3916261 0.3916332 0.3916402 0.3916471
#> [197] 0.3916540 0.3916608 0.3916676 0.3916742 0.3916808 0.3916874 0.3916939
#> [204] 0.3917003 0.3917066 0.3917129 0.3917191 0.3917253 0.3917314 0.3917375
#> [211] 0.3917435 0.3917494 0.3917553 0.3917611 0.3917669 0.3917726 0.3917782
#> [218] 0.3917839 0.3917894 0.3917949 0.3918004 0.3918058 0.3918112 0.3918165
#> [225] 0.3918217 0.3918270 0.3918321 0.3918373 0.3918423 0.3918474 0.3918524
#> [232] 0.3918573 0.3918622 0.3918671 0.3918719 0.3918767 0.3918815 0.3918862
#> [239] 0.3918908 0.3918955 0.3919001 0.3919046 0.3919091 0.3919136 0.3919180
#> [246] 0.3919224 0.3919268 0.3919311 0.3919354 0.3919397 0.3919439 0.3919481
#> [253] 0.3919523 0.3919564 0.3919605 0.3919646 0.3919686 0.3919726 0.3919766
#> [260] 0.3919805 0.3919844 0.3919883 0.3919922 0.3919960 0.3919998 0.3920036
#> [267] 0.3920073 0.3920110 0.3920147 0.3920183 0.3920220 0.3920256 0.3920291
#> [274] 0.3920327 0.3920362 0.3920397 0.3920432 0.3920466 0.3920500 0.3920534
#> [281] 0.3920568 0.3920601 0.3920635 0.3920668 0.3920701 0.3920733 0.3920765
#> [288] 0.3920797 0.3920829 0.3920861 0.3920892 0.3920924 0.3920955 0.3920985
#> [295] 0.3921016 0.3921046 0.3921077 0.3921107 0.3921136 0.3921166 0.3921195
#> [302] 0.3921225 0.3921254 0.3921282 0.3921311 0.3921339 0.3921368 0.3921396
#> [309] 0.3921424 0.3921451 0.3921479 0.3921506 0.3921533 0.3921560 0.3921587
#> [316] 0.3921614 0.3921640 0.3921666 0.3921693 0.3921719 0.3921744 0.3921770
#> [323] 0.3921796 0.3921821 0.3921846 0.3921871 0.3921896 0.3921921 0.3921945
#> [330] 0.3921970 0.3921994 0.3922018 0.3922042 0.3922066 0.3922090 0.3922113
#> [337] 0.3922137 0.3922160 0.3922183 0.3922206 0.3922229 0.3922252 0.3922274
#> [344] 0.3922297 0.3922319 0.3922341 0.3922364 0.3922385 0.3922407 0.3922429
#> [351] 0.3922451 0.3922472 0.3922493 0.3922515 0.3922536 0.3922557 0.3922578
#> [358] 0.3922598 0.3922619 0.3922640 0.3922660 0.3922680 0.3922700 0.3922720
#> [365] 0.3922740 0.3922760 0.3922780 0.3922800 0.3922819 0.3922839 0.3922858
#> [372] 0.3922877 0.3922896 0.3922915 0.3922934 0.3922953 0.3922972 0.3922990
#> [379] 0.3923009 0.3923027 0.3923046 0.3923064 0.3923082 0.3923100 0.3923118
#> [386] 0.3923136 0.3923153 0.3923171 0.3923189 0.3923206 0.3923224 0.3923241
#> [393] 0.3923258 0.3923275 0.3923292 0.3923309 0.3923326 0.3923343 0.3923359
#> [400] 0.3923376 0.3923393 0.3923409 0.3923425 0.3923442 0.3923458 0.3923474
#> [407] 0.3923490 0.3923506 0.3923522 0.3923538 0.3923554 0.3923569 0.3923585
#> [414] 0.3923600 0.3923616 0.3923631 0.3923646 0.3923662 0.3923677 0.3923692
#> [421] 0.3923707 0.3923722 0.3923737 0.3923751 0.3923766 0.3923781 0.3923795
#> [428] 0.3923810 0.3923824 0.3923839 0.3923853 0.3923867 0.3923881 0.3923895
#> [435] 0.3923909 0.3923923 0.3923937 0.3923951 0.3923965 0.3923979 0.3923992
#> [442] 0.3924006 0.3924019 0.3924033 0.3924046 0.3924060 0.3924073 0.3924086
#> [449] 0.3924099 0.3924113 0.3924126 0.3924139 0.3924152 0.3924164 0.3924177
#> [456] 0.3924190 0.3924203 0.3924215 0.3924228 0.3924241 0.3924253 0.3924266
#> [463] 0.3924278 0.3924290 0.3924303 0.3924315 0.3924327 0.3924339 0.3924351
#> [470] 0.3924363 0.3924375 0.3924387 0.3924399 0.3924411 0.3924423 0.3924434
#> [477] 0.3924446 0.3924458 0.3924469 0.3924481 0.3924492 0.3924504 0.3924515
#> [484] 0.3924526 0.3924538 0.3924549 0.3924560 0.3924571 0.3924582 0.3924593
#> [491] 0.3924604 0.3924615 0.3924626 0.3924637 0.3924648 0.3924659 0.3924670
#> [498] 0.3924680 0.3924691 0.3924702

# By default the function uses the shallow community
# light relationship from (Tamir et al., 2019) for a
# range of light values.

   shallow()

#>   [1] 0.0002908956 0.0060304408 0.0116046327 0.0170205156 0.0222847391
#>   [6] 0.0274035855 0.0323829949 0.0372285881 0.0419456879 0.0465393383
#>  [11] 0.0510143230 0.0553751814 0.0596262244 0.0637715486 0.0678150491
#>  [16] 0.0717604322 0.0756112263 0.0793707927 0.0830423354 0.0866289100
#>  [21] 0.0901334323 0.0935586861 0.0969073308 0.1001819083 0.1033848489
#>  [26] 0.1065184780 0.1095850215 0.1125866106 0.1155252877 0.1184030100
#>  [31] 0.1212216546 0.1239830224 0.1266888415 0.1293407714 0.1319404062
#>  [36] 0.1344892776 0.1369888581 0.1394405638 0.1418457572 0.1442057494
#>  [41] 0.1465218027 0.1487951330 0.1510269114 0.1532182671 0.1553702883
#>  [46] 0.1574840246 0.1595604888 0.1616006580 0.1636054756 0.1655758526
#>  [51] 0.1675126689 0.1694167746 0.1712889916 0.1731301142 0.1749409107
#>  [56] 0.1767221243 0.1784744740 0.1801986557 0.1818953433 0.1835651892
#>  [61] 0.1852088253 0.1868268639 0.1884198984 0.1899885038 0.1915332380
#>  [66] 0.1930546415 0.1945532392 0.1960295399 0.1974840375 0.1989172116
#>  [71] 0.2003295276 0.2017214377 0.2030933808 0.2044457836 0.2057790605
#>  [76] 0.2070936145 0.2083898372 0.2096681093 0.2109288013 0.2121722732
#>  [81] 0.2133988755 0.2146089492 0.2158028261 0.2169808293 0.2181432732
#>  [86] 0.2192904642 0.2204227004 0.2215402723 0.2226434632 0.2237325486
#>  [91] 0.2248077975 0.2258694720 0.2269178273 0.2279531126 0.2289755708
#>  [96] 0.2299854389 0.2309829479 0.2319683234 0.2329417854 0.2339035486
#> [101] 0.2348538227 0.2357928123 0.2367207173 0.2376377328 0.2385440493
#> [106] 0.2394398530 0.2403253258 0.2412006454 0.2420659853 0.2429215154
#> [111] 0.2437674014 0.2446038055 0.2454308863 0.2462487988 0.2470576946
#> [116] 0.2478577220 0.2486490259 0.2494317485 0.2502060284 0.2509720015
#> [121] 0.2517298009 0.2524795566 0.2532213960 0.2539554439 0.2546818223
#> [126] 0.2554006509 0.2561120468 0.2568161245 0.2575129965 0.2582027729
#> [131] 0.2588855614 0.2595614677 0.2602305953 0.2608930458 0.2615489185
#> [136] 0.2621983109 0.2628413186 0.2634780354 0.2641085530 0.2647329616
#> [141] 0.2653513495 0.2659638034 0.2665704083 0.2671712476 0.2677664031
#> [146] 0.2683559551 0.2689399823 0.2695185620 0.2700917702 0.2706596811
#> [151] 0.2712223681 0.2717799027 0.2723323555 0.2728797956 0.2734222909
#> [156] 0.2739599082 0.2744927128 0.2750207692 0.2755441406 0.2760628890
#> [161] 0.2765770754 0.2770867597 0.2775920007 0.2780928565 0.2785893837
#> [166] 0.2790816383 0.2795696751 0.2800535482 0.2805333106 0.2810090145
#> [171] 0.2814807111 0.2819484509 0.2824122834 0.2828722575 0.2833284210
#> [176] 0.2837808211 0.2842295042 0.2846745159 0.2851159011 0.2855537040
#> [181] 0.2859879679 0.2864187357 0.2868460493 0.2872699502 0.2876904790
#> [186] 0.2881076759 0.2885215802 0.2889322309 0.2893396661 0.2897439235
#> [191] 0.2901450400 0.2905430522 0.2909379959 0.2913299065 0.2917188188
#> [196] 0.2921047671 0.2924877851 0.2928679061 0.2932451628 0.2936195875
#> [201] 0.2939912119 0.2943600673 0.2947261847 0.2950895942 0.2954503260
#> [206] 0.2958084094 0.2961638736 0.2965167471 0.2968670582 0.2972148346
#> [211] 0.2975601039 0.2979028930 0.2982432285 0.2985811367 0.2989166434
#> [216] 0.2992497743 0.2995805543 0.2999090084 0.3002351609 0.3005590360
#> [221] 0.3008806575 0.3012000487 0.3015172328 0.3018322326 0.3021450707
#> [226] 0.3024557690 0.3027643496 0.3030708341 0.3033752436 0.3036775993
#> [231] 0.3039779217 0.3042762315 0.3045725486 0.3048668931 0.3051592845
#> [236] 0.3054497423 0.3057382855 0.3060249330 0.3063097035 0.3065926153
#> [241] 0.3068736866 0.3071529352 0.3074303788 0.3077060349 0.3079799207
#> [246] 0.3082520532 0.3085224492 0.3087911252 0.3090580976 0.3093233825
#> [251] 0.3095869960 0.3098489537 0.3101092712 0.3103679639 0.3106250469
#> [256] 0.3108805352 0.3111344435 0.3113867866 0.3116375788 0.3118868344
#> [261] 0.3121345675 0.3123807919 0.3126255215 0.3128687697 0.3131105500
#> [266] 0.3133508757 0.3135897597 0.3138272152 0.3140632548 0.3142978911
#> [271] 0.3145311367 0.3147630039 0.3149935048 0.3152226515 0.3154504559
#> [276] 0.3156769297 0.3159020846 0.3161259321 0.3163484835 0.3165697500
#> [281] 0.3167897428 0.3170084728 0.3172259508 0.3174421876 0.3176571937
#> [286] 0.3178709797 0.3180835559 0.3182949325 0.3185051196 0.3187141274
#> [291] 0.3189219656 0.3191286440 0.3193341725 0.3195385604 0.3197418173
#> [296] 0.3199439525 0.3201449753 0.3203448949 0.3205437203 0.3207414605
#> [301] 0.3209381243 0.3211337205 0.3213282578 0.3215217447 0.3217141897
#> [306] 0.3219056013 0.3220959877 0.3222853571 0.3224737177 0.3226610775
#> [311] 0.3228474445 0.3230328265 0.3232172313 0.3234006666 0.3235831401
#> [316] 0.3237646592 0.3239452316 0.3241248645 0.3243035652 0.3244813411
#> [321] 0.3246581992 0.3248341466 0.3250091904 0.3251833376 0.3253565949
#> [326] 0.3255289692 0.3257004672 0.3258710956 0.3260408610 0.3262097699
#> [331] 0.3263778287 0.3265450440 0.3267114219 0.3268769689 0.3270416910
#> [336] 0.3272055945 0.3273686854 0.3275309698 0.3276924536 0.3278531427
#> [341] 0.3280130431 0.3281721604 0.3283305004 0.3284880689 0.3286448714
#> [346] 0.3288009134 0.3289562006 0.3291107384 0.3292645322 0.3294175873
#> [351] 0.3295699091 0.3297215028 0.3298723737 0.3300225268 0.3301719673
#> [356] 0.3303207003 0.3304687308 0.3306160637 0.3307627039 0.3309086564
#> [361] 0.3310539259 0.3311985173 0.3313424352 0.3314856845 0.3316282696
#> [366] 0.3317701953 0.3319114660 0.3320520864 0.3321920609 0.3323313939
#> [371] 0.3324700898 0.3326081531 0.3327455879 0.3328823987 0.3330185895
#> [376] 0.3331541647 0.3332891285 0.3334234848 0.3335572379 0.3336903918
#> [381] 0.3338229504 0.3339549178 0.3340862980 0.3342170948 0.3343473121
#> [386] 0.3344769537 0.3346060235 0.3347345253 0.3348624627 0.3349898395
#> [391] 0.3351166594 0.3352429260 0.3353686429 0.3354938137 0.3356184420
#> [396] 0.3357425313 0.3358660850 0.3359891067 0.3361115997 0.3362335674
#> [401] 0.3363550133 0.3364759406 0.3365963528 0.3367162530 0.3368356445
#> [406] 0.3369545306 0.3370729144 0.3371907992 0.3373081880 0.3374250841
#> [411] 0.3375414904 0.3376574102 0.3377728463 0.3378878019 0.3380022799
#> [416] 0.3381162834 0.3382298151 0.3383428782 0.3384554754 0.3385676096
#> [421] 0.3386792838 0.3387905006 0.3389012630 0.3390115736 0.3391214354
#> [426] 0.3392308509 0.3393398229 0.3394483541 0.3395564472 0.3396641047
#> [431] 0.3397713295 0.3398781239 0.3399844907 0.3400904324 0.3401959515
#> [436] 0.3403010506 0.3404057321 0.3405099986 0.3406138525 0.3407172962
#> [441] 0.3408203322 0.3409229629 0.3410251907 0.3411270179 0.3412284469
#> [446] 0.3413294800 0.3414301195 0.3415303678 0.3416302271 0.3417296996
#> [451] 0.3418287877 0.3419274935 0.3420258192 0.3421237670 0.3422213392
#> [456] 0.3423185378 0.3424153650 0.3425118230 0.3426079138 0.3427036395
#> [461] 0.3427990022 0.3428940040 0.3429886469 0.3430829329 0.3431768641
#> [466] 0.3432704424 0.3433636699 0.3434565485 0.3435490802 0.3436412668
#> [471] 0.3437331104 0.3438246128 0.3439157760 0.3440066018 0.3440970921
#> [476] 0.3441872487 0.3442770735 0.3443665684 0.3444557350 0.3445445753
#> [481] 0.3446330911 0.3447212840 0.3448091559 0.3448967085 0.3449839435
#> [486] 0.3450708627 0.3451574678 0.3452437605 0.3453297425 0.3454154154
#> [491] 0.3455007809 0.3455858407 0.3456705964 0.3457550496 0.3458392019
#> [496] 0.3459230550 0.3460066105 0.3460898698 0.3461728347 0.3462555066

# You can specify your own community light relationship
# by setting parameters explicitly

   shallow(Vm = 0.2, K = 11)

#>   [1] 0.000181653 0.003753321 0.007199546 0.010526824 0.013741209 0.016848350
#>   [7] 0.019853526 0.022761676 0.025577423 0.028305103 0.030948784 0.033512287
#>  [13] 0.035999205 0.038412920 0.040756616 0.043033298 0.045245798 0.047396793
#>  [19] 0.049488813 0.051524250 0.053505369 0.055434316 0.057313125 0.059143725
#>  [25] 0.060927949 0.062667537 0.064364144 0.066019342 0.067634629 0.069211432
#>  [31] 0.070751111 0.072254959 0.073724215 0.075160058 0.076563616 0.077935964
#>  [37] 0.079278132 0.080591106 0.081875827 0.083133197 0.084364081 0.085569308
#>  [43] 0.086749670 0.087905930 0.089038818 0.090149036 0.091237258 0.092304130
#>  [49] 0.093350276 0.094376293 0.095382756 0.096370220 0.097339217 0.098290261
#>  [55] 0.099223846 0.100140447 0.101040526 0.101924524 0.102792868 0.103645970
#>  [61] 0.104484230 0.105308029 0.106117740 0.106913721 0.107696318 0.108465865
#>  [67] 0.109222687 0.109967097 0.110699397 0.111419880 0.112128831 0.112826523
#>  [73] 0.113513224 0.114189190 0.114854672 0.115509911 0.116155143 0.116790594
#>  [79] 0.117416486 0.118033032 0.118640440 0.119238912 0.119828644 0.120409826
#>  [85] 0.120982643 0.121547272 0.122103890 0.122652666 0.123193763 0.123727342
#>  [91] 0.124253558 0.124772563 0.125284505 0.125789526 0.126287765 0.126779359
#>  [97] 0.127264439 0.127743135 0.128215570 0.128681868 0.129142148 0.129596524
#> [103] 0.130045109 0.130488015 0.130925347 0.131357211 0.131783709 0.132204939
#> [109] 0.132620999 0.133031983 0.133437984 0.133839092 0.134235395 0.134626978
#> [115] 0.135013926 0.135396320 0.135774240 0.136147764 0.136516969 0.136881928
#> [121] 0.137242715 0.137599402 0.137952056 0.138300747 0.138645541 0.138986502
#> [127] 0.139323695 0.139657182 0.139987022 0.140313276 0.140636003 0.140955258
#> [133] 0.141271097 0.141583576 0.141892747 0.142198662 0.142501374 0.142800931
#> [139] 0.143097383 0.143390778 0.143681163 0.143968584 0.144253086 0.144534714
#> [145] 0.144813511 0.145089518 0.145362779 0.145633334 0.145901222 0.146166483
#> [151] 0.146429155 0.146689277 0.146946885 0.147202015 0.147454703 0.147704983
#> [157] 0.147952891 0.148198460 0.148441722 0.148682710 0.148921455 0.149157990
#> [163] 0.149392344 0.149624547 0.149854630 0.150082620 0.150308546 0.150532436
#> [169] 0.150754318 0.150974219 0.151192164 0.151408180 0.151622293 0.151834526
#> [175] 0.152044906 0.152253456 0.152460200 0.152665161 0.152868362 0.153069827
#> [181] 0.153269576 0.153467632 0.153664016 0.153858750 0.154051854 0.154243348
#> [187] 0.154433253 0.154621588 0.154808372 0.154993625 0.155177365 0.155359612
#> [193] 0.155540382 0.155719694 0.155897566 0.156074014 0.156249056 0.156422708
#> [199] 0.156594988 0.156765911 0.156935492 0.157103749 0.157270696 0.157436348
#> [205] 0.157600721 0.157763830 0.157925688 0.158086311 0.158245711 0.158403904
#> [211] 0.158560903 0.158716721 0.158871372 0.159024868 0.159177223 0.159328449
#> [217] 0.159478559 0.159627564 0.159775478 0.159922313 0.160068079 0.160212788
#> [223] 0.160356452 0.160499083 0.160640691 0.160781288 0.160920883 0.161059488
#> [229] 0.161197114 0.161333770 0.161469467 0.161604215 0.161738024 0.161870903
#> [235] 0.162002862 0.162133912 0.162264060 0.162393317 0.162521691 0.162649192
#> [241] 0.162775829 0.162901609 0.163026543 0.163150637 0.163273902 0.163396344
#> [247] 0.163517973 0.163638797 0.163758822 0.163878058 0.163996512 0.164114191
#> [253] 0.164231104 0.164347257 0.164462659 0.164577316 0.164691235 0.164804424
#> [259] 0.164916889 0.165028638 0.165139678 0.165250014 0.165359655 0.165468605
#> [265] 0.165576873 0.165684463 0.165791384 0.165897639 0.166003237 0.166108183
#> [271] 0.166212483 0.166316143 0.166419169 0.166521567 0.166623342 0.166724500
#> [277] 0.166825047 0.166924988 0.167024329 0.167123074 0.167221230 0.167318802
#> [283] 0.167415795 0.167512213 0.167608063 0.167703349 0.167798075 0.167892248
#> [289] 0.167985872 0.168078951 0.168171490 0.168263495 0.168354969 0.168445917
#> [295] 0.168536344 0.168626254 0.168715652 0.168804542 0.168892928 0.168980814
#> [301] 0.169068206 0.169155106 0.169241520 0.169327450 0.169412902 0.169497879
#> [307] 0.169582385 0.169666425 0.169750001 0.169833118 0.169915779 0.169997989
#> [313] 0.170079750 0.170161067 0.170241943 0.170322383 0.170402388 0.170481963
#> [319] 0.170561111 0.170639837 0.170718142 0.170796030 0.170873506 0.170950571
#> [325] 0.171027230 0.171103485 0.171179339 0.171254797 0.171329860 0.171404533
#> [331] 0.171478817 0.171552717 0.171626234 0.171699373 0.171772135 0.171844525
#> [337] 0.171916544 0.171988195 0.172059482 0.172130407 0.172200972 0.172271182
#> [343] 0.172341037 0.172410542 0.172479698 0.172548508 0.172616975 0.172685101
#> [349] 0.172752889 0.172820341 0.172887461 0.172954249 0.173020710 0.173086844
#> [355] 0.173152655 0.173218145 0.173283317 0.173348172 0.173412713 0.173476942
#> [361] 0.173540861 0.173604473 0.173667780 0.173730784 0.173793487 0.173855892
#> [367] 0.173918000 0.173979813 0.174041335 0.174102566 0.174163509 0.174224166
#> [373] 0.174284538 0.174344629 0.174404439 0.174463971 0.174523227 0.174582209
#> [379] 0.174640918 0.174699356 0.174757526 0.174815429 0.174873066 0.174930441
#> [385] 0.174987554 0.175044408 0.175101003 0.175157343 0.175213428 0.175269260
#> [391] 0.175324842 0.175380174 0.175435259 0.175490098 0.175544692 0.175599044
#> [397] 0.175653154 0.175707026 0.175760659 0.175814056 0.175867218 0.175920147
#> [403] 0.175972845 0.176025312 0.176077551 0.176129562 0.176181348 0.176232910
#> [409] 0.176284249 0.176335366 0.176386264 0.176436943 0.176487405 0.176537652
#> [415] 0.176587684 0.176637503 0.176687111 0.176736508 0.176785697 0.176834678
#> [421] 0.176883452 0.176932022 0.176980388 0.177028552 0.177076514 0.177124277
#> [427] 0.177171841 0.177219208 0.177266378 0.177313354 0.177360136 0.177406725
#> [433] 0.177453123 0.177499331 0.177545350 0.177591181 0.177636825 0.177682283
#> [439] 0.177727558 0.177772648 0.177817557 0.177862285 0.177906832 0.177951201
#> [445] 0.177995392 0.178039406 0.178083244 0.178126908 0.178170398 0.178213715
#> [451] 0.178256861 0.178299836 0.178342642 0.178385279 0.178427749 0.178470052
#> [457] 0.178512189 0.178554162 0.178595971 0.178637618 0.178679103 0.178720427
#> [463] 0.178761591 0.178802596 0.178843443 0.178884133 0.178924667 0.178965045
#> [469] 0.179005269 0.179045339 0.179085257 0.179125023 0.179164638 0.179204103
#> [475] 0.179243419 0.179282586 0.179321606 0.179360479 0.179399207 0.179437789
#> [481] 0.179476227 0.179514521 0.179552673 0.179590683 0.179628552 0.179666281
#> [487] 0.179703870 0.179741321 0.179778633 0.179815809 0.179852848 0.179889751
#> [493] 0.179926519 0.179963153 0.179999654 0.180036022 0.180072258 0.180108362
#> [499] 0.180144336 0.180180180