Calculate the depth of a light level

This function uses the Lamberte-Beere law to predict the depth in metres a given proportion of surface Photosynthetically Active Radiation (PAR) penetrates to.

depth(light = seq(0.01, 1, length = 500), KdPAR = 0.05, shade = 1)

Arguments

light
The proportion of surface irradiance (PAR), between 0 and 1. Defaults to a vector of light values.

KdPAR
The average light attenuation coefficient of the water column above the reef. Defaults to 0.05.

shade
Additional shading at the seafloor, such as bathymetric effects, between 0 (no light available) and 1 (all light available). Defaults to 1 (all light available).

Value

A depth in metres. When a single light value is provided, a single value is returned. You can pass a vector of light levels for a vector of depths.

Details

The function takes the average light attenuation coefficient (KdPAR) of the water column above a reef and a proportion between 0 and 1 of surface PAR. The Lamberte-Beer law is then used to predict the depth in metres this light environment occurs at.

A shading parameter can be used to scale how much of the light which reaches the seafloor is available to the benthos. According to (Lesser et al., 2018) the shading effect of reef slope at a vertical wall is to limit surface irradiance to 25% (shade = 0.25). The default behaviour is for there to be no shading (Shade = 1). When shading is more severe, the function will return a shallower depth. This is as the function is predicting the depth you would find your specified light level in the shaded environment.

Leaving shade unspecified makes this function equivalent to the "Depth_function" in (Laverick et al. 2020).

Specifying shading makes this function equivalent to the "Shade_function" in (Laverick et al., 2020).

References

Lamberte-Beer law

Lesser

Laverick

Examples

# Specifying a single light value and light
# attenuation returns a single depth.

   depth(light = 0.1, KdPAR = 0.03)

#> [1] 76.75284

# By defualt there is no shading. You can
# switch this on by passing a value.

   depth(light = 0.1, KdPAR = 0.03, shade = 0.8)

#> [1] 69.31472

# Specifying a vector of light values and light
# attenuation returns a vector of depths.

   depth(light = seq(0.0065, 1, length = 500), KdPAR = 0.03)

#>   [1] 167.86510340 158.95835414 151.93664066 146.13977696 141.20324877
#>   [6] 136.90437923 133.09712003 129.68048244 126.58171781 123.74668029
#>  [11] 121.13397940 118.71126362 116.45276615 114.33763086 112.34873867
#>  [16] 110.47186483 108.69506111 107.00819453 105.40259730 103.87079730
#>  [21] 102.40630778 101.00346137  99.65727765  98.36335643  97.11779096
#>  [26]  95.91709682  94.75815315  93.63815375  92.55456619  91.50509728
#>  [31]  90.48766388  89.50036798  88.54147542  87.60939753  86.70267533
#>  [36]  85.81996573  84.96002965  84.12172148  83.30397995  82.50582004
#>  [41]  81.72632582  80.96464419  80.21997922  79.49158715  78.77877195
#>  [46]  78.08088132  77.39730309  76.72746199  76.07081677  75.42685753
#>  [51]  74.79510339  74.17510027  73.56641900  72.96865349  72.38141912
#>  [56]  71.80435122  71.23710377  70.67934812  70.13077182  69.59107762
#>  [61]  69.05998247  68.53721665  68.02252292  67.51565578  67.01638078
#>  [66]  66.52447381  66.03972057  65.56191598  65.09086363  64.62637533
#>  [71]  64.16827067  63.71637654  63.27052682  62.83056192  62.39632854
#>  [76]  61.96767924  61.54447224  61.12657107  60.71384434  60.30616547
#>  [81]  59.90341248  59.50546775  59.11221783  58.72355323  58.33936825
#>  [86]  57.95956081  57.58403227  57.21268729  56.84543369  56.48218229
#>  [91]  56.12284680  55.76734370  55.41559209  55.06751364  54.72303241
#>  [96]  54.38207481  54.04456950  53.71044725  53.37964093  53.05208535
#> [101]  52.72771726  52.40647521  52.08829953  51.77313223  51.46091695
#> [106]  51.15159891  50.84512482  50.54144288  50.24050266  49.94225509
#> [111]  49.64665243  49.35364817  49.06319702  48.77525489  48.48977879
#> [116]  48.20672684  47.92605822  47.64773313  47.37171275  47.09795923
#> [121]  46.82643564  46.55710595  46.28993498  46.02488840  45.76193269
#> [126]  45.50103514  45.24216376  44.98528733  44.73037533  44.47739796
#> [131]  44.22632605  43.97713114  43.72978535  43.48426144  43.24053279
#> [136]  42.99857331  42.75835752  42.51986045  42.28305770  42.04792535
#> [141]  41.81444001  41.58257876  41.35231916  41.12363925  40.89651748
#> [146]  40.67093278  40.44686448  40.22429232  40.00319646  39.78355744
#> [151]  39.56535620  39.34857402  39.13319257  38.91919387  38.70656028
#> [156]  38.49527448  38.28531950  38.07667868  37.86933567  37.66327443
#> [161]  37.45847920  37.25493452  37.05262522  36.85153639  36.65165339
#> [166]  36.45296184  36.25544763  36.05909688  35.86389597  35.66983151
#> [171]  35.47689035  35.28505955  35.09432641  34.90467843  34.71610335
#> [176]  34.52858908  34.34212376  34.15669573  33.97229349  33.78890577
#> [181]  33.60652147  33.42512966  33.24471959  33.06528071  32.88680261
#> [186]  32.70927505  32.53268797  32.35703145  32.18229574  32.00847123
#> [191]  31.83554846  31.66351814  31.49237109  31.32209830  31.15269087
#> [196]  30.98414006  30.81643724  30.64957393  30.48354176  30.31833250
#> [201]  30.15393801  29.99035032  29.82756154  29.66556390  29.50434975
#> [206]  29.34391155  29.18424187  29.02533337  28.86717884  28.70977115
#> [211]  28.55310329  28.39716832  28.24195943  28.08746989  27.93369306
#> [216]  27.78062239  27.62825142  27.47657380  27.32558323  27.17527352
#> [221]  27.02563857  26.87667233  26.72836886  26.58072228  26.43372681
#> [226]  26.28737672  26.14166638  25.99659021  25.85214271  25.70831847
#> [231]  25.56511213  25.42251839  25.28053205  25.13914795  24.99836100
#> [236]  24.85816618  24.71855853  24.57953315  24.44108520  24.30320991
#> [241]  24.16590256  24.02915849  23.89297310  23.75734183  23.62226021
#> [246]  23.48772379  23.35372819  23.22026908  23.08734218  22.95494326
#> [251]  22.82306814  22.69171270  22.56087286  22.43054458  22.30072388
#> [256]  22.17140682  22.04258951  21.91426811  21.78643880  21.65909783
#> [261]  21.53224148  21.40586607  21.27996798  21.15454361  21.02958941
#> [266]  20.90510187  20.78107751  20.65751290  20.53440465  20.41174940
#> [271]  20.28954382  20.16778463  20.04646858  19.92559246  19.80515309
#> [276]  19.68514732  19.56557204  19.44642417  19.32770068  19.20939854
#> [281]  19.09151478  18.97404645  18.85699064  18.74034444  18.62410501
#> [286]  18.50826952  18.39283518  18.27779921  18.16315887  18.04891145
#> [291]  17.93505427  17.82158467  17.70850002  17.59579772  17.48347518
#> [296]  17.37152987  17.25995925  17.14876083  17.03793213  16.92747069
#> [301]  16.81737410  16.70763995  16.59826586  16.48924948  16.38058848
#> [306]  16.27228054  16.16432338  16.05671473  15.94945236  15.84253403
#> [311]  15.73595756  15.62972075  15.52382146  15.41825754  15.31302688
#> [316]  15.20812738  15.10355696  14.99931356  14.89539515  14.79179970
#> [321]  14.68852522  14.58556971  14.48293123  14.38060781  14.27859753
#> [326]  14.17689849  14.07550878  13.97442653  13.87364989  13.77317700
#> [331]  13.67300605  13.57313523  13.47356273  13.37428679  13.27530565
#> [336]  13.17661755  13.07822077  12.98011359  12.88229432  12.78476127
#> [341]  12.68751276  12.59054715  12.49386279  12.39745805  12.30133133
#> [346]  12.20548102  12.10990554  12.01460331  11.91957279  11.82481241
#> [351]  11.73032066  11.63609602  11.54213697  11.44844203  11.35500971
#> [356]  11.26183854  11.16892708  11.07627388  10.98387749  10.89173652
#> [361]  10.79984954  10.70821516  10.61683200  10.52569867  10.43481383
#> [366]  10.34417612  10.25378419  10.16363672  10.07373239   9.98406989
#> [371]   9.89464793   9.80546521   9.71652047   9.62781242   9.53933983
#> [376]   9.45110143   9.36309600   9.27532231   9.18777913   9.10046527
#> [381]   9.01337952   8.92652070   8.83988762   8.75347912   8.66729403
#> [386]   8.58133120   8.49558949   8.41006776   8.32476489   8.23967977
#> [391]   8.15481127   8.07015830   7.98571978   7.90149461   7.81748172
#> [396]   7.73368004   7.65008852   7.56670609   7.48353173   7.40056439
#> [401]   7.31780304   7.23524667   7.15289426   7.07074481   6.98879731
#> [406]   6.90705078   6.82550424   6.74415671   6.66300721   6.58205479
#> [411]   6.50129850   6.42073738   6.34037049   6.26019690   6.18021569
#> [416]   6.10042592   6.02082669   5.94141709   5.86219621   5.78316317
#> [421]   5.70431707   5.62565704   5.54718218   5.46889165   5.39078456
#> [426]   5.31286007   5.23511732   5.15755547   5.08017367   5.00297109
#> [431]   4.92594691   4.84910030   4.77243045   4.69593654   4.61961776
#> [436]   4.54347333   4.46750243   4.39170429   4.31607812   4.24062315
#> [441]   4.16533858   4.09022367   4.01527765   3.94049976   3.86588924
#> [446]   3.79144535   3.71716735   3.64305449   3.56910605   3.49532130
#> [451]   3.42169952   3.34823998   3.27494198   3.20180480   3.12882773
#> [456]   3.05601009   2.98335118   2.91085030   2.83850677   2.76631991
#> [461]   2.69428904   2.62241348   2.55069258   2.47912566   2.40771207
#> [466]   2.33645114   2.26534224   2.19438470   2.12357790   2.05292118
#> [471]   1.98241392   1.91205548   1.84184524   1.77178257   1.70186686
#> [476]   1.63209749   1.56247384   1.49299531   1.42366131   1.35447121
#> [481]   1.28542444   1.21652040   1.14775850   1.07913815   1.01065877
#> [486]   0.94231978   0.87412062   0.80606071   0.73813947   0.67035636
#> [491]   0.60271080   0.53520224   0.46783013   0.40059391   0.33349304
#> [496]   0.26652698   0.19969518   0.13299710   0.06643222   0.00000000

# By default the function uses the reference light
# environment from (Laverick et al., 2020) with no shade.

   depth()

#>   [1] 92.10340372 88.48371056 85.41981742 82.76354192 80.41905127 78.32078150
#>   [7] 76.42189587 74.68776868 73.09207782 71.61434511 70.23832498 68.95091280
#>  [13] 67.74138398 66.60085041 65.52186364 64.49811988 63.52423681 62.59558231
#>  [19] 61.70814119 60.85841030 60.04331499 59.26014196 58.50648476 57.78019920
#>  [25] 57.07936657 56.40226309 55.74733439 55.11317398 54.49850508 53.90216517
#>  [31] 53.32309270 52.76031576 52.21294227 51.68015145 51.16118641 50.65534773
#>  [37] 50.16198772 49.68050550 49.21034256 48.75097891 48.30192954 47.86274144
#>  [43] 47.43299078 47.01228048 46.60023802 46.19651343 45.80077753 45.41272028
#>  [49] 45.03204936 44.65848884 44.29177794 43.93167001 43.57793144 43.23034084
#>  [55] 42.88868816 42.55277391 42.22240851 41.89741161 41.57761154 41.26284469
#>  [61] 40.95295508 40.64779389 40.34721897 40.05109452 39.75929065 39.47168311
#>  [67] 39.18815291 38.90858606 38.63287327 38.36090974 38.09259485 37.82783201
#>  [73] 37.56652839 37.30859477 37.05394533 36.80249748 36.55417172 36.30889147
#>  [79] 36.06658293 35.82717496 35.59059893 35.35678863 35.12568013 34.89721171
#>  [85] 34.67132374 34.44795856 34.22706046 34.00857553 33.79245162 33.57863823
#>  [91] 33.36708649 33.15774907 32.95058007 32.74553504 32.54257086 32.34164574
#>  [97] 32.14271910 31.94575157 31.75070496 31.55754215 31.36622710 31.17672480
#> [103] 30.98900121 30.80302326 30.61875878 30.43617648 30.25524592 30.07593750
#> [109] 29.89822238 29.72207249 29.54746051 29.37435981 29.20274445 29.03258917
#> [115] 28.86386933 28.69656090 28.53064048 28.36608522 28.20287284 28.04098160
#> [121] 27.88039028 27.72107818 27.56302507 27.40621122 27.25061733 27.09622458
#> [127] 26.94301456 26.79096929 26.64007119 26.49030309 26.34164818 26.19409003
#> [133] 26.04761259 25.90220013 25.75783728 25.61450901 25.47220058 25.33089758
#> [139] 25.19058591 25.05125176 24.91288159 24.77546216 24.63898050 24.50342390
#> [145] 24.36877989 24.23503627 24.10218108 23.97020260 23.83908933 23.70882999
#> [151] 23.57941355 23.45082915 23.32306618 23.19611419 23.06996296 22.94460246
#> [157] 22.82002283 22.69621440 22.57316768 22.45087336 22.32932229 22.20850550
#> [163] 22.08841416 21.96903962 21.85037336 21.73240704 21.61513244 21.49854150
#> [169] 21.38262630 21.26737904 21.15279208 21.03885789 20.92556907 20.81291836
#> [175] 20.70089861 20.58950278 20.47872397 20.36855538 20.25899032 20.15002222
#> [181] 20.04164460 19.93385111 19.82663547 19.71999154 19.61391323 19.50839458
#> [187] 19.40342973 19.29901288 19.19513835 19.09180052 18.98899389 18.88671301
#> [193] 18.78495254 18.68370722 18.58297184 18.48274131 18.38301057 18.28377468
#> [199] 18.18502875 18.08676797 17.98898758 17.89168292 17.79484938 17.69848241
#> [205] 17.60257755 17.50713038 17.41213656 17.31759180 17.22349187 17.12983261
#> [211] 17.03660991 16.94381971 16.85145803 16.75952093 16.66800451 16.57690495
#> [217] 16.48621846 16.39594132 16.30606985 16.21660042 16.12752944 16.03885339
#> [223] 15.95056877 15.86267215 15.77516014 15.68802937 15.60127654 15.51489840
#> [229] 15.42889170 15.34325329 15.25798000 15.17306875 15.08851648 15.00432015
#> [235] 14.92047679 14.83698345 14.75383721 14.67103521 14.58857461 14.50645259
#> [241] 14.42466640 14.34321330 14.26209058 14.18129558 14.10082566 14.02067821
#> [247] 13.94085066 13.86134047 13.78214512 13.70326212 13.62468904 13.54642343
#> [253] 13.46846290 13.39080508 13.31344764 13.23638825 13.15962463 13.08315452
#> [259] 13.00697567 12.93108589 12.85548298 12.78016479 12.70512917 12.63037402
#> [265] 12.55589725 12.48169678 12.40777059 12.33411664 12.26073294 12.18761752
#> [271] 12.11476842 12.04218370 11.96986146 11.89779980 11.82599686 11.75445077
#> [277] 11.68315972 11.61212188 11.54133547 11.47079872 11.40050986 11.33046716
#> [283] 11.26066890 11.19111339 11.12179894 11.05272388 10.98388657 10.91528538
#> [289] 10.84691869 10.77878490 10.71088244 10.64320973 10.57576523 10.50854741
#> [295] 10.44155474 10.37478572 10.30823886 10.24191270 10.17580577 10.10991662
#> [301] 10.04424383  9.97878598  9.91354167  9.84850951  9.78368812  9.71907614
#> [307]  9.65467223  9.59047504  9.52648326  9.46269557  9.39911068  9.33572731
#> [313]  9.27254417  9.20956000  9.14677357  9.08418362  9.02178894  8.95958831
#> [319]  8.89758053  8.83576440  8.77413874  8.71270238  8.65145417  8.59039296
#> [325]  8.52951759  8.46882696  8.40831994  8.34799542  8.28785231  8.22788952
#> [331]  8.16810596  8.10850058  8.04907230  7.98982009  7.93074291  7.87183971
#> [337]  7.81310949  7.75455122  7.69616391  7.63794655  7.57989817  7.52201777
#> [343]  7.46430441  7.40675710  7.34937490  7.29215687  7.23510207  7.17820957
#> [349]  7.12147844  7.06490779  7.00849669  6.95224425  6.89614959  6.84021182
#> [355]  6.78443007  6.72880346  6.67333114  6.61801225  6.56284595  6.50783140
#> [361]  6.45296776  6.39825421  6.34368994  6.28927412  6.23500595  6.18088464
#> [367]  6.12690938  6.07307940  6.01939392  5.96585216  5.91245335  5.85919673
#> [373]  5.80608155  5.75310705  5.70027250  5.64757716  5.59502030  5.54260118
#> [379]  5.49031909  5.43817331  5.38616315  5.33428788  5.28254682  5.23093927
#> [385]  5.17946454  5.12812195  5.07691084  5.02583051  4.97488032  4.92405958
#> [391]  4.87336766  4.82280390  4.77236765  4.72205828  4.67187513  4.62181759
#> [397]  4.57188502  4.52207681  4.47239233  4.42283097  4.37339213  4.32407519
#> [403]  4.27487957  4.22580465  4.17684986  4.12801461  4.07929831  4.03070038
#> [409]  3.98222026  3.93385736  3.88561114  3.83748102  3.78946645  3.74156687
#> [415]  3.69378173  3.64611050  3.59855262  3.55110756  3.50377479  3.45655377
#> [421]  3.40944398  3.36244489  3.31555600  3.26877677  3.22210670  3.17554529
#> [427]  3.12909202  3.08274640  3.03650792  2.99037610  2.94435044  2.89843045
#> [433]  2.85261566  2.80690558  2.76129972  2.71579763  2.67039882  2.62510284
#> [439]  2.57990920  2.53481746  2.48982716  2.44493783  2.40014903  2.35546031
#> [445]  2.31087122  2.26638132  2.22199017  2.17769733  2.13350236  2.08940484
#> [451]  2.04540433  2.00150041  1.95769266  1.91398066  1.87036398  1.82684222
#> [457]  1.78341496  1.74008179  1.69684231  1.65369611  1.61064279  1.56768195
#> [463]  1.52481319  1.48203612  1.43935035  1.39675549  1.35425115  1.31183696
#> [469]  1.26951252  1.22727746  1.18513140  1.14307397  1.10110480  1.05922351
#> [475]  1.01742974  0.97572312  0.93410330  0.89256991  0.85112258  0.80976098
#> [481]  0.76848474  0.72729351  0.68618694  0.64516468  0.60422639  0.56337173
#> [487]  0.52260036  0.48191193  0.44130611  0.40078256  0.36034096  0.31998097
#> [493]  0.27970226  0.23950451  0.19938738  0.15935057  0.11939374  0.07951658
#> [499]  0.03971877  0.00000000