Calculate the mesophotic community value for a given light level

This function captures the preferred light environment of the mesophotic coral community.

mesophotic(light = seq(0.0065, 100, length = 500), a = 1.54, b = 8.92)

Arguments

light
The percentage of surface irradiance (PAR) reaching a reef, positive and up to 100.

a
The scale parameter of a Weibull distribution, defaults to (Tamir et al., 2019). a describes the spread of the data, or ‘the peakyness’. Larger numbers create a flatter, wider, curve. Varying a will simultaneously change the width of the preferred light environment of mesophotic taxa, and the maximum similarity a reef patch is expected to achieve in comparison to the idealised mesophotic community.

b
The shape parameter of a Weibull distribution, defaults to (Tamir et al., 2019). A value of 3 approximates a normal distribution. Values larger than 3 introduce a left-skew to the curve. This means mesophotic taxa can be light-limited, while keeping mesophotic communities in low-light environments. Varying b shifts the preferred light environment of mesophotic communities.

Value

A mesophotic community indicator value, as discussed in (Laverick et al., 2017). When a single light value is provided, a single value is returned. You can pass a vector of light levels for a vector of depths.

Details

A Weibull distribution is used to represent the relationship between mesophotic coral communities and light. We assume mesophotic community values (how similar a reef patch is to the idealised mesophotic community) center on a preferred light value, between light limitation and light-induced stress.

This relationship has been verified in the Red Sea (Tamir et al., 2019), but certainly needs more research!

References

(Tamir et al., 2019)

(Laverick et al., 2017)

See also

Examples

# Specifying a single light value
# returns a single community value.

   mesophotic(light = 10)

#> [1] 0.05572972

# Specifying a vector of light values
# returns a vector of community values.

   mesophotic(light = seq(0.0065, 100, length = 500))

#>   [1] 3.490806e-03 2.254945e-02 3.232612e-02 3.982920e-02 4.604889e-02
#>   [6] 5.138064e-02 5.603171e-02 6.012786e-02 6.375299e-02 6.696703e-02
#>  [11] 6.981510e-02 7.233260e-02 7.454827e-02 7.648615e-02 7.816680e-02
#>  [16] 7.960820e-02 8.082631e-02 8.183553e-02 8.264901e-02 8.327887e-02
#>  [21] 8.373638e-02 8.403211e-02 8.417598e-02 8.417741e-02 8.404534e-02
#>  [26] 8.378828e-02 8.341436e-02 8.293135e-02 8.234670e-02 8.166756e-02
#>  [31] 8.090077e-02 8.005291e-02 7.913029e-02 7.813896e-02 7.708474e-02
#>  [36] 7.597319e-02 7.480966e-02 7.359925e-02 7.234685e-02 7.105714e-02
#>  [41] 6.973459e-02 6.838344e-02 6.700776e-02 6.561141e-02 6.419804e-02
#>  [46] 6.277114e-02 6.133400e-02 5.988973e-02 5.844128e-02 5.699140e-02
#>  [51] 5.554271e-02 5.409765e-02 5.265849e-02 5.122739e-02 4.980632e-02
#>  [56] 4.839712e-02 4.700152e-02 4.562109e-02 4.425727e-02 4.291139e-02
#>  [61] 4.158466e-02 4.027818e-02 3.899293e-02 3.772980e-02 3.648957e-02
#>  [66] 3.527294e-02 3.408050e-02 3.291277e-02 3.177019e-02 3.065310e-02
#>  [71] 2.956180e-02 2.849651e-02 2.745737e-02 2.644447e-02 2.545784e-02
#>  [76] 2.449748e-02 2.356330e-02 2.265520e-02 2.177301e-02 2.091653e-02
#>  [81] 2.008554e-02 1.927975e-02 1.849887e-02 1.774258e-02 1.701051e-02
#>  [86] 1.630229e-02 1.561752e-02 1.495579e-02 1.431666e-02 1.369968e-02
#>  [91] 1.310441e-02 1.253036e-02 1.197705e-02 1.144402e-02 1.093075e-02
#>  [96] 1.043676e-02 9.961546e-03 9.504615e-03 9.065463e-03 8.643593e-03
#> [101] 8.238508e-03 7.849713e-03 7.476717e-03 7.119036e-03 6.776188e-03
#> [106] 6.447698e-03 6.133097e-03 5.831924e-03 5.543725e-03 5.268054e-03
#> [111] 5.004472e-03 4.752551e-03 4.511869e-03 4.282017e-03 4.062591e-03
#> [116] 3.853201e-03 3.653462e-03 3.463003e-03 3.281461e-03 3.108482e-03
#> [121] 2.943723e-03 2.786851e-03 2.637543e-03 2.495485e-03 2.360373e-03
#> [126] 2.231914e-03 2.109822e-03 1.993823e-03 1.883650e-03 1.779047e-03
#> [131] 1.679766e-03 1.585569e-03 1.496224e-03 1.411511e-03 1.331216e-03
#> [136] 1.255134e-03 1.183067e-03 1.114825e-03 1.050227e-03 9.890965e-04
#> [141] 9.312672e-04 8.765777e-04 8.248737e-04 7.760076e-04 7.298381e-04
#> [146] 6.862298e-04 6.450533e-04 6.061849e-04 5.695063e-04 5.349047e-04
#> [151] 5.022723e-04 4.715061e-04 4.425081e-04 4.151848e-04 3.894470e-04
#> [156] 3.652098e-04 3.423923e-04 3.209177e-04 3.007127e-04 2.817076e-04
#> [161] 2.638364e-04 2.470362e-04 2.312472e-04 2.164127e-04 2.024790e-04
#> [166] 1.893950e-04 1.771123e-04 1.655849e-04 1.547694e-04 1.446247e-04
#> [171] 1.351117e-04 1.261934e-04 1.178351e-04 1.100035e-04 1.026675e-04
#> [176] 9.579754e-05 8.936573e-05 8.334572e-05 7.771261e-05 7.244292e-05
#> [181] 6.751449e-05 6.290642e-05 5.859899e-05 5.457364e-05 5.081286e-05
#> [186] 4.730015e-05 4.401998e-05 4.095774e-05 3.809968e-05 3.543285e-05
#> [191] 3.294508e-05 3.062492e-05 2.846163e-05 2.644510e-05 2.456583e-05
#> [196] 2.281491e-05 2.118399e-05 1.966520e-05 1.825118e-05 1.693503e-05
#> [201] 1.571026e-05 1.457081e-05 1.351099e-05 1.252547e-05 1.160926e-05
#> [206] 1.075768e-05 9.966373e-06 9.231238e-06 8.548451e-06 7.914434e-06
#> [211] 7.325839e-06 6.779542e-06 6.272619e-06 5.802343e-06 5.366165e-06
#> [216] 4.961706e-06 4.586745e-06 4.239211e-06 3.917171e-06 3.618823e-06
#> [221] 3.342487e-06 3.086596e-06 2.849692e-06 2.630415e-06 2.427498e-06
#> [226] 2.239765e-06 2.066116e-06 1.905531e-06 1.757060e-06 1.619821e-06
#> [231] 1.492990e-06 1.375804e-06 1.267555e-06 1.167581e-06 1.075271e-06
#> [236] 9.900561e-07 9.114073e-07 8.388347e-07 7.718833e-07 7.101310e-07
#> [241] 6.531863e-07 6.006861e-07 5.522940e-07 5.076979e-07 4.666089e-07
#> [246] 4.287591e-07 3.939006e-07 3.618037e-07 3.322559e-07 3.050604e-07
#> [251] 2.800352e-07 2.570119e-07 2.358348e-07 2.163599e-07 1.984541e-07
#> [256] 1.819944e-07 1.668670e-07 1.529671e-07 1.401976e-07 1.284690e-07
#> [261] 1.176987e-07 1.078104e-07 9.873368e-08 9.040364e-08 8.276040e-08
#> [266] 7.574877e-08 6.931783e-08 6.342069e-08 5.801413e-08 5.305832e-08
#> [271] 4.851659e-08 4.435517e-08 4.054298e-08 3.705141e-08 3.385411e-08
#> [276] 3.092688e-08 2.824741e-08 2.579524e-08 2.355150e-08 2.149890e-08
#> [281] 1.962152e-08 1.790473e-08 1.633510e-08 1.490029e-08 1.358899e-08
#> [286] 1.239079e-08 1.129614e-08 1.029630e-08 9.383219e-09 8.549536e-09
#> [291] 7.788492e-09 7.093890e-09 6.460051e-09 5.881768e-09 5.354274e-09
#> [296] 4.873198e-09 4.434539e-09 4.034632e-09 3.670123e-09 3.337941e-09
#> [301] 3.035276e-09 2.759557e-09 2.508432e-09 2.279749e-09 2.071543e-09
#> [306] 1.882014e-09 1.709520e-09 1.552558e-09 1.409756e-09 1.279861e-09
#> [311] 1.161728e-09 1.054312e-09 9.566580e-10 8.678956e-10 7.872299e-10
#> [316] 7.139357e-10 6.473515e-10 5.868740e-10 5.319531e-10 4.820874e-10
#> [321] 4.368197e-10 3.957335e-10 3.584493e-10 3.246213e-10 2.939346e-10
#> [326] 2.661026e-10 2.408642e-10 2.179818e-10 1.972392e-10 1.784397e-10
#> [331] 1.614041e-10 1.459699e-10 1.319889e-10 1.193265e-10 1.078604e-10
#> [336] 9.747949e-11 8.808261e-11 7.957801e-11 7.188233e-11 6.491983e-11
#> [341] 5.862177e-11 5.292574e-11 4.777509e-11 4.311841e-11 3.890906e-11
#> [346] 3.510472e-11 3.166703e-11 2.856119e-11 2.575564e-11 2.322179e-11
#> [351] 2.093372e-11 1.886795e-11 1.700320e-11 1.532019e-11 1.380147e-11
#> [356] 1.243125e-11 1.119520e-11 1.008039e-11 9.075093e-12 8.168701e-12
#> [361] 7.351625e-12 6.615188e-12 5.951544e-12 5.353599e-12 4.814939e-12
#> [366] 4.329769e-12 3.892849e-12 3.499448e-12 3.145290e-12 2.826514e-12
#> [371] 2.539634e-12 2.281500e-12 2.049271e-12 1.840382e-12 1.652518e-12
#> [376] 1.483592e-12 1.331719e-12 1.195201e-12 1.072504e-12 9.622494e-13
#> [381] 8.631902e-13 7.742047e-13 6.942815e-13 6.225095e-13 5.580678e-13
#> [386] 5.002174e-13 4.482925e-13 4.016938e-13 3.598817e-13 3.223707e-13
#> [391] 2.887238e-13 2.585478e-13 2.314891e-13 2.072295e-13 1.854831e-13
#> [396] 1.659926e-13 1.485268e-13 1.328779e-13 1.188591e-13 1.063027e-13
#> [401] 9.505787e-14 8.498926e-14 7.597528e-14 6.790673e-14 6.068562e-14
#> [406] 5.422396e-14 4.844280e-14 4.327129e-14 3.864588e-14 3.450956e-14
#> [411] 3.081119e-14 2.750493e-14 2.454967e-14 2.190856e-14 1.954859e-14
#> [416] 1.744014e-14 1.555673e-14 1.387458e-14 1.237243e-14 1.103122e-14
#> [421] 9.833903e-15 8.765204e-15 7.811456e-15 6.960427e-15 6.201172e-15
#> [426] 5.523899e-15 4.919850e-15 4.381191e-15 3.900917e-15 3.472768e-15
#> [431] 3.091144e-15 2.751041e-15 2.447990e-15 2.177995e-15 1.937487e-15
#> [436] 1.723279e-15 1.532524e-15 1.362680e-15 1.211478e-15 1.076893e-15
#> [441] 9.571154e-16 8.505337e-16 7.557081e-16 6.713546e-16 5.963283e-16
#> [446] 5.296077e-16 4.702826e-16 4.175410e-16 3.706595e-16 3.289932e-16
#> [451] 2.919675e-16 2.590707e-16 2.298465e-16 2.038890e-16 1.808363e-16
#> [456] 1.603666e-16 1.421930e-16 1.260605e-16 1.117420e-16 9.903536e-17
#> [461] 8.776082e-17 7.775846e-17 6.888607e-17 6.101715e-17 5.403924e-17
#> [466] 4.785238e-17 4.236769e-17 3.750620e-17 3.319773e-17 2.937994e-17
#> [471] 2.599745e-17 2.300106e-17 2.034709e-17 1.799676e-17 1.591562e-17
#> [476] 1.407312e-17 1.244214e-17 1.099860e-17 9.721144e-18 8.590831e-18
#> [481] 7.590858e-18 6.706324e-18 5.924015e-18 5.232218e-18 4.620550e-18
#> [486] 4.079808e-18 3.601836e-18 3.179410e-18 2.806128e-18 2.476321e-18
#> [491] 2.184966e-18 1.927619e-18 1.700342e-18 1.499650e-18 1.322460e-18
#> [496] 1.166041e-18 1.027978e-18 9.061350e-19 7.986214e-19 7.037656e-19

# By default the function uses the shallow community
# light relationship from (Tamir et al., 2019) for a
# range of light values.

   mesophotic()

#>   [1] 3.490806e-03 2.254945e-02 3.232612e-02 3.982920e-02 4.604889e-02
#>   [6] 5.138064e-02 5.603171e-02 6.012786e-02 6.375299e-02 6.696703e-02
#>  [11] 6.981510e-02 7.233260e-02 7.454827e-02 7.648615e-02 7.816680e-02
#>  [16] 7.960820e-02 8.082631e-02 8.183553e-02 8.264901e-02 8.327887e-02
#>  [21] 8.373638e-02 8.403211e-02 8.417598e-02 8.417741e-02 8.404534e-02
#>  [26] 8.378828e-02 8.341436e-02 8.293135e-02 8.234670e-02 8.166756e-02
#>  [31] 8.090077e-02 8.005291e-02 7.913029e-02 7.813896e-02 7.708474e-02
#>  [36] 7.597319e-02 7.480966e-02 7.359925e-02 7.234685e-02 7.105714e-02
#>  [41] 6.973459e-02 6.838344e-02 6.700776e-02 6.561141e-02 6.419804e-02
#>  [46] 6.277114e-02 6.133400e-02 5.988973e-02 5.844128e-02 5.699140e-02
#>  [51] 5.554271e-02 5.409765e-02 5.265849e-02 5.122739e-02 4.980632e-02
#>  [56] 4.839712e-02 4.700152e-02 4.562109e-02 4.425727e-02 4.291139e-02
#>  [61] 4.158466e-02 4.027818e-02 3.899293e-02 3.772980e-02 3.648957e-02
#>  [66] 3.527294e-02 3.408050e-02 3.291277e-02 3.177019e-02 3.065310e-02
#>  [71] 2.956180e-02 2.849651e-02 2.745737e-02 2.644447e-02 2.545784e-02
#>  [76] 2.449748e-02 2.356330e-02 2.265520e-02 2.177301e-02 2.091653e-02
#>  [81] 2.008554e-02 1.927975e-02 1.849887e-02 1.774258e-02 1.701051e-02
#>  [86] 1.630229e-02 1.561752e-02 1.495579e-02 1.431666e-02 1.369968e-02
#>  [91] 1.310441e-02 1.253036e-02 1.197705e-02 1.144402e-02 1.093075e-02
#>  [96] 1.043676e-02 9.961546e-03 9.504615e-03 9.065463e-03 8.643593e-03
#> [101] 8.238508e-03 7.849713e-03 7.476717e-03 7.119036e-03 6.776188e-03
#> [106] 6.447698e-03 6.133097e-03 5.831924e-03 5.543725e-03 5.268054e-03
#> [111] 5.004472e-03 4.752551e-03 4.511869e-03 4.282017e-03 4.062591e-03
#> [116] 3.853201e-03 3.653462e-03 3.463003e-03 3.281461e-03 3.108482e-03
#> [121] 2.943723e-03 2.786851e-03 2.637543e-03 2.495485e-03 2.360373e-03
#> [126] 2.231914e-03 2.109822e-03 1.993823e-03 1.883650e-03 1.779047e-03
#> [131] 1.679766e-03 1.585569e-03 1.496224e-03 1.411511e-03 1.331216e-03
#> [136] 1.255134e-03 1.183067e-03 1.114825e-03 1.050227e-03 9.890965e-04
#> [141] 9.312672e-04 8.765777e-04 8.248737e-04 7.760076e-04 7.298381e-04
#> [146] 6.862298e-04 6.450533e-04 6.061849e-04 5.695063e-04 5.349047e-04
#> [151] 5.022723e-04 4.715061e-04 4.425081e-04 4.151848e-04 3.894470e-04
#> [156] 3.652098e-04 3.423923e-04 3.209177e-04 3.007127e-04 2.817076e-04
#> [161] 2.638364e-04 2.470362e-04 2.312472e-04 2.164127e-04 2.024790e-04
#> [166] 1.893950e-04 1.771123e-04 1.655849e-04 1.547694e-04 1.446247e-04
#> [171] 1.351117e-04 1.261934e-04 1.178351e-04 1.100035e-04 1.026675e-04
#> [176] 9.579754e-05 8.936573e-05 8.334572e-05 7.771261e-05 7.244292e-05
#> [181] 6.751449e-05 6.290642e-05 5.859899e-05 5.457364e-05 5.081286e-05
#> [186] 4.730015e-05 4.401998e-05 4.095774e-05 3.809968e-05 3.543285e-05
#> [191] 3.294508e-05 3.062492e-05 2.846163e-05 2.644510e-05 2.456583e-05
#> [196] 2.281491e-05 2.118399e-05 1.966520e-05 1.825118e-05 1.693503e-05
#> [201] 1.571026e-05 1.457081e-05 1.351099e-05 1.252547e-05 1.160926e-05
#> [206] 1.075768e-05 9.966373e-06 9.231238e-06 8.548451e-06 7.914434e-06
#> [211] 7.325839e-06 6.779542e-06 6.272619e-06 5.802343e-06 5.366165e-06
#> [216] 4.961706e-06 4.586745e-06 4.239211e-06 3.917171e-06 3.618823e-06
#> [221] 3.342487e-06 3.086596e-06 2.849692e-06 2.630415e-06 2.427498e-06
#> [226] 2.239765e-06 2.066116e-06 1.905531e-06 1.757060e-06 1.619821e-06
#> [231] 1.492990e-06 1.375804e-06 1.267555e-06 1.167581e-06 1.075271e-06
#> [236] 9.900561e-07 9.114073e-07 8.388347e-07 7.718833e-07 7.101310e-07
#> [241] 6.531863e-07 6.006861e-07 5.522940e-07 5.076979e-07 4.666089e-07
#> [246] 4.287591e-07 3.939006e-07 3.618037e-07 3.322559e-07 3.050604e-07
#> [251] 2.800352e-07 2.570119e-07 2.358348e-07 2.163599e-07 1.984541e-07
#> [256] 1.819944e-07 1.668670e-07 1.529671e-07 1.401976e-07 1.284690e-07
#> [261] 1.176987e-07 1.078104e-07 9.873368e-08 9.040364e-08 8.276040e-08
#> [266] 7.574877e-08 6.931783e-08 6.342069e-08 5.801413e-08 5.305832e-08
#> [271] 4.851659e-08 4.435517e-08 4.054298e-08 3.705141e-08 3.385411e-08
#> [276] 3.092688e-08 2.824741e-08 2.579524e-08 2.355150e-08 2.149890e-08
#> [281] 1.962152e-08 1.790473e-08 1.633510e-08 1.490029e-08 1.358899e-08
#> [286] 1.239079e-08 1.129614e-08 1.029630e-08 9.383219e-09 8.549536e-09
#> [291] 7.788492e-09 7.093890e-09 6.460051e-09 5.881768e-09 5.354274e-09
#> [296] 4.873198e-09 4.434539e-09 4.034632e-09 3.670123e-09 3.337941e-09
#> [301] 3.035276e-09 2.759557e-09 2.508432e-09 2.279749e-09 2.071543e-09
#> [306] 1.882014e-09 1.709520e-09 1.552558e-09 1.409756e-09 1.279861e-09
#> [311] 1.161728e-09 1.054312e-09 9.566580e-10 8.678956e-10 7.872299e-10
#> [316] 7.139357e-10 6.473515e-10 5.868740e-10 5.319531e-10 4.820874e-10
#> [321] 4.368197e-10 3.957335e-10 3.584493e-10 3.246213e-10 2.939346e-10
#> [326] 2.661026e-10 2.408642e-10 2.179818e-10 1.972392e-10 1.784397e-10
#> [331] 1.614041e-10 1.459699e-10 1.319889e-10 1.193265e-10 1.078604e-10
#> [336] 9.747949e-11 8.808261e-11 7.957801e-11 7.188233e-11 6.491983e-11
#> [341] 5.862177e-11 5.292574e-11 4.777509e-11 4.311841e-11 3.890906e-11
#> [346] 3.510472e-11 3.166703e-11 2.856119e-11 2.575564e-11 2.322179e-11
#> [351] 2.093372e-11 1.886795e-11 1.700320e-11 1.532019e-11 1.380147e-11
#> [356] 1.243125e-11 1.119520e-11 1.008039e-11 9.075093e-12 8.168701e-12
#> [361] 7.351625e-12 6.615188e-12 5.951544e-12 5.353599e-12 4.814939e-12
#> [366] 4.329769e-12 3.892849e-12 3.499448e-12 3.145290e-12 2.826514e-12
#> [371] 2.539634e-12 2.281500e-12 2.049271e-12 1.840382e-12 1.652518e-12
#> [376] 1.483592e-12 1.331719e-12 1.195201e-12 1.072504e-12 9.622494e-13
#> [381] 8.631902e-13 7.742047e-13 6.942815e-13 6.225095e-13 5.580678e-13
#> [386] 5.002174e-13 4.482925e-13 4.016938e-13 3.598817e-13 3.223707e-13
#> [391] 2.887238e-13 2.585478e-13 2.314891e-13 2.072295e-13 1.854831e-13
#> [396] 1.659926e-13 1.485268e-13 1.328779e-13 1.188591e-13 1.063027e-13
#> [401] 9.505787e-14 8.498926e-14 7.597528e-14 6.790673e-14 6.068562e-14
#> [406] 5.422396e-14 4.844280e-14 4.327129e-14 3.864588e-14 3.450956e-14
#> [411] 3.081119e-14 2.750493e-14 2.454967e-14 2.190856e-14 1.954859e-14
#> [416] 1.744014e-14 1.555673e-14 1.387458e-14 1.237243e-14 1.103122e-14
#> [421] 9.833903e-15 8.765204e-15 7.811456e-15 6.960427e-15 6.201172e-15
#> [426] 5.523899e-15 4.919850e-15 4.381191e-15 3.900917e-15 3.472768e-15
#> [431] 3.091144e-15 2.751041e-15 2.447990e-15 2.177995e-15 1.937487e-15
#> [436] 1.723279e-15 1.532524e-15 1.362680e-15 1.211478e-15 1.076893e-15
#> [441] 9.571154e-16 8.505337e-16 7.557081e-16 6.713546e-16 5.963283e-16
#> [446] 5.296077e-16 4.702826e-16 4.175410e-16 3.706595e-16 3.289932e-16
#> [451] 2.919675e-16 2.590707e-16 2.298465e-16 2.038890e-16 1.808363e-16
#> [456] 1.603666e-16 1.421930e-16 1.260605e-16 1.117420e-16 9.903536e-17
#> [461] 8.776082e-17 7.775846e-17 6.888607e-17 6.101715e-17 5.403924e-17
#> [466] 4.785238e-17 4.236769e-17 3.750620e-17 3.319773e-17 2.937994e-17
#> [471] 2.599745e-17 2.300106e-17 2.034709e-17 1.799676e-17 1.591562e-17
#> [476] 1.407312e-17 1.244214e-17 1.099860e-17 9.721144e-18 8.590831e-18
#> [481] 7.590858e-18 6.706324e-18 5.924015e-18 5.232218e-18 4.620550e-18
#> [486] 4.079808e-18 3.601836e-18 3.179410e-18 2.806128e-18 2.476321e-18
#> [491] 2.184966e-18 1.927619e-18 1.700342e-18 1.499650e-18 1.322460e-18
#> [496] 1.166041e-18 1.027978e-18 9.061350e-19 7.986214e-19 7.037656e-19

# You can specify your own community light relationship
# by setting parameters explicitly

   mesophotic(a = 1.2, b = 8)

#>   [1] 3.613856e-02 7.132031e-02 8.040122e-02 8.560628e-02 8.896673e-02
#>   [6] 9.120617e-02 9.267809e-02 9.358625e-02 9.406104e-02 9.419208e-02
#>  [11] 9.404436e-02 9.366690e-02 9.309789e-02 9.236781e-02 9.150151e-02
#>  [16] 9.051960e-02 8.943938e-02 8.827560e-02 8.704088e-02 8.574618e-02
#>  [21] 8.440105e-02 8.301387e-02 8.159201e-02 8.014202e-02 7.866969e-02
#>  [26] 7.718021e-02 7.567818e-02 7.416774e-02 7.265260e-02 7.113607e-02
#>  [31] 6.962113e-02 6.811045e-02 6.660644e-02 6.511124e-02 6.362679e-02
#>  [36] 6.215480e-02 6.069682e-02 5.925422e-02 5.782822e-02 5.641990e-02
#>  [41] 5.503021e-02 5.365998e-02 5.230995e-02 5.098074e-02 4.967290e-02
#>  [46] 4.838687e-02 4.712305e-02 4.588173e-02 4.466318e-02 4.346757e-02
#>  [51] 4.229505e-02 4.114569e-02 4.001954e-02 3.891660e-02 3.783682e-02
#>  [56] 3.678013e-02 3.574643e-02 3.473558e-02 3.374741e-02 3.278175e-02
#>  [61] 3.183839e-02 3.091710e-02 3.001764e-02 2.913974e-02 2.828314e-02
#>  [66] 2.744755e-02 2.663268e-02 2.583820e-02 2.506382e-02 2.430921e-02
#>  [71] 2.357403e-02 2.285796e-02 2.216066e-02 2.148179e-02 2.082099e-02
#>  [76] 2.017793e-02 1.955226e-02 1.894364e-02 1.835170e-02 1.777611e-02
#>  [81] 1.721653e-02 1.667259e-02 1.614397e-02 1.563033e-02 1.513132e-02
#>  [86] 1.464661e-02 1.417586e-02 1.371876e-02 1.327498e-02 1.284419e-02
#>  [91] 1.242609e-02 1.202035e-02 1.162667e-02 1.124476e-02 1.087430e-02
#>  [96] 1.051502e-02 1.016661e-02 9.828807e-03 9.501319e-03 9.183877e-03
#> [101] 8.876215e-03 8.578068e-03 8.289179e-03 8.009296e-03 7.738172e-03
#> [106] 7.475564e-03 7.221235e-03 6.974955e-03 6.736495e-03 6.505635e-03
#> [111] 6.282157e-03 6.065849e-03 5.856506e-03 5.653924e-03 5.457907e-03
#> [116] 5.268262e-03 5.084801e-03 4.907341e-03 4.735704e-03 4.569715e-03
#> [121] 4.409205e-03 4.254008e-03 4.103964e-03 3.958914e-03 3.818708e-03
#> [126] 3.683194e-03 3.552230e-03 3.425673e-03 3.303388e-03 3.185240e-03
#> [131] 3.071099e-03 2.960841e-03 2.854341e-03 2.751482e-03 2.652147e-03
#> [136] 2.556223e-03 2.463602e-03 2.374177e-03 2.287846e-03 2.204507e-03
#> [141] 2.124064e-03 2.046423e-03 1.971492e-03 1.899182e-03 1.829407e-03
#> [146] 1.762084e-03 1.697131e-03 1.634469e-03 1.574024e-03 1.515719e-03
#> [151] 1.459485e-03 1.405251e-03 1.352950e-03 1.302518e-03 1.253890e-03
#> [156] 1.207005e-03 1.161805e-03 1.118232e-03 1.076230e-03 1.035745e-03
#> [161] 9.967263e-04 9.591220e-04 9.228839e-04 8.879645e-04 8.543182e-04
#> [166] 8.219008e-04 7.906694e-04 7.605826e-04 7.316004e-04 7.036840e-04
#> [171] 6.767959e-04 6.509001e-04 6.259613e-04 6.019458e-04 5.788209e-04
#> [176] 5.565549e-04 5.351172e-04 5.144783e-04 4.946096e-04 4.754836e-04
#> [181] 4.570736e-04 4.393538e-04 4.222995e-04 4.058866e-04 3.900918e-04
#> [186] 3.748929e-04 3.602681e-04 3.461966e-04 3.326583e-04 3.196337e-04
#> [191] 3.071039e-04 2.950510e-04 2.834573e-04 2.723061e-04 2.615810e-04
#> [196] 2.512662e-04 2.413468e-04 2.318079e-04 2.226356e-04 2.138161e-04
#> [201] 2.053365e-04 1.971840e-04 1.893464e-04 1.818119e-04 1.745693e-04
#> [206] 1.676075e-04 1.609160e-04 1.544847e-04 1.483038e-04 1.423638e-04
#> [211] 1.366556e-04 1.311704e-04 1.258999e-04 1.208358e-04 1.159703e-04
#> [216] 1.112959e-04 1.068052e-04 1.024913e-04 9.834736e-05 9.436694e-05
#> [221] 9.054375e-05 8.687176e-05 8.334515e-05 7.995833e-05 7.670591e-05
#> [226] 7.358270e-05 7.058372e-05 6.770416e-05 6.493938e-05 6.228495e-05
#> [231] 5.973656e-05 5.729010e-05 5.494159e-05 5.268722e-05 5.052331e-05
#> [236] 4.844632e-05 4.645286e-05 4.453964e-05 4.270353e-05 4.094148e-05
#> [241] 3.925059e-05 3.762806e-05 3.607118e-05 3.457737e-05 3.314414e-05
#> [246] 3.176908e-05 3.044989e-05 2.918436e-05 2.797036e-05 2.680583e-05
#> [251] 2.568881e-05 2.461740e-05 2.358979e-05 2.260422e-05 2.165902e-05
#> [256] 2.075256e-05 1.988330e-05 1.904975e-05 1.825046e-05 1.748407e-05
#> [261] 1.674925e-05 1.604472e-05 1.536926e-05 1.472171e-05 1.410093e-05
#> [266] 1.350584e-05 1.293539e-05 1.238860e-05 1.186450e-05 1.136216e-05
#> [271] 1.088071e-05 1.041929e-05 9.977083e-06 9.553310e-06 9.147216e-06
#> [276] 8.758079e-06 8.385204e-06 8.027926e-06 7.685605e-06 7.357628e-06
#> [281] 7.043404e-06 6.742370e-06 6.453981e-06 6.177718e-06 5.913079e-06
#> [286] 5.659585e-06 5.416776e-06 5.184210e-06 4.961463e-06 4.748127e-06
#> [291] 4.543814e-06 4.348148e-06 4.160770e-06 3.981335e-06 3.809514e-06
#> [296] 3.644988e-06 3.487453e-06 3.336619e-06 3.192204e-06 3.053941e-06
#> [301] 2.921572e-06 2.794850e-06 2.673539e-06 2.557412e-06 2.446250e-06
#> [306] 2.339846e-06 2.237999e-06 2.140518e-06 2.047217e-06 1.957922e-06
#> [311] 1.872463e-06 1.790678e-06 1.712412e-06 1.637516e-06 1.565847e-06
#> [316] 1.497268e-06 1.431649e-06 1.368863e-06 1.308790e-06 1.251316e-06
#> [321] 1.196329e-06 1.143724e-06 1.093398e-06 1.045255e-06 9.992023e-07
#> [326] 9.551494e-07 9.130114e-07 8.727062e-07 8.341554e-07 7.972838e-07
#> [331] 7.620193e-07 7.282931e-07 6.960389e-07 6.651937e-07 6.356966e-07
#> [336] 6.074898e-07 5.805175e-07 5.547266e-07 5.300662e-07 5.064873e-07
#> [341] 4.839433e-07 4.623893e-07 4.417827e-07 4.220822e-07 4.032487e-07
#> [346] 3.852446e-07 3.680338e-07 3.515819e-07 3.358559e-07 3.208243e-07
#> [351] 3.064567e-07 2.927244e-07 2.795995e-07 2.670557e-07 2.550675e-07
#> [356] 2.436106e-07 2.326619e-07 2.221991e-07 2.122009e-07 2.026471e-07
#> [361] 1.935180e-07 1.847951e-07 1.764606e-07 1.684974e-07 1.608891e-07
#> [366] 1.536202e-07 1.466757e-07 1.400414e-07 1.337035e-07 1.276491e-07
#> [371] 1.218655e-07 1.163408e-07 1.110636e-07 1.060230e-07 1.012084e-07
#> [376] 9.660994e-08 9.221793e-08 8.802326e-08 8.401718e-08 8.019130e-08
#> [381] 7.653764e-08 7.304853e-08 6.971665e-08 6.653501e-08 6.349691e-08
#> [386] 6.059597e-08 5.782606e-08 5.518134e-08 5.265622e-08 5.024535e-08
#> [391] 4.794364e-08 4.574620e-08 4.364835e-08 4.164565e-08 3.973382e-08
#> [396] 3.790880e-08 3.616669e-08 3.450376e-08 3.291646e-08 3.140139e-08
#> [401] 2.995531e-08 2.857510e-08 2.725781e-08 2.600059e-08 2.480074e-08
#> [406] 2.365567e-08 2.256291e-08 2.152010e-08 2.052498e-08 1.957539e-08
#> [411] 1.866927e-08 1.780466e-08 1.697968e-08 1.619252e-08 1.544148e-08
#> [416] 1.472492e-08 1.404127e-08 1.338903e-08 1.276678e-08 1.217316e-08
#> [421] 1.160686e-08 1.106664e-08 1.055130e-08 1.005973e-08 9.590825e-09
#> [426] 9.143561e-09 8.716946e-09 8.310040e-09 7.921939e-09 7.551785e-09
#> [431] 7.198757e-09 6.862070e-09 6.540976e-09 6.234761e-09 5.942742e-09
#> [436] 5.664268e-09 5.398717e-09 5.145495e-09 4.904037e-09 4.673801e-09
#> [441] 4.454271e-09 4.244954e-09 4.045381e-09 3.855101e-09 3.673687e-09
#> [446] 3.500730e-09 3.335839e-09 3.178642e-09 3.028784e-09 2.885925e-09
#> [451] 2.749742e-09 2.619926e-09 2.496182e-09 2.378229e-09 2.265799e-09
#> [456] 2.158635e-09 2.056493e-09 1.959141e-09 1.866356e-09 1.777925e-09
#> [461] 1.693646e-09 1.613326e-09 1.536782e-09 1.463836e-09 1.394323e-09
#> [466] 1.328081e-09 1.264958e-09 1.204808e-09 1.147494e-09 1.092882e-09
#> [471] 1.040847e-09 9.912672e-10 9.440288e-10 8.990218e-10 8.561420e-10
#> [476] 8.152898e-10 7.763700e-10 7.392922e-10 7.039699e-10 6.703209e-10
#> [481] 6.382665e-10 6.077319e-10 5.786457e-10 5.509398e-10 5.245492e-10
#> [486] 4.994122e-10 4.754696e-10 4.526653e-10 4.309455e-10 4.102592e-10
#> [491] 3.905577e-10 3.717944e-10 3.539251e-10 3.369076e-10 3.207016e-10
#> [496] 3.052687e-10 2.905724e-10 2.765779e-10 2.632519e-10 2.505627e-10