Calculate the whole-reef community value for a given light level

This function returns the community value for the whole reef community at a light level, given shallow and mesophotic community light relationships.

reef(
  light = seq(0.0065, 1, length = 500),
  Vm = 0.393,
  K = 13.5,
  a = 1.54,
  b = 8.92
)

Arguments

light
The proportion of surface irradiance (PAR), between 0 and 1.

Vm
Passed to shallow. The maximum similarity of a reef patch to the idealised shallow reef community, between 0 and 1. Defaults to (Tamir et al., 2019).

K
Passed to shallow. The light value where the community value = Vmax/2, between 0 and 100. Defaults to (Tamir et al., 2019). This can be interpreted as the light level at which shallow reef communities become light limited.

a
Passed to mesophotic. The scale parameter of a Weibull distribution, defaults to (Tamir et al., 2019). a describes the spread of the data, or ‘the peakyness’. Larger numbers create a flatter, wider, curve. Varying a will simultaneously change the width of the preferred light environment of mesophotic taxa, and the maximum similarity a reef patch is expected to achieve in comparison to the idealised mesophotic community.

b
Passed to mesophotic. The shape parameter of a Weibull distribution, defaults to (Tamir et al., 2019). A value of 3 approximates a normal distribution. Values larger than 3 introduce a left-skew to the curve. This means mesophotic taxa can be light-limited, while keeping mesophotic communities in low-light environments. Varying b shifts the preferred light environment of mesophotic communities.

Value

A reef wide community indicator value. This is the basis of the model presented in (Laverick et al., 2020). When a single light value is provided, a single value is returned. You can pass a vector of light levels for a vector of depths.

Details

The function assumes a tradeoff, where an observed patch of reef can approach a mesophotic or shallow community. A patch of reef cannot simultaneously look like the exemplar mesophotic and shallow reef patch. This is achieved by subtracting the shallow community value from the mesophotic community value.

Essentially this function is a wrapper for mesophotic() - shallow(), but it is neccessary for use in boundary when calculating the light level the communities switch at.

References

(Tamir et al., 2019)

(Laverick et al., 2020)

See also

Examples

# Specifying a single light value
# returns a single community value.

   reef(light = 10)

#> [1] -0.3877652

# Specifying a vector of light values
# returns a vector of community values.

   reef(light = seq(0.0065, 1, length = 500))

#>   [1]  0.0231794162  0.0239493283  0.0240371209  0.0236189587  0.0228025766
#>   [6]  0.0216596516  0.0202408890  0.0185839110  0.0167177412  0.0146655162
#>  [11]  0.0124462091  0.0100757669  0.0075678866  0.0049345572  0.0021864474
#>  [16] -0.0006668125 -0.0036164231 -0.0066542599 -0.0097727563 -0.0129648142
#>  [21] -0.0162237360 -0.0195431713 -0.0229170764 -0.0263396828 -0.0298054720
#>  [26] -0.0333091565 -0.0368456647 -0.0404101287 -0.0439978753 -0.0476044180
#>  [31] -0.0512254517 -0.0548568469 -0.0584946463 -0.0621350614 -0.0657744691
#>  [36] -0.0694094091 -0.0730365817 -0.0766528452 -0.0802552135 -0.0838408540
#>  [41] -0.0874070849 -0.0909513729 -0.0944713307 -0.0979647141 -0.1014294195
#>  [46] -0.1048634809 -0.1082650665 -0.1116324762 -0.1149641381 -0.1182586046
#>  [51] -0.1215145500 -0.1247307661 -0.1279061588 -0.1310397448 -0.1341306472
#>  [56] -0.1371780924 -0.1401814059 -0.1431400085 -0.1460534125 -0.1489212180
#>  [61] -0.1517431088 -0.1545188489 -0.1572482782 -0.1599313095 -0.1625679239
#>  [66] -0.1651581680 -0.1677021497 -0.1702000348 -0.1726520437 -0.1750584478
#>  [71] -0.1774195662 -0.1797357627 -0.1820074423 -0.1842350483 -0.1864190596
#>  [76] -0.1885599871 -0.1906583718 -0.1927147816 -0.1947298088 -0.1967040677
#>  [81] -0.1986381920 -0.2005328329 -0.2023886567 -0.2042063427 -0.2059865812
#>  [86] -0.2077300719 -0.2094375215 -0.2111096428 -0.2127471525 -0.2143507699
#>  [91] -0.2159212154 -0.2174592095 -0.2189654710 -0.2204407162 -0.2218856579
#>  [96] -0.2233010040 -0.2246874569 -0.2260457125 -0.2273764596 -0.2286803788
#> [101] -0.2299581421 -0.2312104124 -0.2324378426 -0.2336410754 -0.2348207430
#> [106] -0.2359774662 -0.2371118546 -0.2382245061 -0.2393160066 -0.2403869299
#> [111] -0.2414378374 -0.2424692783 -0.2434817890 -0.2444758936 -0.2454521035
#> [116] -0.2464109174 -0.2473528216 -0.2482782897 -0.2491877830 -0.2500817505
#> [121] -0.2509606287 -0.2518248421 -0.2526748034 -0.2535109132 -0.2543335606
#> [126] -0.2551431233 -0.2559399677 -0.2567244489 -0.2574969115 -0.2582576893
#> [131] -0.2590071056 -0.2597454735 -0.2604730964 -0.2611902677 -0.2618972715
#> [136] -0.2625943823 -0.2632818661 -0.2639599797 -0.2646289716 -0.2652890818
#> [141] -0.2659405424 -0.2665835777 -0.2672184043 -0.2678452314 -0.2684642612
#> [146] -0.2690756888 -0.2696797028 -0.2702764852 -0.2708662118 -0.2714490522
#> [151] -0.2720251704 -0.2725947246 -0.2731578674 -0.2737147464 -0.2742655040
#> [156] -0.2748102776 -0.2753492001 -0.2758823996 -0.2764100001 -0.2769321211
#> [161] -0.2774488783 -0.2779603832 -0.2784667437 -0.2789680641 -0.2794644453
#> [166] -0.2799559846 -0.2804427762 -0.2809249113 -0.2814024781 -0.2818755618
#> [171] -0.2823442451 -0.2828086078 -0.2832687273 -0.2837246786 -0.2841765344
#> [176] -0.2846243650 -0.2850682388 -0.2855082218 -0.2859443782 -0.2863767705
#> [181] -0.2868054590 -0.2872305025 -0.2876519580 -0.2880698810 -0.2884843253
#> [186] -0.2888953432 -0.2893029858 -0.2897073026 -0.2901083420 -0.2905061507
#> [191] -0.2909007748 -0.2912922586 -0.2916806458 -0.2920659787 -0.2924482986
#> [196] -0.2928276459 -0.2932040601 -0.2935775795 -0.2939482417 -0.2943160834
#> [201] -0.2946811406 -0.2950434483 -0.2954030408 -0.2957599516 -0.2961142138
#> [206] -0.2964658593 -0.2968149198 -0.2971614261 -0.2975054084 -0.2978468964
#> [211] -0.2981859191 -0.2985225051 -0.2988566823 -0.2991884782 -0.2995179196
#> [216] -0.2998450330 -0.3001698444 -0.3004923793 -0.3008126627 -0.3011307193
#> [221] -0.3014465731 -0.3017602481 -0.3020717676 -0.3023811545 -0.3026884316
#> [226] -0.3029936210 -0.3032967447 -0.3035978242 -0.3038968807 -0.3041939351
#> [231] -0.3044890080 -0.3047821196 -0.3050732899 -0.3053625385 -0.3056498847
#> [236] -0.3059353477 -0.3062189462 -0.3065006988 -0.3067806237 -0.3070587389
#> [241] -0.3073350622 -0.3076096110 -0.3078824027 -0.3081534542 -0.3084227823
#> [246] -0.3086904037 -0.3089563345 -0.3092205910 -0.3094831891 -0.3097441444
#> [251] -0.3100034725 -0.3102611885 -0.3105173077 -0.3107718448 -0.3110248147
#> [256] -0.3112762318 -0.3115261104 -0.3117744647 -0.3120213087 -0.3122666561
#> [261] -0.3125105206 -0.3127529157 -0.3129938546 -0.3132333504 -0.3134714162
#> [266] -0.3137080646 -0.3139433085 -0.3141771603 -0.3144096323 -0.3146407367
#> [271] -0.3148704856 -0.3150988909 -0.3153259644 -0.3155517178 -0.3157761624
#> [276] -0.3159993097 -0.3162211709 -0.3164417572 -0.3166610794 -0.3168791484
#> [281] -0.3170959750 -0.3173115698 -0.3175259431 -0.3177391055 -0.3179510670
#> [286] -0.3181618379 -0.3183714282 -0.3185798477 -0.3187871063 -0.3189932136
#> [291] -0.3191981792 -0.3194020126 -0.3196047230 -0.3198063199 -0.3200068122
#> [296] -0.3202062092 -0.3204045197 -0.3206017526 -0.3207979167 -0.3209930206
#> [301] -0.3211870729 -0.3213800821 -0.3215720565 -0.3217630046 -0.3219529344
#> [306] -0.3221418541 -0.3223297718 -0.3225166953 -0.3227026327 -0.3228875916
#> [311] -0.3230715797 -0.3232546047 -0.3234366742 -0.3236177955 -0.3237979761
#> [316] -0.3239772233 -0.3241555443 -0.3243329463 -0.3245094364 -0.3246850216
#> [321] -0.3248597087 -0.3250335048 -0.3252064166 -0.3253784508 -0.3255496141
#> [326] -0.3257199131 -0.3258893543 -0.3260579441 -0.3262256891 -0.3263925955
#> [331] -0.3265586696 -0.3267239176 -0.3268883456 -0.3270519598 -0.3272147662
#> [336] -0.3273767707 -0.3275379792 -0.3276983976 -0.3278580318 -0.3280168873
#> [341] -0.3281749700 -0.3283322854 -0.3284888391 -0.3286446367 -0.3287996835
#> [346] -0.3289539851 -0.3291075467 -0.3292603737 -0.3294124713 -0.3295638448
#> [351] -0.3297144993 -0.3298644399 -0.3300136717 -0.3301621997 -0.3303100288
#> [356] -0.3304571640 -0.3306036102 -0.3307493722 -0.3308944547 -0.3310388626
#> [361] -0.3311826004 -0.3313256729 -0.3314680847 -0.3316098402 -0.3317509442
#> [366] -0.3318914010 -0.3320312151 -0.3321703908 -0.3323089326 -0.3324468448
#> [371] -0.3325841316 -0.3327207973 -0.3328568461 -0.3329922821 -0.3331271096
#> [376] -0.3332613325 -0.3333949550 -0.3335279811 -0.3336604147 -0.3337922598
#> [381] -0.3339235203 -0.3340542001 -0.3341843030 -0.3343138329 -0.3344427936
#> [386] -0.3345711886 -0.3346990219 -0.3348262970 -0.3349530177 -0.3350791875
#> [391] -0.3352048099 -0.3353298887 -0.3354544272 -0.3355784290 -0.3357018975
#> [396] -0.3358248362 -0.3359472485 -0.3360691378 -0.3361905073 -0.3363113604
#> [401] -0.3364317005 -0.3365515307 -0.3366708543 -0.3367896745 -0.3369079945
#> [406] -0.3370258174 -0.3371431463 -0.3372599845 -0.3373763348 -0.3374922004
#> [411] -0.3376075844 -0.3377224896 -0.3378369191 -0.3379508758 -0.3380643627
#> [416] -0.3381773826 -0.3382899384 -0.3384020330 -0.3385136692 -0.3386248498
#> [421] -0.3387355776 -0.3388458554 -0.3389556858 -0.3390650716 -0.3391740156
#> [426] -0.3392825203 -0.3393905884 -0.3394982226 -0.3396054255 -0.3397121996
#> [431] -0.3398185475 -0.3399244718 -0.3400299750 -0.3401350595 -0.3402397280
#> [436] -0.3403439828 -0.3404478264 -0.3405512612 -0.3406542896 -0.3407569141
#> [441] -0.3408591370 -0.3409609606 -0.3410623873 -0.3411634193 -0.3412640591
#> [446] -0.3413643089 -0.3414641709 -0.3415636474 -0.3416627406 -0.3417614527
#> [451] -0.3418597860 -0.3419577425 -0.3420553245 -0.3421525341 -0.3422493734
#> [456] -0.3423458446 -0.3424419496 -0.3425376908 -0.3426330699 -0.3427280892
#> [461] -0.3428227507 -0.3429170564 -0.3430110082 -0.3431046082 -0.3431978583
#> [466] -0.3432907606 -0.3433833168 -0.3434755291 -0.3435673992 -0.3436589291
#> [471] -0.3437501207 -0.3438409758 -0.3439314963 -0.3440216841 -0.3441115410
#> [476] -0.3442010687 -0.3442902692 -0.3443791441 -0.3444676953 -0.3445559246
#> [481] -0.3446438336 -0.3447314242 -0.3448186980 -0.3449056568 -0.3449923023
#> [486] -0.3450786361 -0.3451646600 -0.3452503756 -0.3453357846 -0.3454208885
#> [491] -0.3455056891 -0.3455901880 -0.3456743867 -0.3457582869 -0.3458418901
#> [496] -0.3459251979 -0.3460082120 -0.3460909337 -0.3461733648 -0.3462555066

# By default the function uses the shallow and mesophotic
# community light relationships from (Tamir et al., 2019)
# for a range of light values.

   reef()

#>   [1]  0.0231794162  0.0239493283  0.0240371209  0.0236189587  0.0228025766
#>   [6]  0.0216596516  0.0202408890  0.0185839110  0.0167177412  0.0146655162
#>  [11]  0.0124462091  0.0100757669  0.0075678866  0.0049345572  0.0021864474
#>  [16] -0.0006668125 -0.0036164231 -0.0066542599 -0.0097727563 -0.0129648142
#>  [21] -0.0162237360 -0.0195431713 -0.0229170764 -0.0263396828 -0.0298054720
#>  [26] -0.0333091565 -0.0368456647 -0.0404101287 -0.0439978753 -0.0476044180
#>  [31] -0.0512254517 -0.0548568469 -0.0584946463 -0.0621350614 -0.0657744691
#>  [36] -0.0694094091 -0.0730365817 -0.0766528452 -0.0802552135 -0.0838408540
#>  [41] -0.0874070849 -0.0909513729 -0.0944713307 -0.0979647141 -0.1014294195
#>  [46] -0.1048634809 -0.1082650665 -0.1116324762 -0.1149641381 -0.1182586046
#>  [51] -0.1215145500 -0.1247307661 -0.1279061588 -0.1310397448 -0.1341306472
#>  [56] -0.1371780924 -0.1401814059 -0.1431400085 -0.1460534125 -0.1489212180
#>  [61] -0.1517431088 -0.1545188489 -0.1572482782 -0.1599313095 -0.1625679239
#>  [66] -0.1651581680 -0.1677021497 -0.1702000348 -0.1726520437 -0.1750584478
#>  [71] -0.1774195662 -0.1797357627 -0.1820074423 -0.1842350483 -0.1864190596
#>  [76] -0.1885599871 -0.1906583718 -0.1927147816 -0.1947298088 -0.1967040677
#>  [81] -0.1986381920 -0.2005328329 -0.2023886567 -0.2042063427 -0.2059865812
#>  [86] -0.2077300719 -0.2094375215 -0.2111096428 -0.2127471525 -0.2143507699
#>  [91] -0.2159212154 -0.2174592095 -0.2189654710 -0.2204407162 -0.2218856579
#>  [96] -0.2233010040 -0.2246874569 -0.2260457125 -0.2273764596 -0.2286803788
#> [101] -0.2299581421 -0.2312104124 -0.2324378426 -0.2336410754 -0.2348207430
#> [106] -0.2359774662 -0.2371118546 -0.2382245061 -0.2393160066 -0.2403869299
#> [111] -0.2414378374 -0.2424692783 -0.2434817890 -0.2444758936 -0.2454521035
#> [116] -0.2464109174 -0.2473528216 -0.2482782897 -0.2491877830 -0.2500817505
#> [121] -0.2509606287 -0.2518248421 -0.2526748034 -0.2535109132 -0.2543335606
#> [126] -0.2551431233 -0.2559399677 -0.2567244489 -0.2574969115 -0.2582576893
#> [131] -0.2590071056 -0.2597454735 -0.2604730964 -0.2611902677 -0.2618972715
#> [136] -0.2625943823 -0.2632818661 -0.2639599797 -0.2646289716 -0.2652890818
#> [141] -0.2659405424 -0.2665835777 -0.2672184043 -0.2678452314 -0.2684642612
#> [146] -0.2690756888 -0.2696797028 -0.2702764852 -0.2708662118 -0.2714490522
#> [151] -0.2720251704 -0.2725947246 -0.2731578674 -0.2737147464 -0.2742655040
#> [156] -0.2748102776 -0.2753492001 -0.2758823996 -0.2764100001 -0.2769321211
#> [161] -0.2774488783 -0.2779603832 -0.2784667437 -0.2789680641 -0.2794644453
#> [166] -0.2799559846 -0.2804427762 -0.2809249113 -0.2814024781 -0.2818755618
#> [171] -0.2823442451 -0.2828086078 -0.2832687273 -0.2837246786 -0.2841765344
#> [176] -0.2846243650 -0.2850682388 -0.2855082218 -0.2859443782 -0.2863767705
#> [181] -0.2868054590 -0.2872305025 -0.2876519580 -0.2880698810 -0.2884843253
#> [186] -0.2888953432 -0.2893029858 -0.2897073026 -0.2901083420 -0.2905061507
#> [191] -0.2909007748 -0.2912922586 -0.2916806458 -0.2920659787 -0.2924482986
#> [196] -0.2928276459 -0.2932040601 -0.2935775795 -0.2939482417 -0.2943160834
#> [201] -0.2946811406 -0.2950434483 -0.2954030408 -0.2957599516 -0.2961142138
#> [206] -0.2964658593 -0.2968149198 -0.2971614261 -0.2975054084 -0.2978468964
#> [211] -0.2981859191 -0.2985225051 -0.2988566823 -0.2991884782 -0.2995179196
#> [216] -0.2998450330 -0.3001698444 -0.3004923793 -0.3008126627 -0.3011307193
#> [221] -0.3014465731 -0.3017602481 -0.3020717676 -0.3023811545 -0.3026884316
#> [226] -0.3029936210 -0.3032967447 -0.3035978242 -0.3038968807 -0.3041939351
#> [231] -0.3044890080 -0.3047821196 -0.3050732899 -0.3053625385 -0.3056498847
#> [236] -0.3059353477 -0.3062189462 -0.3065006988 -0.3067806237 -0.3070587389
#> [241] -0.3073350622 -0.3076096110 -0.3078824027 -0.3081534542 -0.3084227823
#> [246] -0.3086904037 -0.3089563345 -0.3092205910 -0.3094831891 -0.3097441444
#> [251] -0.3100034725 -0.3102611885 -0.3105173077 -0.3107718448 -0.3110248147
#> [256] -0.3112762318 -0.3115261104 -0.3117744647 -0.3120213087 -0.3122666561
#> [261] -0.3125105206 -0.3127529157 -0.3129938546 -0.3132333504 -0.3134714162
#> [266] -0.3137080646 -0.3139433085 -0.3141771603 -0.3144096323 -0.3146407367
#> [271] -0.3148704856 -0.3150988909 -0.3153259644 -0.3155517178 -0.3157761624
#> [276] -0.3159993097 -0.3162211709 -0.3164417572 -0.3166610794 -0.3168791484
#> [281] -0.3170959750 -0.3173115698 -0.3175259431 -0.3177391055 -0.3179510670
#> [286] -0.3181618379 -0.3183714282 -0.3185798477 -0.3187871063 -0.3189932136
#> [291] -0.3191981792 -0.3194020126 -0.3196047230 -0.3198063199 -0.3200068122
#> [296] -0.3202062092 -0.3204045197 -0.3206017526 -0.3207979167 -0.3209930206
#> [301] -0.3211870729 -0.3213800821 -0.3215720565 -0.3217630046 -0.3219529344
#> [306] -0.3221418541 -0.3223297718 -0.3225166953 -0.3227026327 -0.3228875916
#> [311] -0.3230715797 -0.3232546047 -0.3234366742 -0.3236177955 -0.3237979761
#> [316] -0.3239772233 -0.3241555443 -0.3243329463 -0.3245094364 -0.3246850216
#> [321] -0.3248597087 -0.3250335048 -0.3252064166 -0.3253784508 -0.3255496141
#> [326] -0.3257199131 -0.3258893543 -0.3260579441 -0.3262256891 -0.3263925955
#> [331] -0.3265586696 -0.3267239176 -0.3268883456 -0.3270519598 -0.3272147662
#> [336] -0.3273767707 -0.3275379792 -0.3276983976 -0.3278580318 -0.3280168873
#> [341] -0.3281749700 -0.3283322854 -0.3284888391 -0.3286446367 -0.3287996835
#> [346] -0.3289539851 -0.3291075467 -0.3292603737 -0.3294124713 -0.3295638448
#> [351] -0.3297144993 -0.3298644399 -0.3300136717 -0.3301621997 -0.3303100288
#> [356] -0.3304571640 -0.3306036102 -0.3307493722 -0.3308944547 -0.3310388626
#> [361] -0.3311826004 -0.3313256729 -0.3314680847 -0.3316098402 -0.3317509442
#> [366] -0.3318914010 -0.3320312151 -0.3321703908 -0.3323089326 -0.3324468448
#> [371] -0.3325841316 -0.3327207973 -0.3328568461 -0.3329922821 -0.3331271096
#> [376] -0.3332613325 -0.3333949550 -0.3335279811 -0.3336604147 -0.3337922598
#> [381] -0.3339235203 -0.3340542001 -0.3341843030 -0.3343138329 -0.3344427936
#> [386] -0.3345711886 -0.3346990219 -0.3348262970 -0.3349530177 -0.3350791875
#> [391] -0.3352048099 -0.3353298887 -0.3354544272 -0.3355784290 -0.3357018975
#> [396] -0.3358248362 -0.3359472485 -0.3360691378 -0.3361905073 -0.3363113604
#> [401] -0.3364317005 -0.3365515307 -0.3366708543 -0.3367896745 -0.3369079945
#> [406] -0.3370258174 -0.3371431463 -0.3372599845 -0.3373763348 -0.3374922004
#> [411] -0.3376075844 -0.3377224896 -0.3378369191 -0.3379508758 -0.3380643627
#> [416] -0.3381773826 -0.3382899384 -0.3384020330 -0.3385136692 -0.3386248498
#> [421] -0.3387355776 -0.3388458554 -0.3389556858 -0.3390650716 -0.3391740156
#> [426] -0.3392825203 -0.3393905884 -0.3394982226 -0.3396054255 -0.3397121996
#> [431] -0.3398185475 -0.3399244718 -0.3400299750 -0.3401350595 -0.3402397280
#> [436] -0.3403439828 -0.3404478264 -0.3405512612 -0.3406542896 -0.3407569141
#> [441] -0.3408591370 -0.3409609606 -0.3410623873 -0.3411634193 -0.3412640591
#> [446] -0.3413643089 -0.3414641709 -0.3415636474 -0.3416627406 -0.3417614527
#> [451] -0.3418597860 -0.3419577425 -0.3420553245 -0.3421525341 -0.3422493734
#> [456] -0.3423458446 -0.3424419496 -0.3425376908 -0.3426330699 -0.3427280892
#> [461] -0.3428227507 -0.3429170564 -0.3430110082 -0.3431046082 -0.3431978583
#> [466] -0.3432907606 -0.3433833168 -0.3434755291 -0.3435673992 -0.3436589291
#> [471] -0.3437501207 -0.3438409758 -0.3439314963 -0.3440216841 -0.3441115410
#> [476] -0.3442010687 -0.3442902692 -0.3443791441 -0.3444676953 -0.3445559246
#> [481] -0.3446438336 -0.3447314242 -0.3448186980 -0.3449056568 -0.3449923023
#> [486] -0.3450786361 -0.3451646600 -0.3452503756 -0.3453357846 -0.3454208885
#> [491] -0.3455056891 -0.3455901880 -0.3456743867 -0.3457582869 -0.3458418901
#> [496] -0.3459251979 -0.3460082120 -0.3460909337 -0.3461733648 -0.3462555066

# You can specify your own community light relationships
# by setting parameters explicitly

   reef(Vm = 0.2, K = 11, a = 1.2, b = 8)

#>   [1]  0.075277738  0.075170075  0.074151254  0.072522890  0.070462023
#>   [6]  0.068082694  0.065463297  0.062660359  0.059716159  0.056663243
#>  [11]  0.053527244  0.050328735  0.047084481  0.043808321  0.040511795
#>  [16]  0.037204613  0.033895007  0.030589992  0.027295577  0.024016930
#>  [21]  0.020758503  0.017524144  0.014317178  0.011140481  0.007996535
#>  [26]  0.004887481  0.001815161 -0.001218851 -0.004213214 -0.007166795
#>  [31] -0.010078647 -0.012947989 -0.015774190 -0.018556751 -0.021295291
#>  [36] -0.023989540 -0.026639323 -0.029244553 -0.031805222 -0.034321395
#>  [41] -0.036793198 -0.039220817 -0.041604491 -0.043944504 -0.046241185
#>  [46] -0.048494899 -0.050706046 -0.052875056 -0.055002388 -0.057088522
#>  [51] -0.059133963 -0.061139233 -0.063104870 -0.065031427 -0.066919471
#>  [56] -0.068769577 -0.070582330 -0.072358321 -0.074098150 -0.075802418
#>  [61] -0.077471731 -0.079106697 -0.080707924 -0.082276022 -0.083811599
#>  [66] -0.085315261 -0.086787614 -0.088229258 -0.089640791 -0.091022807
#>  [71] -0.092375896 -0.093700640 -0.094997619 -0.096267404 -0.097510561
#>  [76] -0.098727651 -0.099919225 -0.101085829 -0.102228001 -0.103346271
#>  [81] -0.104441162 -0.105513188 -0.106562857 -0.107590666 -0.108597107
#>  [86] -0.109582661 -0.110547802 -0.111492996 -0.112418701 -0.113325363
#>  [91] -0.114213425 -0.115083317 -0.115935464 -0.116770281 -0.117588174
#>  [96] -0.118389544 -0.119174780 -0.119944264 -0.120698373 -0.121437471
#> [101] -0.122161919 -0.122872066 -0.123568257 -0.124250828 -0.124920105
#> [106] -0.125576411 -0.126220060 -0.126851357 -0.127470601 -0.128078087
#> [111] -0.128674098 -0.129258915 -0.129832809 -0.130396047 -0.130948887
#> [116] -0.131491583 -0.132024383 -0.132547527 -0.133061250 -0.133565782
#> [121] -0.134061346 -0.134548162 -0.135026440 -0.135496389 -0.135958211
#> [126] -0.136412103 -0.136858258 -0.137296861 -0.137728095 -0.138152139
#> [131] -0.138569165 -0.138979343 -0.139382835 -0.139779803 -0.140170403
#> [136] -0.140554785 -0.140933098 -0.141305487 -0.141672090 -0.142033045
#> [141] -0.142388485 -0.142738539 -0.143083333 -0.143422990 -0.143757629
#> [146] -0.144087366 -0.144412315 -0.144732585 -0.145048285 -0.145359518
#> [151] -0.145666385 -0.145968987 -0.146267419 -0.146561774 -0.146852144
#> [156] -0.147138617 -0.147421281 -0.147700218 -0.147975511 -0.148247240
#> [161] -0.148515481 -0.148780311 -0.149041802 -0.149300027 -0.149555054
#> [166] -0.149806951 -0.150055785 -0.150301619 -0.150544516 -0.150784536
#> [171] -0.151021739 -0.151256183 -0.151487923 -0.151717014 -0.151943509
#> [176] -0.152167460 -0.152388918 -0.152607931 -0.152824547 -0.153038813
#> [181] -0.153250775 -0.153460475 -0.153667958 -0.153873265 -0.154076438
#> [186] -0.154277515 -0.154476536 -0.154673538 -0.154868559 -0.155061633
#> [191] -0.155252797 -0.155442084 -0.155629527 -0.155815159 -0.155999011
#> [196] -0.156181115 -0.156361499 -0.156540194 -0.156717229 -0.156892630
#> [201] -0.157066425 -0.157238641 -0.157409304 -0.157578438 -0.157746069
#> [206] -0.157912220 -0.158076916 -0.158240178 -0.158402030 -0.158562493
#> [211] -0.158721589 -0.158879338 -0.159035762 -0.159190879 -0.159344709
#> [216] -0.159497272 -0.159648586 -0.159798669 -0.159947539 -0.160095213
#> [221] -0.160241709 -0.160387043 -0.160531230 -0.160674288 -0.160816232
#> [226] -0.160957076 -0.161096837 -0.161235528 -0.161373163 -0.161509758
#> [231] -0.161645324 -0.161779877 -0.161913429 -0.162045992 -0.162177580
#> [236] -0.162308205 -0.162437879 -0.162566614 -0.162694421 -0.162821312
#> [241] -0.162947298 -0.163072390 -0.163196600 -0.163319936 -0.163442411
#> [246] -0.163564033 -0.163684814 -0.163804762 -0.163923887 -0.164042200
#> [251] -0.164159708 -0.164276422 -0.164392350 -0.164507500 -0.164621883
#> [256] -0.164735505 -0.164848376 -0.164960503 -0.165071896 -0.165182560
#> [261] -0.165292505 -0.165401738 -0.165510267 -0.165618098 -0.165725240
#> [266] -0.165831699 -0.165937483 -0.166042598 -0.166147051 -0.166250849
#> [271] -0.166353999 -0.166456506 -0.166558378 -0.166659621 -0.166760241
#> [276] -0.166860244 -0.166959636 -0.167058424 -0.167156612 -0.167254207
#> [281] -0.167351214 -0.167447640 -0.167543489 -0.167638767 -0.167733479
#> [286] -0.167827631 -0.167921228 -0.168014274 -0.168106776 -0.168198738
#> [291] -0.168290165 -0.168381062 -0.168471433 -0.168561284 -0.168650619
#> [296] -0.168739442 -0.168827759 -0.168915574 -0.169002890 -0.169089714
#> [301] -0.169176048 -0.169261897 -0.169347266 -0.169432158 -0.169516578
#> [306] -0.169600529 -0.169684016 -0.169767043 -0.169849613 -0.169931730
#> [311] -0.170013399 -0.170094623 -0.170175405 -0.170255750 -0.170335661
#> [316] -0.170415141 -0.170494195 -0.170572825 -0.170651035 -0.170728829
#> [321] -0.170806209 -0.170883180 -0.170959745 -0.171035906 -0.171111667
#> [326] -0.171187032 -0.171262003 -0.171336583 -0.171410776 -0.171484585
#> [331] -0.171558013 -0.171631062 -0.171703736 -0.171776037 -0.171847969
#> [336] -0.171919535 -0.171990736 -0.172061577 -0.172132059 -0.172202186
#> [341] -0.172271959 -0.172341383 -0.172410459 -0.172479191 -0.172547580
#> [346] -0.172615630 -0.172683342 -0.172750720 -0.172817766 -0.172884482
#> [351] -0.172950871 -0.173016935 -0.173082677 -0.173148098 -0.173213203
#> [356] -0.173277991 -0.173342467 -0.173406632 -0.173470489 -0.173534039
#> [361] -0.173597285 -0.173660229 -0.173722874 -0.173785221 -0.173847273
#> [366] -0.173909031 -0.173970498 -0.174031676 -0.174092567 -0.174153173
#> [371] -0.174213495 -0.174273537 -0.174333299 -0.174392784 -0.174451994
#> [376] -0.174510931 -0.174569596 -0.174627992 -0.174686119 -0.174743981
#> [381] -0.174801579 -0.174858915 -0.174915990 -0.174972807 -0.175029366
#> [386] -0.175085671 -0.175141722 -0.175197521 -0.175253070 -0.175308371
#> [391] -0.175363425 -0.175418235 -0.175472800 -0.175527124 -0.175581208
#> [396] -0.175635053 -0.175688662 -0.175742034 -0.175795173 -0.175848080
#> [401] -0.175900755 -0.175953202 -0.176005420 -0.176057412 -0.176109180
#> [406] -0.176160724 -0.176212046 -0.176263147 -0.176314029 -0.176364694
#> [411] -0.176415142 -0.176465375 -0.176515395 -0.176565203 -0.176614799
#> [416] -0.176664187 -0.176713366 -0.176762338 -0.176811104 -0.176859666
#> [421] -0.176908026 -0.176956183 -0.177004140 -0.177051898 -0.177099458
#> [426] -0.177146821 -0.177193989 -0.177240962 -0.177287742 -0.177334331
#> [431] -0.177380728 -0.177426936 -0.177472956 -0.177518788 -0.177564434
#> [436] -0.177609895 -0.177655172 -0.177700267 -0.177745180 -0.177789912
#> [441] -0.177834465 -0.177878839 -0.177923036 -0.177967057 -0.178010903
#> [446] -0.178054574 -0.178098073 -0.178141399 -0.178184554 -0.178227539
#> [451] -0.178270355 -0.178313003 -0.178355484 -0.178397798 -0.178439948
#> [456] -0.178481934 -0.178523756 -0.178565416 -0.178606914 -0.178648253
#> [461] -0.178689431 -0.178730452 -0.178771314 -0.178812020 -0.178852571
#> [466] -0.178892966 -0.178933207 -0.178973295 -0.179013231 -0.179053015
#> [471] -0.179092649 -0.179132133 -0.179171468 -0.179210655 -0.179249696
#> [476] -0.179288589 -0.179327337 -0.179365941 -0.179404400 -0.179442717
#> [481] -0.179480891 -0.179518924 -0.179556815 -0.179594567 -0.179632180
#> [486] -0.179669655 -0.179706991 -0.179744191 -0.179781255 -0.179818183
#> [491] -0.179854977 -0.179891636 -0.179928163 -0.179964557 -0.180000819
#> [496] -0.180036951 -0.180072952 -0.180108823 -0.180144566 -0.180180180