Calculate the shallow community value for a given light level

This function captures the preferred light environment of the "shallow" coral community.

shallow(light = seq(0.0065, 100, length = 500), Vm = 0.393, K = 13.5)

Arguments

light
The percentage of surface irradiance (PAR) reaching a reef, positive and up to 100.

Vm
The maximum similarity of a reef patch to the idealised shallow reef community, between 0 and 1. Defaults to (Tamir et al., 2019).

K
The light value where the community value = Vmax/2, between 0 and 100. Defaults to (Tamir et al., 2019). This can be interpreted as the light level at which shallow reef communities become light limited.

Value

A shallow community indicator value, as discussed in (Laverick et al., 2017). When a single light value is provided, a single value is returned. You can pass a vector of light levels for a vector of depths.

Details

A Michaelis-Menten equation is used to represent the relationship between shallow coral communities and light. We assume shallow corals are light-limited beneath a light threshold, and that community values (how similar a reef patch is to the idealised shallow community) increase asymptotically with increasing light.

This relationship has been verified in the Red Sea (Tamir et al., 2019), but certainly needs more research!

References

(Tamir et al., 2019)

(Laverick et al., 2017)

See also

Examples

# Specifying a single light value
# returns a single community value.

   shallow(light = 10)

#> [1] 0.167234

# Specifying a vector of light values
# returns a vector of community values.

   shallow(light = seq(0.0065, 100, length = 500))

#>   [1] 0.0001891312 0.0059318276 0.0115090329 0.0169277989 0.0221947826
#>   [6] 0.0273162731 0.0322982170 0.0371462409 0.0418656731 0.0464615629
#>  [11] 0.0509386986 0.0553016244 0.0595546552 0.0637018917 0.0677472328
#>  [16] 0.0716943881 0.0755468895 0.0793081013 0.0829812304 0.0865693352
#>  [21] 0.0900753341 0.0935020133 0.0968520346 0.1001279418 0.1033321676
#>  [26] 0.1064670393 0.1095347843 0.1125375360 0.1154773382 0.1183561497
#>  [31] 0.1211758492 0.1239382388 0.1266450481 0.1292979379 0.1318985034
#>  [36] 0.1344482774 0.1369487337 0.1394012894 0.1418073078 0.1441681011
#>  [41] 0.1464849325 0.1487590186 0.1509915317 0.1531836013 0.1553363166
#>  [46] 0.1574507281 0.1595278489 0.1615686570 0.1635740963 0.1655450785
#>  [51] 0.1674824839 0.1693871635 0.1712599393 0.1731016064 0.1749129335
#>  [56] 0.1766946641 0.1784475178 0.1801721911 0.1818693580 0.1835396713
#>  [61] 0.1851837635 0.1868022471 0.1883957159 0.1899647453 0.1915098933
#>  [66] 0.1930317010 0.1945306934 0.1960073795 0.1974622538 0.1988957959
#>  [71] 0.2003084716 0.2017007332 0.2030730199 0.2044257587 0.2057593642
#>  [76] 0.2070742396 0.2083707766 0.2096493562 0.2109103490 0.2121541153
#>  [81] 0.2133810058 0.2145913616 0.2157855147 0.2169637882 0.2181264970
#>  [86] 0.2192739472 0.2204064374 0.2215242582 0.2226276928 0.2237170172
#>  [91] 0.2247925002 0.2258544041 0.2269029843 0.2279384902 0.2289611646
#>  [96] 0.2299712446 0.2309689615 0.2319545409 0.2329282029 0.2338901623
#> [101] 0.2348406289 0.2357798075 0.2367078978 0.2376250952 0.2385315903
#> [106] 0.2394275693 0.2403132142 0.2411887027 0.2420542086 0.2429099015
#> [111] 0.2437559475 0.2445925088 0.2454197440 0.2462378081 0.2470468529
#> [116] 0.2478470267 0.2486384745 0.2494213384 0.2501957573 0.2509618670
#> [121] 0.2517198006 0.2524696883 0.2532116575 0.2539458329 0.2546723368
#> [126] 0.2553912888 0.2561028059 0.2568070030 0.2575039925 0.2581938843
#> [131] 0.2588767864 0.2595528045 0.2602220421 0.2608846007 0.2615405799
#> [136] 0.2621900772 0.2628331880 0.2634700063 0.2641006239 0.2647251309
#> [141] 0.2653436157 0.2659561650 0.2665628638 0.2671637955 0.2677590421
#> [146] 0.2683486838 0.2689327993 0.2695114660 0.2700847599 0.2706527553
#> [151] 0.2712155254 0.2717731420 0.2723256755 0.2728731952 0.2734157689
#> [156] 0.2739534634 0.2744863442 0.2750144756 0.2755379209 0.2760567422
#> [161] 0.2765710004 0.2770807556 0.2775860665 0.2780869911 0.2785835862
#> [166] 0.2790759077 0.2795640105 0.2800479487 0.2805277753 0.2810035425
#> [171] 0.2814753016 0.2819431029 0.2824069962 0.2828670302 0.2833232528
#> [176] 0.2837757113 0.2842244519 0.2846695204 0.2851109616 0.2855488198
#> [181] 0.2859831383 0.2864139598 0.2868413266 0.2872652799 0.2876858604
#> [186] 0.2881031084 0.2885170632 0.2889277636 0.2893352479 0.2897395538
#> [191] 0.2901407182 0.2905387776 0.2909337680 0.2913257247 0.2917146825
#> [196] 0.2921006757 0.2924837380 0.2928639028 0.2932412028 0.2936156702
#> [201] 0.2939873368 0.2943562340 0.2947223925 0.2950858427 0.2954466147
#> [206] 0.2958047378 0.2961602411 0.2965131534 0.2968635027 0.2972113170
#> [211] 0.2975566236 0.2978994496 0.2982398216 0.2985777658 0.2989133082
#> [216] 0.2992464742 0.2995772890 0.2999057775 0.3002319640 0.3005558727
#> [221] 0.3008775273 0.3011969514 0.3015141679 0.3018291998 0.3021420695
#> [226] 0.3024527992 0.3027614108 0.3030679259 0.3033723657 0.3036747513
#> [231] 0.3039751033 0.3042734423 0.3045697884 0.3048641615 0.3051565812
#> [236] 0.3054470670 0.3057356379 0.3060223128 0.3063071103 0.3065900489
#> [241] 0.3068711466 0.3071504214 0.3074278909 0.3077035727 0.3079774838
#> [246] 0.3082496414 0.3085200621 0.3087887627 0.3090557594 0.3093210683
#> [251] 0.3095847055 0.3098466868 0.3101070275 0.3103657432 0.3106228490
#> [256] 0.3108783598 0.3111322904 0.3113846556 0.3116354696 0.3118847468
#> [261] 0.3121325013 0.3123787469 0.3126234973 0.3128667663 0.3131085671
#> [266] 0.3133489131 0.3135878172 0.3138252926 0.3140613519 0.3142960077
#> [271] 0.3145292726 0.3147611588 0.3149916786 0.3152208440 0.3154486669
#> [276] 0.3156751591 0.3159003322 0.3161241976 0.3163467668 0.3165680510
#> [281] 0.3167880612 0.3170068084 0.3172243036 0.3174405573 0.3176555802
#> [286] 0.3178693828 0.3180819754 0.3182933683 0.3185035716 0.3187125953
#> [291] 0.3189204493 0.3191271434 0.3193326874 0.3195370906 0.3197403627
#> [296] 0.3199425131 0.3201435508 0.3203434852 0.3205423252 0.3207400799
#> [301] 0.3209367581 0.3211323685 0.3213269199 0.3215204208 0.3217128797
#> [306] 0.3219043050 0.3220947049 0.3222840878 0.3224724617 0.3226598347
#> [311] 0.3228462148 0.3230316097 0.3232160274 0.3233994754 0.3235819615
#> [316] 0.3237634932 0.3239440779 0.3241237230 0.3243024359 0.3244802238
#> [321] 0.3246570939 0.3248330532 0.3250081087 0.3251822675 0.3253555363
#> [326] 0.3255279220 0.3256994313 0.3258700709 0.3260398474 0.3262087673
#> [331] 0.3263768371 0.3265440631 0.3267104518 0.3268760093 0.3270407420
#> [336] 0.3272046559 0.3273677572 0.3275300518 0.3276915458 0.3278522450
#> [341] 0.3280121553 0.3281712825 0.3283296324 0.3284872106 0.3286440227
#> [346] 0.3288000743 0.3289553710 0.3291099181 0.3292637212 0.3294167856
#> [351] 0.3295691165 0.3297207193 0.3298715992 0.3300217612 0.3301712106
#> [356] 0.3303199524 0.3304679915 0.3306153331 0.3307619819 0.3309079428
#> [361] 0.3310532208 0.3311978205 0.3313417467 0.3314850041 0.3316275974
#> [366] 0.3317695311 0.3319108099 0.3320514382 0.3321914205 0.3323307613
#> [371] 0.3324694650 0.3326075359 0.3327449783 0.3328817966 0.3330179950
#> [376] 0.3331535776 0.3332885487 0.3334229123 0.3335566727 0.3336898337
#> [381] 0.3338223995 0.3339543740 0.3340857611 0.3342165648 0.3343467890
#> [386] 0.3344764375 0.3346055141 0.3347340226 0.3348619667 0.3349893501
#> [391] 0.3351161766 0.3352424497 0.3353681731 0.3354933503 0.3356179850
#> [396] 0.3357420805 0.3358656405 0.3359886684 0.3361111675 0.3362331414
#> [401] 0.3363545933 0.3364755267 0.3365959448 0.3367158509 0.3368352483
#> [406] 0.3369541402 0.3370725298 0.3371904203 0.3373078149 0.3374247166
#> [411] 0.3375411285 0.3376570538 0.3377724955 0.3378874566 0.3380019400
#> [416] 0.3381159488 0.3382294860 0.3383425543 0.3384551568 0.3385672963
#> [421] 0.3386789756 0.3387901976 0.3389009651 0.3390112808 0.3391211475
#> [426] 0.3392305680 0.3393395450 0.3394480811 0.3395561791 0.3396638415
#> [431] 0.3397710710 0.3398778702 0.3399842418 0.3400901881 0.3401957119
#> [436] 0.3403008156 0.3404055017 0.3405097728 0.3406136312 0.3407170794
#> [441] 0.3408201199 0.3409227550 0.3410249871 0.3411268187 0.3412282520
#> [446] 0.3413292894 0.3414299332 0.3415301857 0.3416300492 0.3417295259
#> [451] 0.3418286181 0.3419273279 0.3420256577 0.3421236096 0.3422211858
#> [456] 0.3423183884 0.3424152196 0.3425116814 0.3426077761 0.3427035057
#> [461] 0.3427988722 0.3428938778 0.3429885245 0.3430828143 0.3431767492
#> [466] 0.3432703313 0.3433635624 0.3434564447 0.3435489799 0.3436411702
#> [471] 0.3437330173 0.3438245233 0.3439156900 0.3440065193 0.3440970130
#> [476] 0.3441871731 0.3442770014 0.3443664996 0.3444556696 0.3445445133
#> [481] 0.3446330323 0.3447212285 0.3448091037 0.3448966595 0.3449838978
#> [486] 0.3450708202 0.3451574285 0.3452437244 0.3453297094 0.3454153855
#> [491] 0.3455007541 0.3455858169 0.3456705756 0.3457550319 0.3458391872
#> [496] 0.3459230433 0.3460066017 0.3460898640 0.3461728318 0.3462555066

# By default the function uses the shallow community
# light relationship from (Tamir et al., 2019) for a
# range of light values.

   shallow()

#>   [1] 0.0001891312 0.0059318276 0.0115090329 0.0169277989 0.0221947826
#>   [6] 0.0273162731 0.0322982170 0.0371462409 0.0418656731 0.0464615629
#>  [11] 0.0509386986 0.0553016244 0.0595546552 0.0637018917 0.0677472328
#>  [16] 0.0716943881 0.0755468895 0.0793081013 0.0829812304 0.0865693352
#>  [21] 0.0900753341 0.0935020133 0.0968520346 0.1001279418 0.1033321676
#>  [26] 0.1064670393 0.1095347843 0.1125375360 0.1154773382 0.1183561497
#>  [31] 0.1211758492 0.1239382388 0.1266450481 0.1292979379 0.1318985034
#>  [36] 0.1344482774 0.1369487337 0.1394012894 0.1418073078 0.1441681011
#>  [41] 0.1464849325 0.1487590186 0.1509915317 0.1531836013 0.1553363166
#>  [46] 0.1574507281 0.1595278489 0.1615686570 0.1635740963 0.1655450785
#>  [51] 0.1674824839 0.1693871635 0.1712599393 0.1731016064 0.1749129335
#>  [56] 0.1766946641 0.1784475178 0.1801721911 0.1818693580 0.1835396713
#>  [61] 0.1851837635 0.1868022471 0.1883957159 0.1899647453 0.1915098933
#>  [66] 0.1930317010 0.1945306934 0.1960073795 0.1974622538 0.1988957959
#>  [71] 0.2003084716 0.2017007332 0.2030730199 0.2044257587 0.2057593642
#>  [76] 0.2070742396 0.2083707766 0.2096493562 0.2109103490 0.2121541153
#>  [81] 0.2133810058 0.2145913616 0.2157855147 0.2169637882 0.2181264970
#>  [86] 0.2192739472 0.2204064374 0.2215242582 0.2226276928 0.2237170172
#>  [91] 0.2247925002 0.2258544041 0.2269029843 0.2279384902 0.2289611646
#>  [96] 0.2299712446 0.2309689615 0.2319545409 0.2329282029 0.2338901623
#> [101] 0.2348406289 0.2357798075 0.2367078978 0.2376250952 0.2385315903
#> [106] 0.2394275693 0.2403132142 0.2411887027 0.2420542086 0.2429099015
#> [111] 0.2437559475 0.2445925088 0.2454197440 0.2462378081 0.2470468529
#> [116] 0.2478470267 0.2486384745 0.2494213384 0.2501957573 0.2509618670
#> [121] 0.2517198006 0.2524696883 0.2532116575 0.2539458329 0.2546723368
#> [126] 0.2553912888 0.2561028059 0.2568070030 0.2575039925 0.2581938843
#> [131] 0.2588767864 0.2595528045 0.2602220421 0.2608846007 0.2615405799
#> [136] 0.2621900772 0.2628331880 0.2634700063 0.2641006239 0.2647251309
#> [141] 0.2653436157 0.2659561650 0.2665628638 0.2671637955 0.2677590421
#> [146] 0.2683486838 0.2689327993 0.2695114660 0.2700847599 0.2706527553
#> [151] 0.2712155254 0.2717731420 0.2723256755 0.2728731952 0.2734157689
#> [156] 0.2739534634 0.2744863442 0.2750144756 0.2755379209 0.2760567422
#> [161] 0.2765710004 0.2770807556 0.2775860665 0.2780869911 0.2785835862
#> [166] 0.2790759077 0.2795640105 0.2800479487 0.2805277753 0.2810035425
#> [171] 0.2814753016 0.2819431029 0.2824069962 0.2828670302 0.2833232528
#> [176] 0.2837757113 0.2842244519 0.2846695204 0.2851109616 0.2855488198
#> [181] 0.2859831383 0.2864139598 0.2868413266 0.2872652799 0.2876858604
#> [186] 0.2881031084 0.2885170632 0.2889277636 0.2893352479 0.2897395538
#> [191] 0.2901407182 0.2905387776 0.2909337680 0.2913257247 0.2917146825
#> [196] 0.2921006757 0.2924837380 0.2928639028 0.2932412028 0.2936156702
#> [201] 0.2939873368 0.2943562340 0.2947223925 0.2950858427 0.2954466147
#> [206] 0.2958047378 0.2961602411 0.2965131534 0.2968635027 0.2972113170
#> [211] 0.2975566236 0.2978994496 0.2982398216 0.2985777658 0.2989133082
#> [216] 0.2992464742 0.2995772890 0.2999057775 0.3002319640 0.3005558727
#> [221] 0.3008775273 0.3011969514 0.3015141679 0.3018291998 0.3021420695
#> [226] 0.3024527992 0.3027614108 0.3030679259 0.3033723657 0.3036747513
#> [231] 0.3039751033 0.3042734423 0.3045697884 0.3048641615 0.3051565812
#> [236] 0.3054470670 0.3057356379 0.3060223128 0.3063071103 0.3065900489
#> [241] 0.3068711466 0.3071504214 0.3074278909 0.3077035727 0.3079774838
#> [246] 0.3082496414 0.3085200621 0.3087887627 0.3090557594 0.3093210683
#> [251] 0.3095847055 0.3098466868 0.3101070275 0.3103657432 0.3106228490
#> [256] 0.3108783598 0.3111322904 0.3113846556 0.3116354696 0.3118847468
#> [261] 0.3121325013 0.3123787469 0.3126234973 0.3128667663 0.3131085671
#> [266] 0.3133489131 0.3135878172 0.3138252926 0.3140613519 0.3142960077
#> [271] 0.3145292726 0.3147611588 0.3149916786 0.3152208440 0.3154486669
#> [276] 0.3156751591 0.3159003322 0.3161241976 0.3163467668 0.3165680510
#> [281] 0.3167880612 0.3170068084 0.3172243036 0.3174405573 0.3176555802
#> [286] 0.3178693828 0.3180819754 0.3182933683 0.3185035716 0.3187125953
#> [291] 0.3189204493 0.3191271434 0.3193326874 0.3195370906 0.3197403627
#> [296] 0.3199425131 0.3201435508 0.3203434852 0.3205423252 0.3207400799
#> [301] 0.3209367581 0.3211323685 0.3213269199 0.3215204208 0.3217128797
#> [306] 0.3219043050 0.3220947049 0.3222840878 0.3224724617 0.3226598347
#> [311] 0.3228462148 0.3230316097 0.3232160274 0.3233994754 0.3235819615
#> [316] 0.3237634932 0.3239440779 0.3241237230 0.3243024359 0.3244802238
#> [321] 0.3246570939 0.3248330532 0.3250081087 0.3251822675 0.3253555363
#> [326] 0.3255279220 0.3256994313 0.3258700709 0.3260398474 0.3262087673
#> [331] 0.3263768371 0.3265440631 0.3267104518 0.3268760093 0.3270407420
#> [336] 0.3272046559 0.3273677572 0.3275300518 0.3276915458 0.3278522450
#> [341] 0.3280121553 0.3281712825 0.3283296324 0.3284872106 0.3286440227
#> [346] 0.3288000743 0.3289553710 0.3291099181 0.3292637212 0.3294167856
#> [351] 0.3295691165 0.3297207193 0.3298715992 0.3300217612 0.3301712106
#> [356] 0.3303199524 0.3304679915 0.3306153331 0.3307619819 0.3309079428
#> [361] 0.3310532208 0.3311978205 0.3313417467 0.3314850041 0.3316275974
#> [366] 0.3317695311 0.3319108099 0.3320514382 0.3321914205 0.3323307613
#> [371] 0.3324694650 0.3326075359 0.3327449783 0.3328817966 0.3330179950
#> [376] 0.3331535776 0.3332885487 0.3334229123 0.3335566727 0.3336898337
#> [381] 0.3338223995 0.3339543740 0.3340857611 0.3342165648 0.3343467890
#> [386] 0.3344764375 0.3346055141 0.3347340226 0.3348619667 0.3349893501
#> [391] 0.3351161766 0.3352424497 0.3353681731 0.3354933503 0.3356179850
#> [396] 0.3357420805 0.3358656405 0.3359886684 0.3361111675 0.3362331414
#> [401] 0.3363545933 0.3364755267 0.3365959448 0.3367158509 0.3368352483
#> [406] 0.3369541402 0.3370725298 0.3371904203 0.3373078149 0.3374247166
#> [411] 0.3375411285 0.3376570538 0.3377724955 0.3378874566 0.3380019400
#> [416] 0.3381159488 0.3382294860 0.3383425543 0.3384551568 0.3385672963
#> [421] 0.3386789756 0.3387901976 0.3389009651 0.3390112808 0.3391211475
#> [426] 0.3392305680 0.3393395450 0.3394480811 0.3395561791 0.3396638415
#> [431] 0.3397710710 0.3398778702 0.3399842418 0.3400901881 0.3401957119
#> [436] 0.3403008156 0.3404055017 0.3405097728 0.3406136312 0.3407170794
#> [441] 0.3408201199 0.3409227550 0.3410249871 0.3411268187 0.3412282520
#> [446] 0.3413292894 0.3414299332 0.3415301857 0.3416300492 0.3417295259
#> [451] 0.3418286181 0.3419273279 0.3420256577 0.3421236096 0.3422211858
#> [456] 0.3423183884 0.3424152196 0.3425116814 0.3426077761 0.3427035057
#> [461] 0.3427988722 0.3428938778 0.3429885245 0.3430828143 0.3431767492
#> [466] 0.3432703313 0.3433635624 0.3434564447 0.3435489799 0.3436411702
#> [471] 0.3437330173 0.3438245233 0.3439156900 0.3440065193 0.3440970130
#> [476] 0.3441871731 0.3442770014 0.3443664996 0.3444556696 0.3445445133
#> [481] 0.3446330323 0.3447212285 0.3448091037 0.3448966595 0.3449838978
#> [486] 0.3450708202 0.3451574285 0.3452437244 0.3453297094 0.3454153855
#> [491] 0.3455007541 0.3455858169 0.3456705756 0.3457550319 0.3458391872
#> [496] 0.3459230433 0.3460066017 0.3460898640 0.3461728318 0.3462555066

# You can specify your own community light relationship
# by setting parameters explicitly

   shallow(Vm = 0.2, K = 11)

#>   [1] 0.000118112 0.003692154 0.007140628 0.010470037 0.013686443 0.016795503
#>   [7] 0.019802503 0.022712387 0.025529786 0.028259038 0.030904218 0.033469151
#>  [13] 0.035957434 0.038372453 0.040717396 0.042995269 0.045208911 0.047360998
#>  [19] 0.049454065 0.051490506 0.053472588 0.055402459 0.057282156 0.059113609
#>  [25] 0.060898653 0.062639030 0.064336395 0.065992323 0.067608314 0.069185795
#>  [31] 0.070726127 0.072230606 0.073700470 0.075136900 0.076541023 0.077913918
#>  [37] 0.079256615 0.080570100 0.081855315 0.083113163 0.084344510 0.085550184
#>  [43] 0.086730980 0.087887660 0.089020955 0.090131567 0.091220171 0.092287414
#>  [49] 0.093333919 0.094360285 0.095367087 0.096354879 0.097324195 0.098275549
#>  [55] 0.099209435 0.100126330 0.101026693 0.101910967 0.102779580 0.103632945
#>  [61] 0.104471459 0.105295506 0.106105459 0.106901674 0.107684500 0.108454271
#>  [67] 0.109211310 0.109955931 0.110688437 0.111409122 0.112118269 0.112816153
#>  [73] 0.113503040 0.114179188 0.114844848 0.115500260 0.116145661 0.116781278
#>  [79] 0.117407330 0.118024034 0.118631597 0.119230219 0.119820098 0.120401424
#>  [85] 0.120974381 0.121539148 0.122095900 0.122644806 0.123186032 0.123719736
#>  [91] 0.124246076 0.124765201 0.125277260 0.125782396 0.126280749 0.126772453
#>  [97] 0.127257641 0.127736442 0.128208982 0.128675381 0.129135760 0.129590234
#> [103] 0.130038915 0.130481914 0.130919338 0.131351292 0.131777878 0.132199194
#> [109] 0.132615339 0.133026407 0.133432489 0.133833677 0.134230058 0.134621719
#> [115] 0.135008742 0.135391210 0.135769203 0.136142798 0.136512073 0.136877101
#> [121] 0.137237956 0.137594709 0.137947428 0.138296183 0.138641040 0.138982063
#> [127] 0.139319317 0.139652863 0.139982762 0.140309074 0.140631856 0.140951167
#> [133] 0.141267061 0.141579593 0.141888817 0.142194785 0.142497547 0.142797155
#> [139] 0.143093656 0.143387100 0.143677532 0.143965000 0.144249549 0.144531222
#> [145] 0.144810063 0.145086115 0.145359419 0.145630016 0.145897946 0.146163249
#> [151] 0.146425962 0.146686123 0.146943771 0.147198939 0.147451665 0.147701984
#> [157] 0.147949929 0.148195534 0.148438832 0.148679855 0.148918635 0.149155204
#> [163] 0.149389592 0.149621829 0.149851944 0.150079966 0.150305924 0.150529846
#> [169] 0.150751759 0.150971690 0.151189665 0.151405711 0.151619852 0.151832115
#> [175] 0.152042523 0.152251101 0.152457872 0.152662861 0.152866089 0.153067580
#> [181] 0.153267355 0.153465437 0.153661846 0.153856605 0.154049733 0.154241252
#> [187] 0.154431181 0.154619539 0.154806347 0.154991623 0.155175386 0.155357655
#> [193] 0.155538447 0.155717781 0.155895674 0.156072144 0.156247207 0.156420880
#> [199] 0.156593180 0.156764123 0.156933724 0.157102001 0.157268967 0.157434639
#> [205] 0.157599031 0.157762158 0.157924034 0.158084675 0.158244094 0.158402304
#> [211] 0.158559321 0.158715156 0.158869824 0.159023337 0.159175709 0.159326951
#> [217] 0.159477077 0.159626099 0.159774029 0.159920879 0.160066660 0.160211385
#> [223] 0.160355064 0.160497710 0.160639333 0.160779944 0.160919554 0.161058173
#> [229] 0.161195813 0.161332483 0.161468194 0.161602955 0.161736777 0.161869670
#> [235] 0.162001643 0.162132705 0.162262866 0.162392136 0.162520523 0.162648036
#> [241] 0.162774684 0.162900477 0.163025422 0.163149529 0.163272805 0.163395259
#> [247] 0.163516900 0.163637734 0.163757771 0.163877018 0.163995483 0.164113173
#> [253] 0.164230096 0.164346260 0.164461672 0.164576339 0.164690269 0.164803468
#> [259] 0.164915944 0.165027703 0.165138752 0.165249098 0.165358748 0.165467708
#> [265] 0.165575985 0.165683585 0.165790514 0.165896779 0.166002386 0.166107341
#> [271] 0.166211650 0.166315319 0.166418353 0.166520759 0.166622543 0.166723709
#> [277] 0.166824265 0.166924214 0.167023562 0.167122316 0.167220480 0.167318060
#> [283] 0.167415060 0.167511486 0.167607344 0.167702637 0.167797371 0.167891551
#> [289] 0.167985182 0.168078268 0.168170815 0.168262827 0.168354308 0.168445263
#> [295] 0.168535697 0.168625614 0.168715018 0.168803915 0.168892308 0.168980201
#> [301] 0.169067599 0.169154506 0.169240925 0.169326862 0.169412320 0.169497304
#> [307] 0.169581816 0.169665861 0.169749444 0.169832566 0.169915234 0.169997449
#> [313] 0.170079216 0.170160539 0.170241421 0.170321866 0.170401877 0.170481457
#> [319] 0.170560611 0.170639342 0.170717652 0.170795546 0.170873027 0.170950097
#> [325] 0.171026761 0.171103021 0.171178881 0.171254344 0.171329412 0.171404090
#> [331] 0.171478379 0.171552283 0.171625806 0.171698949 0.171771716 0.171844110
#> [337] 0.171916134 0.171987790 0.172059081 0.172130011 0.172200581 0.172270794
#> [343] 0.172340654 0.172410163 0.172479323 0.172548138 0.172616609 0.172684739
#> [349] 0.172752532 0.172819988 0.172887112 0.172953904 0.173020369 0.173086507
#> [355] 0.173152322 0.173217816 0.173282991 0.173347850 0.173412395 0.173476628
#> [361] 0.173540551 0.173604166 0.173667477 0.173730485 0.173793191 0.173855600
#> [367] 0.173917711 0.173979528 0.174041053 0.174102288 0.174163234 0.174223895
#> [373] 0.174284271 0.174344364 0.174404178 0.174463714 0.174522973 0.174581957
#> [379] 0.174640670 0.174699111 0.174757284 0.174815190 0.174872831 0.174930209
#> [385] 0.174987325 0.175044181 0.175100780 0.175157122 0.175213211 0.175269046
#> [391] 0.175324630 0.175379966 0.175435053 0.175489895 0.175544492 0.175598846
#> [397] 0.175652960 0.175706834 0.175760470 0.175813869 0.175867034 0.175919966
#> [403] 0.175972666 0.176025136 0.176077377 0.176129391 0.176181180 0.176232744
#> [409] 0.176284085 0.176335205 0.176386106 0.176436787 0.176487252 0.176537501
#> [415] 0.176587535 0.176637357 0.176686967 0.176736367 0.176785557 0.176834541
#> [421] 0.176883318 0.176931890 0.176980258 0.177028424 0.177076389 0.177124154
#> [427] 0.177171720 0.177219089 0.177266261 0.177313239 0.177360023 0.177406614
#> [433] 0.177453014 0.177499224 0.177545245 0.177591078 0.177636724 0.177682185
#> [439] 0.177727461 0.177772554 0.177817465 0.177862194 0.177906744 0.177951114
#> [445] 0.177995307 0.178039323 0.178083163 0.178126828 0.178170320 0.178213639
#> [451] 0.178256787 0.178299764 0.178342572 0.178385211 0.178427682 0.178469987
#> [457] 0.178512126 0.178554101 0.178595911 0.178637560 0.178679046 0.178720372
#> [463] 0.178761537 0.178802544 0.178843393 0.178884084 0.178924620 0.178965000
#> [469] 0.179005225 0.179045297 0.179085217 0.179124984 0.179164601 0.179204067
#> [475] 0.179243385 0.179282554 0.179321575 0.179360450 0.179399178 0.179437762
#> [481] 0.179476201 0.179514497 0.179552651 0.179590662 0.179628532 0.179666263
#> [487] 0.179703853 0.179741305 0.179778619 0.179815796 0.179852836 0.179889741
#> [493] 0.179926510 0.179963146 0.179999647 0.180036017 0.180072254 0.180108360
#> [499] 0.180144335 0.180180180